@@ -5,11 +5,9 @@
     "homepage": "https://developers.onelogin.com/saml/php",
     "keywords": ["saml", "saml2", "onelogin"],
     "autoload": {
-        "classmap": [
-            "extlib/xmlseclibs",
-            "lib/Saml",
-            "lib/Saml2"
-        ]
+        "psr-4": {
+            "OneLogin\\": "src/"
+        }
     },
     "support": {
         "email": "sixto.garcia@onelogin.com",
@@ -17,22 +15,20 @@
         "source": "https://github.com/onelogin/php-saml/"
     },
     "require": {
-        "php": ">=5.3.2 <7.2",
-        "ext-curl": "*",
-        "ext-openssl": "*",
-        "ext-dom": "*",
-        "ext-mcrypt": "*",
-        "ext-zlib": "*"
+        "php": ">=5.4",
+        "robrichards/xmlseclibs": ">=3.1.1",
+        "phpunit/phpunit": "<7.5.18"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.8",
-        "satooshi/php-coveralls": "1.0.1",
-        "sebastian/phpcpd": "*",
-        "phploc/phploc": "*",
-        "pdepend/pdepend" : "1.1.0",
-        "squizlabs/php_codesniffer": "2.9.0"
+        "php-coveralls/php-coveralls": "^1.0.2 || ^2.0",
+        "sebastian/phpcpd": "^2.0 || ^3.0 || ^4.0",
+        "phploc/phploc": "^2.1 || ^3.0 || ^4.0",
+        "pdepend/pdepend": "^2.5.0",
+        "squizlabs/php_codesniffer": "^3.1.1"
     },
     "suggest": {
+        "ext-openssl": "Install openssl lib in order to handle with x509 certs (require to support sign and encryption)",
+        "ext-curl": "Install curl lib to be able to use the IdPMetadataParser for parsing remote XMLs",
         "ext-gettext": "Install gettext and php5-gettext libs to handle translations"
     }
 }
