@@ -1,56 +1,69 @@
 <?php
 
+namespace OneLogin\Saml2\Tests;
+
+use OneLogin\Saml2\Auth;
+use OneLogin\Saml2\Error;
+use OneLogin\Saml2\LogoutRequest;
+use OneLogin\Saml2\Settings;
+use OneLogin\Saml2\Utils;
+use OneLogin\Saml2\ValidationError;
+
+use RobRichards\XMLSecLibs\XMLSecurityKey;
+
+use Exception;
+
 /**
  * Unit tests for Auth class
  */
-class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
+class AuthTest extends \PHPUnit\Framework\TestCase
 {
     private $_auth;
     private $_settingsInfo;
 
     /**
-    * Initializes the Test Suite
-    */
+     * Initializes the Test Suite
+     */
     public function setUp()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
         $this->_settingsInfo = $settingsInfo;
-        $this->_auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $this->_auth = new Auth($settingsInfo);
     }
 
     /**
-    * Tests the getSettings method of the OneLogin_Saml2_Auth class
-    * Build a OneLogin_Saml2_Settings object with a setting array
-    * and compare the value returned from the method of the
-    * $auth object
-    *
-    * @covers OneLogin_Saml2_Auth::getSettings
-    */
+     * Tests the getSettings method of the OneLogin\Saml2\Auth class
+     * Build a Settings object with a setting array
+     * and compare the value returned from the method of the
+     * $auth object
+     *
+     * @covers OneLogin\Saml2\Auth::getSettings
+     */
     public function testGetSettings()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $authSettings = $this->_auth->getSettings();
         $this->assertEquals($authSettings, $settings);
     }
 
     /**
-    * Tests the getLastRequestID method of the OneLogin_Saml2_Auth class
-    *
-    * @covers OneLogin_Saml2_Auth::getLastRequestID
-    */
+     * Tests the getLastRequestID method of the Auth class
+     *
+     * @covers OneLogin\Saml2\Auth::getLastRequestID
+     */
     public function testGetLastRequestID()
     {
-        $targetSSOURL = $this->_auth->login(null, array(), false, false, true, false, false);
+        $targetSSOURL = $this->_auth->login(null, array(), false, false, true, false);
         $id1 = $this->_auth->getLastRequestID();
         $this->assertNotNull($id1);
 
-        $targetSLOURL = $this->_auth->logout(null, array(), null, null, true, null, null);
+        $targetSLOURL = $this->_auth->logout(null, array(), null, null, true, null);
         $id2 = $this->_auth->getLastRequestID();
         $this->assertNotNull($id2);
 
@@ -58,10 +71,10 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getSSOurl method of the OneLogin_Saml2_Auth class
-    *
-    * @covers OneLogin_Saml2_Auth::getSSOurl
-    */
+     * Tests the getSSOurl method of the Auth class
+     *
+     * @covers OneLogin\Saml2\Auth::getSSOurl
+     */
     public function testGetSSOurl()
     {
         $ssoUrl = $this->_settingsInfo['idp']['singleSignOnService']['url'];
@@ -69,10 +82,10 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getSLOurl method of the OneLogin_Saml2_Auth class
-    *
-    * @covers OneLogin_Saml2_Auth::getSLOurl
-    */
+     * Tests the getSLOurl method of the Auth class
+     *
+     * @covers OneLogin\Saml2\Auth::getSLOurl
+     */
     public function testGetSLOurl()
     {
         $sloUrl = $this->_settingsInfo['idp']['singleLogoutService']['url'];
@@ -80,28 +93,29 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests the getSLOResponseUrl method of the OneLogin_Saml2_Auth class
+     * Tests the getSLOResponseUrl method of the Auth class
      *
-     * @covers OneLogin_Saml2_Auth::getSLOResponseUrl
+     * @covers OneLogin\Saml2\Auth::getSLOResponseUrl
      */
     public function testGetSLOResponseUrl()
     {
-        $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
-        $this->assertEquals($this->_auth->getSLOResponseUrl(), $sloResponseUrl);
+        $sloUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
+        $this->assertEquals($this->_auth->getSLOResponseUrl(), $sloUrl);
     }
 
+
     /**
-    * Tests the processResponse method of the OneLogin_Saml2_Auth class
-    * Case No Response, An exception is throw
-    *
-    * @covers OneLogin_Saml2_Auth::processResponse
-    */
+     * Tests the processResponse method of the Auth class
+     * Case No Response, An exception is throw
+     *
+     * @covers OneLogin\Saml2\Auth::processResponse
+     */
     public function testProcessNoResponse()
     {
         try {
             $this->_auth->processResponse();
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('SAML Response not found', $e->getMessage());
         }
 
@@ -109,24 +123,25 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processResponse method of the OneLogin_Saml2_Auth class
-    * Case Invalid Response, After processing the response the user
-    * is not authenticated, attributes are notreturned, no nameID and
-    * the error array is not empty, contains 'invalid_response
-    *
-    * @covers OneLogin_Saml2_Auth::processResponse
-    * @covers OneLogin_Saml2_Auth::isAuthenticated
-    * @covers OneLogin_Saml2_Auth::getAttributes
-    * @covers OneLogin_Saml2_Auth::getAttribute
-    * @covers OneLogin_Saml2_Auth::getNameId
-    * @covers OneLogin_Saml2_Auth::getNameIdFormat
-    * @covers OneLogin_Saml2_Auth::getNameIdNameQualifier
-    * @covers OneLogin_Saml2_Auth::getNameIdSPNameQualifier
-    * @covers OneLogin_Saml2_Auth::getErrors
-    * @covers OneLogin_Saml2_Auth::getSessionIndex
-    * @covers OneLogin_Saml2_Auth::getSessionExpiration
-    * @covers OneLogin_Saml2_Auth::getLastErrorReason
-    */
+     * Tests the processResponse method of the Auth class
+     * Case Invalid Response, After processing the response the user
+     * is not authenticated, attributes are notreturned, no nameID and
+     * the error array is not empty, contains 'invalid_response
+     *
+     * @covers OneLogin\Saml2\Auth::processResponse
+     * @covers OneLogin\Saml2\Auth::isAuthenticated
+     * @covers OneLogin\Saml2\Auth::getAttributes
+     * @covers OneLogin\Saml2\Auth::getAttribute
+     * @covers OneLogin\Saml2\Auth::getNameId
+     * @covers OneLogin\Saml2\Auth::getNameIdFormat
+     * @covers OneLogin\Saml2\Auth::getNameIdNameQualifier
+     * @covers OneLogin\Saml2\Auth::getNameIdSPNameQualifier
+     * @covers OneLogin\Saml2\Auth::getErrors
+     * @covers OneLogin\Saml2\Auth::getSessionIndex
+     * @covers OneLogin\Saml2\Auth::getSessionExpiration
+     * @covers OneLogin\Saml2\Auth::getLastErrorReason
+     * @covers OneLogin\Saml2\Auth::getLastErrorException
+     */
     public function testProcessResponseInvalid()
     {
         $message = file_get_contents(TEST_ROOT . '/data/responses/response1.xml.base64');
@@ -145,20 +160,22 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         $this->assertNull($this->_auth->getAttribute('uid'));
         $this->assertEquals($this->_auth->getErrors(), array('invalid_response'));
         $this->assertEquals($this->_auth->getLastErrorReason(), "Reference validation failed");
+        $errorException = $this->_auth->getLastErrorException();
+        $this->assertEquals("Reference validation failed", $errorException->getMessage());
     }
 
     /**
-    * Tests the processResponse method of the OneLogin_Saml2_Auth class
-    * Case Invalid Response, Invalid requestID
-    *
-    * @covers OneLogin_Saml2_Auth::processResponse
-    */
+     * Tests the processResponse method of the Auth class
+     * Case Invalid Response, Invalid requestID
+     *
+     * @covers OneLogin\Saml2\Auth::processResponse
+     */
     public function testProcessResponseInvalidRequestId()
     {
         $message = file_get_contents(TEST_ROOT . '/data/responses/unsigned_response.xml.base64');
 
         $plainMessage = base64_decode($message);
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/acs.php', $currentURL, $plainMessage);
 
         $_POST['SAMLResponse'] = base64_encode($plainMessage);
@@ -167,6 +184,8 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         $this->_auth->processResponse($requestId);
 
         $this->assertEquals("No Signature found. SAML Response rejected", $this->_auth->getLastErrorReason());
+        $errorException = $this->_auth->getLastErrorException();
+        $this->assertEquals("No Signature found. SAML Response rejected", $errorException->getMessage());
 
         $this->_auth->setStrict(true);
         $this->_auth->processResponse($requestId);
@@ -178,21 +197,21 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processResponse method of the OneLogin_Saml2_Auth class
-    * Case Valid Response, After processing the response the user
-    * is authenticated, attributes are returned, also has a nameID and
-    * the error array is empty
-    *
-    * @covers OneLogin_Saml2_Auth::processResponse
-    * @covers OneLogin_Saml2_Auth::isAuthenticated
-    * @covers OneLogin_Saml2_Auth::getAttributes
-    * @covers OneLogin_Saml2_Auth::getAttribute
-    * @covers OneLogin_Saml2_Auth::getNameId
-    * @covers OneLogin_Saml2_Auth::getNameIdFormat
-    * @covers OneLogin_Saml2_Auth::getSessionIndex
-    * @covers OneLogin_Saml2_Auth::getSessionExpiration
-    * @covers OneLogin_Saml2_Auth::getErrors
-    */
+     * Tests the processResponse method of the Auth class
+     * Case Valid Response, After processing the response the user
+     * is authenticated, attributes are returned, also has a nameID and
+     * the error array is empty
+     *
+     * @covers OneLogin\Saml2\Auth::processResponse
+     * @covers OneLogin\Saml2\Auth::isAuthenticated
+     * @covers OneLogin\Saml2\Auth::getAttributes
+     * @covers OneLogin\Saml2\Auth::getAttribute
+     * @covers OneLogin\Saml2\Auth::getNameId
+     * @covers OneLogin\Saml2\Auth::getNameIdFormat
+     * @covers OneLogin\Saml2\Auth::getSessionIndex
+     * @covers OneLogin\Saml2\Auth::getSessionExpiration
+     * @covers OneLogin\Saml2\Auth::getErrors
+     */
     public function testProcessResponseValid()
     {
         $message = file_get_contents(TEST_ROOT . '/data/responses/valid_response.xml.base64');
@@ -216,7 +235,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     /**
      * Tests the getNameIdNameQualifier method of the Auth class
      * Case found
-     * @covers OneLogin_Saml2_Auth::getNameIdNameQualifier
+     * @covers OneLogin\Saml2\Auth::getNameIdNameQualifier
      */
     public function testGetNameIdNameQualifier()
     {
@@ -227,10 +246,11 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         $this->assertTrue($this->_auth->isAuthenticated());
         $this->assertEquals('https://test.example.com/saml/metadata', $this->_auth->getNameIdNameQualifier());
     }
-     /**
+
+    /**
      * Tests the getNameIdNameQualifier method of the Auth class
      * Case Null
-     * @covers OneLogin_Saml2_Auth::getNameIdNameQualifier
+     * @covers OneLogin\Saml2\Auth::getNameIdNameQualifier
      */
     public function testGetNameIdNameQualifier2()
     {
@@ -241,10 +261,11 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         $this->assertTrue($this->_auth->isAuthenticated());
         $this->assertNull($this->_auth->getNameIdNameQualifier());
     }
-     /**
+
+    /**
      * Tests the getNameIdSPNameQualifier method of the Auth class
      * Case Found
-     * @covers OneLogin_Saml2_Auth::getNameIdSPNameQualifier
+     * @covers OneLogin\Saml2\Auth::getNameIdSPNameQualifier
      */
     public function testGetNameIdSPNameQualifier()
     {
@@ -255,10 +276,11 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         $this->assertTrue($this->_auth->isAuthenticated());
         $this->assertNull($this->_auth->getNameIdSPNameQualifier());
     }
-     /**
+
+    /**
      * Tests the getNameIdSPNameQualifier method of the Auth class
      * Case Null
-     * @covers OneLogin_Saml2_Auth::getNameIdSPNameQualifier
+     * @covers OneLogin\Saml2\Auth::getNameIdSPNameQualifier
      */
     public function testGetNameIdSPNameQualifier2()
     {
@@ -271,20 +293,18 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getAttributes and getAttributesWithFriendlyName methods
-    * @covers OneLogin_Saml2_Auth::getAttributes
-    * @covers OneLogin_Saml2_Auth::getAttribute
-    * @covers OneLogin_Saml2_Auth::getAttributesWithFriendlyName
-    * @covers OneLogin_Saml2_Auth::getAttributeWithFriendlyName
+     * Tests the getAttributes and getAttributesWithFriendlyName methods
+     * @covers OneLogin\Saml2\Auth::getAttributes
+     * @covers OneLogin\Saml2\Auth::getAttribute
+     * @covers OneLogin\Saml2\Auth::getAttributesWithFriendlyName
+     * @covers OneLogin\Saml2\Auth::getAttributeWithFriendlyName
      */
     public function testGetAttributes()
     {
-        $auth = new OneLogin_Saml2_Auth($this->_settingsInfo);
-
+        $auth = new Auth($this->_settingsInfo);
         $response = file_get_contents(TEST_ROOT . '/data/responses/response6.xml.base64');
         $_POST['SAMLResponse'] = $response;
         $auth->processResponse();
-
         $expectedAttributes = array(
             'urn:oid:0.9.2342.19200300.100.1.1' => array(
                 'demo'
@@ -303,60 +323,56 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         );
         $this->assertEquals($expectedAttributes, $auth->getAttributes());
         $this->assertEquals($expectedFriendlyNameAttributes, $auth->getAttributesWithFriendlyName());
-
         $this->assertNull($auth->getAttribute('givenName'));
         $this->assertEquals(array('value'), $auth->getAttributeWithFriendlyName('givenName'));
-
         $this->assertEquals(array('value'), $auth->getAttribute('urn:oid:2.5.4.42'));
         $this->assertNull($auth->getAttributeWithFriendlyName('urn:oid:2.5.4.42'));
-
         // An assertion that has no attributes should return an empty array when asked for the attributes
         $response2 = file_get_contents(TEST_ROOT . '/data/responses/response2.xml.base64');
         $_POST['SAMLResponse'] = $response2;
-        $auth2 = new OneLogin_Saml2_Auth($this->_settingsInfo);
+        $auth2 = new Auth($this->_settingsInfo);
         $auth2->processResponse();
         $this->assertEmpty($auth2->getAttributes());
         $this->assertEmpty($auth2->getAttributesWithFriendlyName());
-
         // Encrypted Attributes are not supported
         $response3 = file_get_contents(TEST_ROOT . '/data/responses/invalids/encrypted_attrs.xml.base64');
         $_POST['SAMLResponse'] = $response3;
-        $auth3 = new OneLogin_Saml2_Auth($this->_settingsInfo);
+        $auth3 = new Auth($this->_settingsInfo);
         $auth3->processResponse();
         $this->assertEmpty($auth3->getAttributes());
         $this->assertEmpty($auth3->getAttributesWithFriendlyName());
-
         // Duplicated Attribute names
         $response4 = file_get_contents(TEST_ROOT . '/data/responses/invalids/duplicated_attributes_with_friendly_names.xml.base64');
         $_POST['SAMLResponse'] = $response4;
-        $auth4 = new OneLogin_Saml2_Auth($this->_settingsInfo);
+        $auth4 = new Auth($this->_settingsInfo);
         try {
             $auth4->processResponse();
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $this->fail('OneLogin\Saml2\ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertContains('Found an Attribute element with duplicated FriendlyName', $e->getMessage());
         }
-
         $response5 = file_get_contents(TEST_ROOT . '/data/responses/invalids/duplicated_attributes.xml.base64');
         $_POST['SAMLResponse'] = $response5;
-        $auth5 = new OneLogin_Saml2_Auth($this->_settingsInfo);
+        $auth5 = new Auth($this->_settingsInfo);
         try {
             $auth5->processResponse();
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $this->fail('OneLogin\Saml2\ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertContains('Found an Attribute element with duplicated Name', $e->getMessage());
         }
     }
 
     /**
-    * Tests the redirectTo method of the OneLogin_Saml2_Auth class
-    * (phpunit raises an exception when a redirect is executed, the
-    * exception is catched and we check that the targetURL is correct)
-    * Case redirect without url parameter
-    *
-    * @covers OneLogin_Saml2_Auth::redirectTo
-    * @runInSeparateProcess
-    */
+     * Tests the redirectTo method of the Auth class
+     * (phpunit raises an exception when a redirect is executed, the
+     * exception is catched and we check that the targetURL is correct)
+     * Case redirect without url parameter
+     *
+     * @covers OneLogin\Saml2\Auth::redirectTo
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testRedirectTo()
     {
         try {
@@ -376,14 +392,16 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the redirectTo method of the OneLogin_Saml2_Auth class
-    * (phpunit raises an exception when a redirect is executed, the
-    * exception is catched and we check that the targetURL is correct)
-    * Case redirect with url parameter
-    *
-    * @covers OneLogin_Saml2_Auth::redirectTo
-    * @runInSeparateProcess
-    */
+     * Tests the redirectTo method of the Auth class
+     * (phpunit raises an exception when a redirect is executed, the
+     * exception is catched and we check that the targetURL is correct)
+     * Case redirect with url parameter
+     *
+     * @covers OneLogin\Saml2\Auth::redirectTo
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testRedirectTowithUrl()
     {
         try {
@@ -404,17 +422,17 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case No Message, An exception is throw
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    */
+     * Tests the processSLO method of the Auth class
+     * Case No Message, An exception is throw
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     */
     public function testProcessNoSLO()
     {
         try {
             $this->_auth->processSLO(true);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('SAML LogoutRequest/LogoutResponse not found', $e->getMessage());
         }
 
@@ -422,11 +440,11 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Invalid Logout Response
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Invalid Logout Response
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     */
     public function testProcessSLOResponseInvalid()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_responses/logout_response_deflated.xml.base64');
@@ -446,18 +464,18 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Logout Response not sucess
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Logout Response not sucess
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     */
     public function testProcessSLOResponseNoSucess()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_responses/invalids/status_code_responder.xml.base64');
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -469,18 +487,18 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Logout Response with valid and invalid Request ID
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Logout Response with valid and invalid Request ID
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     */
     public function testProcessSLOResponseRequestId()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_responses/logout_response_deflated.xml.base64');
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -497,18 +515,18 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Valid Logout Response
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Valid Logout Response
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     */
     public function testProcessSLOResponseValid()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_responses/logout_response_deflated.xml.base64');
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -530,11 +548,11 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Valid Logout Response, validating deleting the local session
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Valid Logout Response, validating deleting the local session
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     */
     public function testProcessSLOResponseValidDeletingSession()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_responses/logout_response_deflated.xml.base64');
@@ -546,7 +564,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -561,10 +579,10 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests the processSLO method of the OneLogin_Saml2_Auth class
+     * Tests the processSLO method of the Auth class
      * Case Valid Logout Response, validating deleting the local session
      *
-     * @covers OneLogin_Saml2_Auth::processSLO
+     * @covers OneLogin\Saml2\Auth::processSLO
      */
     public function testProcessSLOResponseValidDeletingSessionCallback()
     {
@@ -575,13 +593,13 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         }
         $_SESSION['samltest'] = true;
 
-        $callback = function() {
+        $callback = function () {
             $_SESSION['samltest'] = false;
         };
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -597,72 +615,60 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Invalid Logout Request
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    * @runInSeparateProcess
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Invalid Logout Request
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testProcessSLORequestInvalidValid()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_deflated.xml.base64');
         $_GET['SAMLRequest'] = $message;
 
-        try {
-            // The Header of the redirect produces an Exception
-            $this->_auth->processSLO(true);
-            // Do not ever get here
-            $this->assertFalse(true);
-        } catch (Exception $e) {
-            $this->assertContains('Cannot modify header information', $e->getMessage());
-            $trace = $e->getTrace();
-            $targetUrl = getUrlFromRedirect($trace);
-            $parsedQuery = getParamsFromUrl($targetUrl);
+        $targetUrl = $this->_auth->processSLO(true, null, false, null, true);
+        $parsedQuery = getParamsFromUrl($targetUrl);
 
-            $this->assertEmpty($this->_auth->getErrors());
-            $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
-            $this->assertContains($sloResponseUrl, $targetUrl);
-            $this->assertArrayHasKey('SAMLResponse', $parsedQuery);
-            $this->assertArrayNotHasKey('RelayState', $parsedQuery);
-        }
+        $this->assertEmpty($this->_auth->getErrors());
+        $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
+        $this->assertContains($sloResponseUrl, $targetUrl);
+        $this->assertArrayHasKey('SAMLResponse', $parsedQuery);
+        $this->assertArrayNotHasKey('RelayState', $parsedQuery);
 
         $this->_auth->setStrict(true);
         $this->_auth->processSLO(true);
         // Fail due destination missmatch
         $this->assertEquals($this->_auth->getErrors(), array('invalid_logout_request'));
 
-        try {
-            $this->_auth->setStrict(false);
-            $this->_auth->processSLO(true);
-            $this->assertFalse(true);
-        } catch (Exception $e) {
-            $this->assertContains('Cannot modify header information', $e->getMessage());
-            $trace = $e->getTrace();
-            $targetUrl = getUrlFromRedirect($trace);
-            $parsedQuery = getParamsFromUrl($targetUrl);
+        $this->_auth->setStrict(false);
+        $targetUrl = $this->_auth->processSLO(true, null, false, null, true);
+        $parsedQuery = getParamsFromUrl($targetUrl);
 
-            $this->assertEmpty($this->_auth->getErrors());
-            $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
-            $this->assertContains($sloResponseUrl, $targetUrl);
-            $this->assertArrayHasKey('SAMLResponse', $parsedQuery);
-            $this->assertArrayNotHasKey('RelayState', $parsedQuery);
-        }
+        $this->assertEmpty($this->_auth->getErrors());
+        $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
+        $this->assertContains($sloResponseUrl, $targetUrl);
+        $this->assertArrayHasKey('SAMLResponse', $parsedQuery);
+        $this->assertArrayNotHasKey('RelayState', $parsedQuery);
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Logout Request NotOnOrAfter failed
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    */
-
+     * Tests the processSLO method of the Auth class
+     * Case Logout Request NotOnOrAfter failed
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testProcessSLORequestNotOnOrAfterFailed()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_requests/invalids/not_after_failed.xml.base64');
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -674,20 +680,22 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Valid Logout Request, validating that the local session is deleted,
-    * a LogoutResponse is created and a redirection executed
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    * @runInSeparateProcess
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Valid Logout Request, validating that the local session is deleted,
+     * a LogoutResponse is created and a redirection executed
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testProcessSLORequestDeletingSession()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_deflated.xml.base64');
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -741,12 +749,15 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests the processSLO method of the OneLogin_Saml2_Auth class
-     * Case Valid Logout Request, validating that the local session is deleted with callback,
-     * a LogoutResponse is created and a redirection executed
+     * Tests the processSLO method of the Auth class
+     * Case Valid Logout Request, validating that the local session is
+     * deleted with callback, a LogoutResponse is created and
+     * a redirection executed
      *
-     * @covers OneLogin_Saml2_Auth::processSLO
+     * @covers OneLogin\Saml2\Auth::processSLO
+     *
      * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testProcessSLORequestDeletingSessionCallback()
     {
@@ -754,7 +765,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -765,7 +776,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         }
         $_SESSION['samltest'] = true;
 
-        $callback = function() {
+        $callback = function () {
             $_SESSION['samltest'] = false;
         };
 
@@ -779,8 +790,8 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
             $targetUrl = getUrlFromRedirect($trace);
             $parsedQuery = getParamsFromUrl($targetUrl);
 
-            $sloUrl = $this->_settingsInfo['idp']['singleLogoutService']['url'];
-            $this->assertContains($sloUrl, $targetUrl);
+            $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];;;
+            $this->assertContains($sloResponseUrl, $targetUrl);
             $this->assertArrayHasKey('SAMLResponse', $parsedQuery);
             $this->assertArrayNotHasKey('RelayState', $parsedQuery);
 
@@ -797,20 +808,22 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Valid Logout Request, validating the relayState,
-    * a LogoutResponse is created and a redirection executed
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    * @runInSeparateProcess
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Valid Logout Request, validating the relayState,
+     * a LogoutResponse is created and a redirection executed
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testProcessSLORequestRelayState()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_deflated.xml.base64');
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -827,8 +840,8 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
             $targetUrl = getUrlFromRedirect($trace);
             $parsedQuery = getParamsFromUrl($targetUrl);
 
-            $sloUrl = $this->_settingsInfo['idp']['singleLogoutService']['url'];
-            $this->assertContains($sloUrl, $targetUrl);
+            $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
+            $this->assertContains($sloResponseUrl, $targetUrl);
             $this->assertArrayHasKey('SAMLResponse', $parsedQuery);
             $this->assertArrayHasKey('RelayState', $parsedQuery);
             $this->assertEquals('http://relaystate.com', $parsedQuery['RelayState']);
@@ -836,13 +849,15 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the processSLO method of the OneLogin_Saml2_Auth class
-    * Case Valid Logout Request, validating the relayState,
-    * a signed LogoutResponse is created and a redirection executed
-    *
-    * @covers OneLogin_Saml2_Auth::processSLO
-    * @runInSeparateProcess
-    */
+     * Tests the processSLO method of the Auth class
+     * Case Valid Logout Request, validating the relayState,
+     * a signed LogoutResponse is created and a redirection executed
+     *
+     * @covers OneLogin\Saml2\Auth::processSLO
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testProcessSLORequestSignedResponse()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -850,13 +865,13 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['logoutResponseSigned'] = true;
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         $message = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_deflated.xml.base64');
 
         // In order to avoid the destination problem
         $plainMessage = gzinflate(base64_decode($message));
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $plainMessage = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $plainMessage);
         $message = base64_encode(gzdeflate($plainMessage));
 
@@ -873,8 +888,8 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
             $targetUrl = getUrlFromRedirect($trace);
             $parsedQuery = getParamsFromUrl($targetUrl);
 
-            $sloUrl = $settingsInfo['idp']['singleLogoutService']['url'];
-            $this->assertContains($sloUrl, $targetUrl);
+            $sloResponseUrl = $this->_settingsInfo['idp']['singleLogoutService']['responseUrl'];
+            $this->assertContains($sloResponseUrl, $targetUrl);
             $this->assertArrayHasKey('SAMLResponse', $parsedQuery);
             $this->assertArrayHasKey('RelayState', $parsedQuery);
             $this->assertArrayHasKey('SigAlg', $parsedQuery);
@@ -885,12 +900,14 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
-    * Case Login with no parameters. An AuthnRequest is built an redirection executed
-    *
-    * @covers OneLogin_Saml2_Auth::login
-    * @runInSeparateProcess
-    */
+     * Tests the login method of the Auth class
+     * Case Login with no parameters. An AuthnRequest is built an redirection executed
+     *
+     * @covers OneLogin\Saml2\Auth::login
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLogin()
     {
         try {
@@ -908,18 +925,20 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
             $this->assertContains($ssoUrl, $targetUrl);
             $this->assertArrayHasKey('SAMLRequest', $parsedQuery);
             $this->assertArrayHasKey('RelayState', $parsedQuery);
-            $this->assertEquals($parsedQuery['RelayState'], OneLogin_Saml2_Utils::getSelfRoutedURLNoQuery());
+            $this->assertEquals($parsedQuery['RelayState'], Utils::getSelfRoutedURLNoQuery());
         }
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
-    * Case Login with relayState. An AuthnRequest is built. GET with SAMLRequest,
-    * and RelayState. A redirection is executed
-    *
-    * @covers OneLogin_Saml2_Auth::login
-    * @runInSeparateProcess
-    */
+     * Tests the login method of the Auth class
+     * Case Login with relayState. An AuthnRequest is built. GET with SAMLRequest,
+     * and RelayState. A redirection is executed
+     *
+     * @covers OneLogin\Saml2\Auth::login
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLoginWithRelayState()
     {
         try {
@@ -943,18 +962,20 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
-    * Case Login with $elaySate and $parameters. An AuthnRequest is built. GET with
-    * SAMLRequest, RelayState and extra parameters in the GET. A redirection is executed
-    *
-    * @covers OneLogin_Saml2_Auth::login
-    * @runInSeparateProcess
-    */
+     * Tests the login method of the Auth class
+     * Case Login with $elaySate and $parameters. An AuthnRequest is built. GET with
+     * SAMLRequest, RelayState and extra parameters in the GET. A redirection is executed
+     *
+     * @covers OneLogin\Saml2\Auth::login
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLoginWithRelayStateAndParameters()
     {
         try {
             $relayState = 'http://sp.example.com';
-            $parameters = array ('test1' => 'value1', 'test2' => 'value2');
+            $parameters = array('test1' => 'value1', 'test2' => 'value2');
 
             // The Header of the redirect produces an Exception
             $this->_auth->login($relayState, $parameters);
@@ -979,12 +1000,14 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
-    * Case Login signed. An AuthnRequest signed is built an redirect executed
-    *
-    * @covers OneLogin_Saml2_Auth::login
-    * @runInSeparateProcess
-    */
+     * Tests the login method of the Auth class
+     * Case Login signed. An AuthnRequest signed is built an redirect executed
+     *
+     * @covers OneLogin\Saml2\Auth::login
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLoginSigned()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -992,7 +1015,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['authnRequestsSigned'] = true;
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         try {
             // The Header of the redirect produces an Exception
@@ -1018,12 +1041,14 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
-    * Case Login with no parameters. A AuthN Request is built with ForceAuthn and redirect executed
-    *
-    * @covers OneLogin_Saml2_Auth::login
-    * @runInSeparateProcess
-    */
+     * Tests the login method of the Auth class
+     * Case Login with no parameters. A AuthN Request is built with ForceAuthn and redirect executed
+     *
+     * @covers OneLogin\Saml2\Auth::login
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLoginForceAuthN()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -1031,7 +1056,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['authnRequestsSigned'] = true;
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         try {
             // The Header of the redirect produces an Exception
@@ -1100,12 +1125,14 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
-    * Case Login with no parameters. A AuthN Request is built with IsPassive and redirect executed
-    *
-    * @covers OneLogin_Saml2_Auth::login
-    * @runInSeparateProcess
-    */
+     * Tests the login method of the Auth class
+     * Case Login with no parameters. A AuthN Request is built with IsPassive and redirect executed
+     *
+     * @covers OneLogin\Saml2\Auth::login
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLoginIsPassive()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -1113,7 +1140,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['authnRequestsSigned'] = true;
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         try {
             // The Header of the redirect produces an Exception
@@ -1180,18 +1207,20 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
-    * Case Login with no parameters. A AuthN Request is built with and without NameIDPolicy
-    *
-    * @covers OneLogin_Saml2_Auth::login
-    * @runInSeparateProcess
-    */
+     * Tests the login method of the Auth class
+     * Case Login with no parameters. A AuthN Request is built with and without NameIDPolicy
+     *
+     * @covers OneLogin\Saml2\Auth::login
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLoginNameIDPolicy()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         try {
             // The Header of the redirect produces an Exception
@@ -1258,18 +1287,17 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the login method of the OneLogin_Saml2_Auth class
+    * Tests the login method of the Auth class
     * Case Login with no parameters. A AuthN Request is built with and without Subject
     *
-    * @covers OneLogin_Saml2_Auth::login
+    * @covers OneLogin\Saml2\Auth::login
     * @runInSeparateProcess
     */
     public function testLoginSubject()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
-
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         try {
             // The Header of the redirect produces an Exception
@@ -1319,7 +1347,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
             // The Header of the redirect produces an Exception
             $returnTo = 'http://example.com/returnto';
             $settingsInfo['sp']['NameIDFormat'] = "urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress";
-            $auth2 = new OneLogin_Saml2_Auth($settingsInfo);
+            $auth2 = new Auth($settingsInfo);
             $auth2->login($returnTo);
             // Do not ever get here
             $this->assertFalse(true);
@@ -1342,12 +1370,14 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the logout method of the OneLogin_Saml2_Auth class
-    * Case Logout with no parameters. A logout Request is built and redirect executed
-    *
-    * @covers OneLogin_Saml2_Auth::logout
-    * @runInSeparateProcess
-    */
+     * Tests the logout method of the Auth class
+     * Case Logout with no parameters. A logout Request is built and redirect executed
+     *
+     * @covers OneLogin\Saml2\Auth::logout
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLogout()
     {
         try {
@@ -1365,18 +1395,20 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
             $this->assertContains($sloUrl, $targetUrl);
             $this->assertArrayHasKey('SAMLRequest', $parsedQuery);
             $this->assertArrayHasKey('RelayState', $parsedQuery);
-            $this->assertEquals($parsedQuery['RelayState'], OneLogin_Saml2_Utils::getSelfRoutedURLNoQuery());
+            $this->assertEquals($parsedQuery['RelayState'], Utils::getSelfRoutedURLNoQuery());
         }
     }
 
     /**
-    * Tests the logout method of the OneLogin_Saml2_Auth class
-    * Case Logout with relayState. A logout Request is build. GET with SAMLRequest,
-    * RelayState. A redirection is executed
-    *
-    * @covers OneLogin_Saml2_Auth::logout
-    * @runInSeparateProcess
-    */
+     * Tests the logout method of the Auth class
+     * Case Logout with relayState. A logout Request is build. GET with SAMLRequest,
+     * RelayState. A redirection is executed
+     *
+     * @covers OneLogin\Saml2\Auth::logout
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLogoutWithRelayState()
     {
         try {
@@ -1400,18 +1432,20 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the logout method of the OneLogin_Saml2_Auth class
-    * Case Logout with relayState + parameters. A logout Request is build. GET with SAMLRequest,
-    * RelayState and extra parameters. A redirection is executed
-    *
-    * @covers OneLogin_Saml2_Auth::logout
-    * @runInSeparateProcess
-    */
+     * Tests the logout method of the Auth class
+     * Case Logout with relayState + parameters. A logout Request is build. GET with SAMLRequest,
+     * RelayState and extra parameters. A redirection is executed
+     *
+     * @covers OneLogin\Saml2\Auth::logout
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLogoutWithRelayStateAndParameters()
     {
         try {
             $relayState = 'http://sp.example.com';
-            $parameters = array ('test1' => 'value1', 'test2' => 'value2');
+            $parameters = array('test1' => 'value1', 'test2' => 'value2');
 
             // The Header of the redirect produces an Exception
             $this->_auth->logout($relayState, $parameters);
@@ -1436,13 +1470,15 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the logout method of the OneLogin_Saml2_Auth class
-    * Case Logout with relayState + NameID + SessionIndex. A logout Request is build. GET with SAMLRequest.
-    * A redirection is executed
-    *
-    * @covers OneLogin_Saml2_Auth::logout
-    * @runInSeparateProcess
-    */
+     * Tests the logout method of the Auth class
+     * Case Logout with relayState + NameID + SessionIndex. A logout Request is build. GET with SAMLRequest.
+     * A redirection is executed
+     *
+     * @covers OneLogin\Saml2\Auth::logout
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLogoutWithNameIdAndSessionIndex()
     {
         try {
@@ -1466,12 +1502,14 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the logout method of the OneLogin_Saml2_Auth class
-    * Case nameID loaded after process SAML Response
-    *
-    * @covers OneLogin_Saml2_Auth::logout
-    * @runInSeparateProcess
-    */
+     * Tests the logout method of the Auth class
+     * Case nameID loaded after process SAML Response
+     *
+     * @covers OneLogin\Saml2\Auth::logout
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLogoutNameID()
     {
         $message = file_get_contents(TEST_ROOT . '/data/responses/valid_response.xml.base64');
@@ -1495,19 +1533,21 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
             $this->assertArrayHasKey('SAMLRequest', $parsedQuery);
 
             $logoutRequest = gzinflate(base64_decode($parsedQuery['SAMLRequest']));
-            $nameIdFromRequest = OneLogin_Saml2_LogoutRequest::getNameId($logoutRequest);
+            $nameIdFromRequest = LogoutRequest::getNameId($logoutRequest);
             $this->assertEquals($nameIdFromResponse, $nameIdFromRequest);
         }
     }
 
     /**
-    * Tests the logout method of the OneLogin_Saml2_Auth class
-    * Case Logout signed. A logout Request signed in
-    * the assertion is built and redirect executed
-    *
-    * @covers OneLogin_Saml2_Auth::logout
-    * @runInSeparateProcess
-    */
+     * Tests the logout method of the Auth class
+     * Case Logout signed. A logout Request signed in
+     * the assertion is built and redirect executed
+     *
+     * @covers OneLogin\Saml2\Auth::logout
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testLogoutSigned()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -1515,7 +1555,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['logoutRequestSigned'] = true;
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         try {
             // The Header of the redirect produces an Exception
@@ -1541,11 +1581,11 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the logout method of the OneLogin_Saml2_Auth class
-    * Case IdP no SLO endpoint.
-    *
-    * @covers OneLogin_Saml2_Auth::logout
-    */
+     * Tests the logout method of the Auth class
+     * Case IdP no SLO endpoint.
+     *
+     * @covers OneLogin\Saml2\Auth::logout
+     */
     public function testLogoutNoSLO()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -1553,29 +1593,29 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
 
         unset($settingsInfo['idp']['singleLogoutService']);
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         try {
             $returnTo = 'http://example.com/returnto';
             $auth->logout($returnTo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('The IdP does not support Single Log Out', $e->getMessage());
         }
     }
 
     /**
-    * Tests the setStrict method of the OneLogin_Saml2_Auth
-    *
-    * @covers OneLogin_Saml2_Auth::setStrict
-    */
+     * Tests the setStrict method of the Auth
+     *
+     * @covers OneLogin\Saml2\Auth::setStrict
+     */
     public function testSetStrict()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
         $settingsInfo['strict'] = false;
 
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         $settings = $auth->getSettings();
         $this->assertFalse($settings->isStrict());
@@ -1597,30 +1637,30 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the buildRequestSignature method of the OneLogin_Saml2_Auth
-    *
-    * @covers OneLogin_Saml2_Auth::buildRequestSignature
-    */
+     * Tests the buildRequestSignature method of the Auth
+     *
+     * @covers OneLogin\Saml2\Auth::buildRequestSignature
+     */
     public function testBuildRequestSignature()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_deflated.xml.base64');
         $relayState = 'http://relaystate.com';
         $signature = $this->_auth->buildRequestSignature($message, $relayState);
-        $validSignature = 'Pb1EXAX5TyipSJ1SndEKZstLQTsT+1D00IZAhEepBM+OkAZQSToivu3njgJu47HZiZAqgXZFgloBuuWE/+GdcSsRYEMkEkiSDWTpUr25zKYLJDSg6GNo6iAHsKSuFt46Z54Xe/keYxYP03Hdy97EwuuSjBzzgRc5tmpV+KC7+a0=';
+        $validSignature = 'CqdIlbO6GieeJFV+PYqyqz1QVJunQXdZZl+ZyIby9O3/eMJM0XHi+TWReRrpgNxKkbmmvx5fp/t7mphbLiVYNMgGINEaaa/OfoaGwU9GM5YCVULA2t7qZBel1yrIXGMxijJizB7UPR2ZMo4G+Wdhx1zbmbB0GYM0A27w6YCe/+k=';
         $this->assertEquals($validSignature, $signature);
     }
 
     /**
-    * Tests the buildResponseSignature method of the OneLogin_Saml2_Auth
-    *
-    * @covers OneLogin_Saml2_Auth::buildResponseSignature
-    */
+     * Tests the buildResponseSignature method of the Auth
+     *
+     * @covers OneLogin\Saml2\Auth::buildResponseSignature
+     */
     public function testBuildResponseSignature()
     {
         $message = file_get_contents(TEST_ROOT . '/data/logout_responses/logout_response_deflated.xml.base64');
         $relayState = 'http://relaystate.com';
         $signature = $this->_auth->buildResponseSignature($message, $relayState);
-        $validSignature = 'IcyWLRX6Dz3wHBfpcUaNLVDMGM3uo6z2Z11Gjq0/APPJaHboKGljffsgMVAGBml497yckq+eYKmmz+jpURV9yTj2sF9qfD6CwX2dEzSzMdRzB40X7pWyHgEJGIhs6BhaOt5oXEk4T+h3AczERqpVYFpL00yo7FNtyQkhZFpHFhM=';
+        $validSignature = 'fFGaOuO/2+ch/xlwU5o7iS6R+v2quWchLAtiDyQTxStFQZKY1NsBs/eYIin2Meq7oTl1Ks6tpT6JshH5OwhPh/08K7M2oa6FIKb99cjg+jIJ/WwpuJ5h9SH0XXP8y3RLhCxLIomHDsBOGQK8WvOlXFUg+9nvOaEMNi6raUWrGhA=';
         $this->assertEquals($validSignature, $signature);
     }
 
@@ -1628,7 +1668,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
      * Tests that we can get most recently constructed
      * SAML AuthNRequest
      *
-     * @covers OneLogin_Saml2_Auth::getLastRequestXML()
+     * @covers OneLogin\Saml2\Auth::getLastRequestXML()
      */
     public function testGetLastAuthNRequest()
     {
@@ -1642,7 +1682,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
      * Tests that we can get most recently constructed
      * LogoutResponse.
      *
-     * @covers OneLogin_Saml2_Auth::getLastRequestXML()
+     * @covers OneLogin\Saml2\Auth::getLastRequestXML()
      */
     public function testGetLastLogoutRequestSent()
     {
@@ -1656,7 +1696,10 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
      * Tests that we can get most recently processed
      * LogoutRequest.
      *
-     * @covers OneLogin_Saml2_Auth::getLastRequestXML()
+     * @covers OneLogin\Saml2\Auth::getLastRequestXML()
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetLastLogoutRequestReceived()
     {
@@ -1670,7 +1713,10 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
      * Tests that we can get most recently processed
      * SAML Response
      *
-     * @covers OneLogin_Saml2_Auth::getLastResponseXML()
+     * @covers OneLogin\Saml2\Auth::getLastResponseXML()
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetLastSAMLResponse()
     {
@@ -1690,22 +1736,51 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
      * Tests that we can get most recently constructed
      * LogoutResponse.
      *
-     * @covers OneLogin_Saml2_Auth::getLastResponseXML()
+     * @covers OneLogin\Saml2\Auth::getLastResponseXML()
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetLastLogoutResponseSent()
     {
+        $settingsDir = TEST_ROOT .'/settings/';
+        include $settingsDir.'settings1.php';
+
         $_GET['SAMLRequest'] = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml.base64');
-        $targetSLOURL = $this->_auth->processSLO(false, null, false, null, true);
+
+        $auth = new Auth($settingsInfo);
+        $targetSLOURL = $auth->processSLO(false, null, false, null, true);
         $parsedQuery = getParamsFromUrl($targetSLOURL);
         $decodedLogoutResponse = gzinflate(base64_decode($parsedQuery['SAMLResponse']));
-        $this->assertEquals($decodedLogoutResponse, $this->_auth->getLastResponseXML());
+        $this->assertEquals($decodedLogoutResponse, $auth->getLastResponseXML());
+
+        $settingsInfo['compress'] = array(
+            'responses' => true
+        );
+        $auth2 = new Auth($settingsInfo);
+        $targetSLOURL2 = $auth2->processSLO(false, null, false, null, true);
+        $parsedQuery2 = getParamsFromUrl($targetSLOURL2);
+        $decodedLogoutResponse2 = gzinflate(base64_decode($parsedQuery2['SAMLResponse']));
+        $this->assertEquals($decodedLogoutResponse2, $auth2->getLastResponseXML());
+
+        $settingsInfo['compress'] = array(
+            'responses' => false
+        );
+        $auth3 = new Auth($settingsInfo);
+        $targetSLOURL3 = $auth3->processSLO(false, null, false, null, true);
+        $parsedQuery3 = getParamsFromUrl($targetSLOURL3);
+        $decodedLogoutResponse3 = base64_decode($parsedQuery3['SAMLResponse']);
+        $this->assertEquals($decodedLogoutResponse3, $auth3->getLastResponseXML());
     }
 
     /**
      * Tests that we can get most recently processed
      * LogoutResponse.
      *
-     * @covers OneLogin_Saml2_Auth::getLastResponseXML()
+     * @covers OneLogin\Saml2\Auth::getLastResponseXML()
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetLastLogoutResponseReceived()
     {
@@ -1719,11 +1794,12 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
      * Tests that we can get the Id of the SAMLResponse and
      * the assertion processed and the NotOnOrAfter value
      *
-     * @covers OneLogin_Saml2_Auth::getLastMessageId()
-     * @covers OneLogin_Saml2_Auth::getLastAssertionId()
-     * @covers OneLogin_Saml2_Auth::getLastAssertionNotOnOrAfter()
+     * @covers OneLogin\Saml2\Auth::getLastMessageId()
+     * @covers OneLogin\Saml2\Auth::getLastAssertionId()
+     * @covers OneLogin\Saml2\Auth::getLastAssertionNotOnOrAfter()
      *
      * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetInfoFromLastResponseReceived()
     {
@@ -1749,7 +1825,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         include $settingsDir.'settings1.php';
 
         $settingsInfo['strict'] = true;
-        $auth = new OneLogin_Saml2_Auth($settingsInfo);
+        $auth = new Auth($settingsInfo);
 
         $auth->processResponse();
         $this->assertNotEmpty($auth->getErrors());
@@ -1758,8 +1834,8 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
         $this->assertNull($auth->getLastAssertionId());
         $this->assertNull($auth->getLastAssertionNotOnOrAfter());
 
-        OneLogin_Saml2_Utils::setSelfProtocol('https');
-        OneLogin_Saml2_Utils::setSelfHost('pitbulk.no-ip.org');
+        Utils::setSelfProtocol('https');
+        Utils::setSelfHost('pitbulk.no-ip.org');
         $auth->processResponse();
         $this->assertEmpty($auth->getErrors());
         $this->assertEquals('pfx42be40bf-39c3-77f0-c6ae-8bf2e23a1a2e', $auth->getLastMessageId());
@@ -1770,7 +1846,10 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     /**
      * Tests that we can get the Id of the LogoutRequest processed
      *
-     * @covers OneLogin_Saml2_Auth::getLastMessageId()
+     * @covers OneLogin\Saml2\Auth::getLastMessageId()
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetIdFromLastLogoutRequest()
     {
@@ -1783,7 +1862,7 @@ class OneLogin_Saml2_AuthTest extends PHPUnit_Framework_TestCase
     /**
      * Tests that we can get the Id of the LogoutResponse processed
      *
-     * @covers OneLogin_Saml2_Auth::getLastMessageId()
+     * @covers OneLogin\Saml2\Auth::getLastMessageId()
      */
     public function testGetIdFromLastLogoutResponse()
     {
