@@ -1,29 +1,40 @@
 <?php
 
+namespace OneLogin\Saml2\Tests;
+
+use OneLogin\Saml2\Constants;
+use OneLogin\Saml2\Error;
+use OneLogin\Saml2\LogoutRequest;
+use OneLogin\Saml2\Settings;
+use OneLogin\Saml2\Utils;
+use OneLogin\Saml2\ValidationError;
+
+use DomDocument;
+
 /**
  * Unit tests for Logout Request
  */
-class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
+class LogoutRequestTest extends \PHPUnit\Framework\TestCase
 {
     private $_settings;
 
     /**
-    * Initializes the Test Suite
-    */
+     * Initializes the Test Suite
+     */
     public function setUp()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $this->_settings = $settings;
     }
 
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
-    *
-    * @covers OneLogin_Saml2_LogoutRequest
-    */
+     * Tests the LogoutRequest Constructor.
+     *
+     * @covers OneLogin\Saml2\LogoutRequest
+     */
     public function testConstructor()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -31,12 +42,12 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['nameIdEncrypted'] = true;
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $logoutRequest = new LogoutRequest($settings);
 
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -48,23 +59,23 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
-    *
-    * @covers OneLogin_Saml2_LogoutRequest
-    */
+     * Tests the LogoutRequest Constructor.
+     *
+     * @covers OneLogin\Saml2\LogoutRequest
+     */
     public function testConstructorWithRequest()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $encodedDeflatedRequest = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_deflated.xml.base64');
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings, $encodedDeflatedRequest);
+        $logoutRequest = new LogoutRequest($settings, $encodedDeflatedRequest);
 
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -75,22 +86,22 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
-    *
-    * @covers OneLogin_Saml2_LogoutRequest
-    */
+     * Tests the LogoutRequest Constructor.
+     *
+     * @covers OneLogin\Saml2\LogoutRequest
+     */
     public function testConstructorWithSessionIndex()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
         $sessionIndex = '_51be37965feb5579d803141076936dc2e9d1d98ebf';
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings, null, null, $sessionIndex);
+        $logoutRequest = new LogoutRequest($settings, null, null, $sessionIndex);
 
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -99,16 +110,16 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $inflated = gzinflate($decoded);
         $this->assertRegExp('#^<samlp:LogoutRequest#', $inflated);
 
-        $sessionIndexes = OneLogin_Saml2_LogoutRequest::getSessionIndexes($inflated);
+        $sessionIndexes = LogoutRequest::getSessionIndexes($inflated);
         $this->assertInternalType('array', $sessionIndexes);
         $this->assertEquals(array($sessionIndex), $sessionIndexes);
     }
 
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
-    *
-    * @covers OneLogin_Saml2_LogoutRequest
-    */
+     * Tests the LogoutRequest Constructor.
+     *
+     * @covers OneLogin\Saml2\LogoutRequest
+     */
     public function testConstructorWithNameIdFormatOnParameter()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -116,12 +127,12 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
 
         $nameId = 'test@example.com';
         $nameIdFormat = 'urn:oasis:names:tc:SAML:2.0:nameid-format:transient';
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings, null, $nameId, null, $nameIdFormat);
+        $logoutRequest = new LogoutRequest($settings, null, $nameId, null, $nameIdFormat);
 
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -130,32 +141,29 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $inflated = gzinflate($decoded);
         $this->assertRegExp('#^<samlp:LogoutRequest#', $inflated);
 
-        $logoutNameId = OneLogin_Saml2_LogoutRequest::getNameId($inflated);
+        $logoutNameId = LogoutRequest::getNameId($inflated);
         $this->assertEquals($nameId, $logoutNameId);
 
-        $logoutNameIdData = OneLogin_Saml2_LogoutRequest::getNameIdData($inflated);
+        $logoutNameIdData = LogoutRequest::getNameIdData($inflated);
         $this->assertEquals($nameIdFormat, $logoutNameIdData['Format']);
     }
 
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
+    * Tests the LogoutRequest Constructor.
     *
-    * @covers OneLogin_Saml2_LogoutRequest
+    * @covers OneLogin\Saml2\LogoutRequest
     */
     public function testConstructorWithNameIdFormatOnSettings()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
-
         $nameId = 'test@example.com';
         $nameIdFormat = 'urn:oasis:names:tc:SAML:2.0:nameid-format:transient';
         $settingsInfo['sp']['NameIDFormat'] = $nameIdFormat;
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings, null, $nameId, null, null);
-
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings, null, $nameId, null, null);
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -163,33 +171,28 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $decoded = base64_decode($payload);
         $inflated = gzinflate($decoded);
         $this->assertRegExp('#^<samlp:LogoutRequest#', $inflated);
-
-        $logoutNameId = OneLogin_Saml2_LogoutRequest::getNameId($inflated);
+        $logoutNameId = LogoutRequest::getNameId($inflated);
         $this->assertEquals($nameId, $logoutNameId);
-
-        $logoutNameIdData = OneLogin_Saml2_LogoutRequest::getNameIdData($inflated);
+        $logoutNameIdData = LogoutRequest::getNameIdData($inflated);
         $this->assertEquals($nameIdFormat, $logoutNameIdData['Format']);
     }
 
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
+    * Tests the LogoutRequest Constructor.
     *
-    * @covers OneLogin_Saml2_LogoutRequest
+    * @covers OneLogin\Saml2\LogoutRequest
     */
     public function testConstructorWithoutNameIdFormat()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
-
         $nameId = 'test@example.com';
         $nameIdFormat = 'urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified';
         $settingsInfo['sp']['NameIDFormat'] = $nameIdFormat;
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings, null, $nameId, null, null);
-
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings, null, $nameId, null, null);
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -197,33 +200,27 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $decoded = base64_decode($payload);
         $inflated = gzinflate($decoded);
         $this->assertRegExp('#^<samlp:LogoutRequest#', $inflated);
-
-        $logoutNameId = OneLogin_Saml2_LogoutRequest::getNameId($inflated);
+        $logoutNameId = LogoutRequest::getNameId($inflated);
         $this->assertEquals($nameId, $logoutNameId);
-
-        $logoutNameIdData = OneLogin_Saml2_LogoutRequest::getNameIdData($inflated);
+        $logoutNameIdData = LogoutRequest::getNameIdData($inflated);
         $this->assertFalse(isset($logoutNameIdData['Format']));
     }
-
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
+    * Tests the LogoutRequest Constructor.
     *
-    * @covers OneLogin_Saml2_LogoutRequest
+    * @covers OneLogin\Saml2\LogoutRequest
     */
     public function testConstructorWithNameIdNameQualifier()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
-
         $nameId = 'test@example.com';
         $nameIdFormat = 'urn:oasis:names:tc:SAML:2.0:nameid-format:transient';
         $nameIdNameQualifier = 'https://test.example.com/saml/metadata';
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings, null, $nameId, null, $nameIdFormat, $nameIdNameQualifier);
-
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings, null, $nameId, null, $nameIdFormat, $nameIdNameQualifier);
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -231,28 +228,25 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $decoded = base64_decode($payload);
         $inflated = gzinflate($decoded);
         $this->assertRegExp('#^<samlp:LogoutRequest#', $inflated);
-
-        $logoutNameId = OneLogin_Saml2_LogoutRequest::getNameId($inflated);
+        $logoutNameId = LogoutRequest::getNameId($inflated);
         $this->assertEquals($nameId, $logoutNameId);
-
-        $logoutNameIdData = OneLogin_Saml2_LogoutRequest::getNameIdData($inflated);
+        $logoutNameIdData = LogoutRequest::getNameIdData($inflated);
         $this->assertEquals($nameIdFormat, $logoutNameIdData['Format']);
         $this->assertEquals($nameIdNameQualifier, $logoutNameIdData['NameQualifier']);
     }
 
-
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
-    * The creation of a deflated SAML Logout Request
-    *
-    * @covers OneLogin_Saml2_LogoutRequest
-    */
+     * Tests the LogoutRequest Constructor.
+     * The creation of a deflated SAML Logout Request
+     *
+     * @covers OneLogin\Saml2\LogoutRequest
+     */
     public function testCreateDeflatedSAMLLogoutRequestURLParameter()
     {
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings);
+        $logoutRequest = new LogoutRequest($this->_settings);
 
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -263,11 +257,11 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the OneLogin_Saml2_LogoutRequest Constructor.
-    * Case: Able to generate encryptedID with MultiCert
-    *
-    * @covers OneLogin_Saml2_LogoutRequest
-    */
+     * Tests the LogoutRequest Constructor.
+     * Case: Able to generate encryptedID with MultiCert
+     *
+     * @covers OneLogin\Saml2\LogoutRequest
+     */
     public function testConstructorEncryptIdUsingX509certMulti()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -275,12 +269,12 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['nameIdEncrypted'] = true;
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $logoutRequest = new LogoutRequest($settings);
 
         $parameters = array('SAMLRequest' => $logoutRequest->getRequest());
-        $logoutUrl = OneLogin_Saml2_Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
+        $logoutUrl = Utils::redirect('http://idp.example.com/SingleLogoutService.php', $parameters, true);
         $this->assertRegExp('#^http://idp\.example\.com\/SingleLogoutService\.php\?SAMLRequest=#', $logoutUrl);
         parse_str(parse_url($logoutUrl, PHP_URL_QUERY), $exploded);
         // parse_url already urldecode de params so is not required.
@@ -292,44 +286,44 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getID method of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getID
-    */
+     * Tests the getID method of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getID
+     */
     public function testGetIDFromSAMLLogoutRequest()
     {
         $logoutRequest = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
-        $id = OneLogin_Saml2_LogoutRequest::getID($logoutRequest);
+        $id = LogoutRequest::getID($logoutRequest);
         $this->assertEquals('ONELOGIN_21584ccdfaca36a145ae990442dcd96bfe60151e', $id);
 
         $dom = new DOMDocument;
         $dom->loadXML($logoutRequest);
-        $id2 = OneLogin_Saml2_LogoutRequest::getID($dom);
+        $id2 = LogoutRequest::getID($dom);
         $this->assertEquals('ONELOGIN_21584ccdfaca36a145ae990442dcd96bfe60151e', $id2);
     }
 
     /**
-    * Tests the getID method of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getID
-    */
+     * Tests the getID method of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getID
+     */
     public function testGetIDFromDeflatedSAMLLogoutRequest()
     {
         $deflatedLogoutRequest = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_deflated.xml.base64');
         $decoded = base64_decode($deflatedLogoutRequest);
         $logoutRequest = gzinflate($decoded);
-        $id = OneLogin_Saml2_LogoutRequest::getID($logoutRequest);
+        $id = LogoutRequest::getID($logoutRequest);
         $this->assertEquals('ONELOGIN_21584ccdfaca36a145ae990442dcd96bfe60151e', $id);
     }
 
     /**
-    * Tests the getNameIdData method of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getNameIdData
-    */
+     * Tests the getNameIdData method of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getNameIdData
+     */
     public function testGetNameIdData()
     {
-        $expectedNameIdData = array (
+        $expectedNameIdData = array(
             'Value' => 'ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c',
             'Format' => 'urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified',
             'SPNameQualifier' => 'http://idp.example.com/'
@@ -337,28 +331,28 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
 
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
 
-        $nameIdData = OneLogin_Saml2_LogoutRequest::getNameIdData($request);
+        $nameIdData = LogoutRequest::getNameIdData($request);
 
         $this->assertEquals($expectedNameIdData, $nameIdData);
 
         $dom = new DOMDocument();
         $dom->loadXML($request);
-        $nameIdData2 = OneLogin_Saml2_LogoutRequest::getNameIdData($dom);
+        $nameIdData2 = LogoutRequest::getNameIdData($dom);
         $this->assertEquals($expectedNameIdData, $nameIdData2);
 
         $request2 = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_encrypted_nameid.xml');
 
         try {
-            $nameIdData3 = OneLogin_Saml2_LogoutRequest::getNameIdData($request2);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $nameIdData3 = LogoutRequest::getNameIdData($request2);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('Key is required in order to decrypt the NameID', $e->getMessage());
         }
 
         $key = $this->_settings->getSPkey();
-        $nameIdData4 = OneLogin_Saml2_LogoutRequest::getNameIdData($request2, $key);
+        $nameIdData4 = LogoutRequest::getNameIdData($request2, $key);
 
-        $expectedNameIdData = array (
+        $expectedNameIdData = array(
             'Value' => 'ONELOGIN_9c86c4542ab9d6fce07f2f7fd335287b9b3cdf69',
             'Format' => 'urn:oasis:names:tc:SAML:2.0:nameid-format:emailAddress',
             'SPNameQualifier' => 'https://pitbulk.no-ip.org/newonelogin/demo1/metadata.php'
@@ -368,27 +362,28 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
 
         $invRequest = file_get_contents(TEST_ROOT . '/data/logout_requests/invalids/no_nameId.xml');
         try {
-            $nameIdData3 = OneLogin_Saml2_LogoutRequest::getNameIdData($invRequest);
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $nameIdData3 = LogoutRequest::getNameIdData($invRequest);
+            $this->fail('ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertContains('NameID not found in the Logout Request', $e->getMessage());
         }
 
-        $idpData = $this->_settings->getIdPData();
-        $spData = $this->_settings->getSPData();
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings, null, "ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c", null, OneLogin_Saml2_Constants::NAMEID_PERSISTENT, $idpData['entityId'], $spData['entityId']);
+
+        $logoutRequest = new LogoutRequest($this->_settings, null, "ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c", null, Constants::NAMEID_PERSISTENT, $this->_settings->getIdPData()['entityId'], $this->_settings->getSPData()['entityId']);
         $logoutRequestStr = $logoutRequest->getXML();
         $this->assertContains('ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c', $logoutRequestStr);
-        $this->assertContains('Format="'.OneLogin_Saml2_Constants::NAMEID_PERSISTENT, $logoutRequestStr);
-        $this->assertContains('NameQualifier="'.$idpData['entityId'], $logoutRequestStr);
-        $this->assertContains('SPNameQualifier="'.$spData['entityId'], $logoutRequestStr);
-         $logoutRequest2 = new OneLogin_Saml2_LogoutRequest($this->_settings, null, "ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c", null, OneLogin_Saml2_Constants::NAMEID_ENTITY, $idpData['entityId'], $spData['entityId']);
+        $this->assertContains('Format="'.Constants::NAMEID_PERSISTENT, $logoutRequestStr);
+        $this->assertContains('NameQualifier="'.$this->_settings->getIdPData()['entityId'], $logoutRequestStr);
+        $this->assertContains('SPNameQualifier="'.$this->_settings->getSPData()['entityId'], $logoutRequestStr);
+
+        $logoutRequest2 = new LogoutRequest($this->_settings, null, "ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c", null, Constants::NAMEID_ENTITY, $this->_settings->getIdPData()['entityId'], $this->_settings->getSPData()['entityId']);
         $logoutRequestStr2 = $logoutRequest2->getXML();
         $this->assertContains('ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c', $logoutRequestStr2);
-        $this->assertContains('Format="'.OneLogin_Saml2_Constants::NAMEID_ENTITY, $logoutRequestStr2);
+        $this->assertContains('Format="'.Constants::NAMEID_ENTITY, $logoutRequestStr2);
         $this->assertNotContains('NameQualifier', $logoutRequestStr2);
         $this->assertNotContains('SPNameQualifier', $logoutRequestStr2);
-         $logoutRequest3 = new OneLogin_Saml2_LogoutRequest($this->_settings, null, "ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c", null, OneLogin_Saml2_Constants::NAMEID_UNSPECIFIED);
+
+        $logoutRequest3 = new LogoutRequest($this->_settings, null, "ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c", null, Constants::NAMEID_UNSPECIFIED);
         $logoutRequestStr3 = $logoutRequest3->getXML();
         $this->assertContains('ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c', $logoutRequestStr3);
         $this->assertNotContains('Format', $logoutRequestStr3);
@@ -397,74 +392,74 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getNameIdmethod of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getNameId
-    */
+     * Tests the getNameIdmethod of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getNameId
+     */
     public function testGetNameId()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
 
-        $nameId = OneLogin_Saml2_LogoutRequest::getNameId($request);
+        $nameId = LogoutRequest::getNameId($request);
         $this->assertEquals('ONELOGIN_1e442c129e1f822c8096086a1103c5ee2c7cae1c', $nameId);
 
         $request2 = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_encrypted_nameid.xml');
         try {
-            $nameId2 = OneLogin_Saml2_LogoutRequest::getNameId($request2);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $nameId2 = LogoutRequest::getNameId($request2);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('Key is required in order to decrypt the NameID', $e->getMessage());
         }
         $key = $this->_settings->getSPkey();
-        $nameId3 = OneLogin_Saml2_LogoutRequest::getNameId($request2, $key);
+        $nameId3 = LogoutRequest::getNameId($request2, $key);
         $this->assertEquals('ONELOGIN_9c86c4542ab9d6fce07f2f7fd335287b9b3cdf69', $nameId3);
     }
 
     /**
-    * Tests the getIssuer of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getIssuer
-    */
+     * Tests the getIssuer of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getIssuer
+     */
     public function testGetIssuer()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
 
-        $issuer = OneLogin_Saml2_LogoutRequest::getIssuer($request);
+        $issuer = LogoutRequest::getIssuer($request);
         $this->assertEquals('http://idp.example.com/', $issuer);
 
         $dom = new DOMDocument();
         $dom->loadXML($request);
-        $issuer2 = OneLogin_Saml2_LogoutRequest::getIssuer($dom);
+        $issuer2 = LogoutRequest::getIssuer($dom);
         $this->assertEquals('http://idp.example.com/', $issuer2);
     }
 
     /**
-    * Tests the getSessionIndexes of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getSessionIndexes
-    */
+     * Tests the getSessionIndexes of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getSessionIndexes
+     */
     public function testGetSessionIndexes()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
 
-        $sessionIndexes = OneLogin_Saml2_LogoutRequest::getSessionIndexes($request);
+        $sessionIndexes = LogoutRequest::getSessionIndexes($request);
         $this->assertEmpty($sessionIndexes);
 
         $dom = new DOMDocument();
         $dom->loadXML($request);
-        $sessionIndexes = OneLogin_Saml2_LogoutRequest::getSessionIndexes($dom);
+        $sessionIndexes = LogoutRequest::getSessionIndexes($dom);
         $this->assertEmpty($sessionIndexes);
 
         $request2 = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request_with_sessionindex.xml');
-        $sessionIndexes2 = OneLogin_Saml2_LogoutRequest::getSessionIndexes($request2);
+        $sessionIndexes2 = LogoutRequest::getSessionIndexes($request2);
         $this->assertEquals(array('_ac72a76526cb6ca19f8438e73879a0e6c8ae5131'), $sessionIndexes2);
     }
 
     /**
-    * Tests the getError method of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getError
-    */
+     * Tests the getError method of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getError
+     */
     public function testGetError()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
@@ -472,7 +467,7 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $deflatedRequest = gzdeflate($request);
         $encodedRequest = base64_encode($deflatedRequest);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest = new LogoutRequest($this->_settings, $encodedRequest);
 
         $this->assertNull($logoutRequest->getError());
 
@@ -480,43 +475,71 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $this->assertNull($logoutRequest->getError());
 
         $this->_settings->setStrict(true);
-        $logoutRequest2 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest2 = new LogoutRequest($this->_settings, $encodedRequest);
 
         $this->assertFalse($logoutRequest2->isValid());
         $this->assertContains('The LogoutRequest was received at', $logoutRequest2->getError());
     }
 
     /**
-    * Tests the isValid method of the OneLogin_Saml2_LogoutRequest
-    * Case Invalid Issuer
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::isValid
-    */
+     * Tests the getErrorException method of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getErrorException
+     */
+    public function testGetErrorException()
+    {
+        $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
+
+        $deflatedRequest = gzdeflate($request);
+        $encodedRequest = base64_encode($deflatedRequest);
+
+        $logoutRequest = new LogoutRequest($this->_settings, $encodedRequest);
+
+        $this->assertNull($logoutRequest->getError());
+
+        $this->assertTrue($logoutRequest->isValid());
+        $this->assertNull($logoutRequest->getError());
+
+        $this->_settings->setStrict(true);
+        $logoutRequest2 = new LogoutRequest($this->_settings, $encodedRequest);
+
+        $this->assertFalse($logoutRequest2->isValid());
+        $errorException = $logoutRequest2->getErrorException();
+        $this->assertContains('The LogoutRequest was received at', $errorException->getMessage());
+        $this->assertEquals($errorException->getMessage(), $logoutRequest2->getError());
+    }
+
+    /**
+     * Tests the isValid method of the LogoutRequest
+     * Case Invalid Issuer
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::isValid
+     */
     public function testIsInvalidIssuer()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/invalids/invalid_issuer.xml');
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $request = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $request);
 
         $deflatedRequest = gzdeflate($request);
         $encodedRequest = base64_encode($deflatedRequest);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest = new LogoutRequest($this->_settings, $encodedRequest);
         $this->assertTrue($logoutRequest->isValid());
 
         $this->_settings->setStrict(true);
-        $logoutRequest2 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest2 = new LogoutRequest($this->_settings, $encodedRequest);
 
         $this->assertFalse($logoutRequest2->isValid());
         $this->assertContains('Invalid issuer in the Logout Request', $logoutRequest2->getError());
     }
 
     /**
-    * Tests the isValid method of the OneLogin_Saml2_LogoutRequest
-    * Case invalid xml
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::isValid
-    */
+     * Tests the isValid method of the LogoutRequest
+     * Case invalid xml
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::isValid
+     */
     public function testIsInValidWrongXML()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -524,37 +547,37 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['wantXMLValidation'] = false;
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $settings->setStrict(false);
 
         $message = file_get_contents(TEST_ROOT . '/data/logout_requests/invalids/invalid_xml.xml.base64');
-        $response = new OneLogin_Saml2_LogoutRequest($settings, $message);
+        $response = new LogoutRequest($settings, $message);
 
         $this->assertTrue($response->isValid());
 
         $settings->setStrict(true);
-        $response2 = new OneLogin_Saml2_LogoutRequest($settings, $message);
+        $response2 = new LogoutRequest($settings, $message);
         $response2->isValid();
         $this->assertNotEquals('Invalid SAML Logout Request. Not match the saml-schema-protocol-2.0.xsd', $response2->getError());
 
         $settingsInfo['security']['wantXMLValidation'] = true;
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
         $settings2->setStrict(false);
-        $response3 = new OneLogin_Saml2_LogoutRequest($settings2, $message);
+        $response3 = new LogoutRequest($settings2, $message);
         $this->assertTrue($response3->isValid());
 
         $settings2->setStrict(true);
-        $response4 = new OneLogin_Saml2_LogoutRequest($settings2, $message);
+        $response4 = new LogoutRequest($settings2, $message);
         $this->assertFalse($response4->isValid());
         $this->assertEquals('Invalid SAML Logout Request. Not match the saml-schema-protocol-2.0.xsd', $response4->getError());
     }
 
     /**
-    * Tests the isValid method of the OneLogin_Saml2_LogoutRequest
-    * Case Invalid Destination
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::isValid
-    */
+     * Tests the isValid method of the LogoutRequest
+     * Case Invalid Destination
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::isValid
+     */
     public function testIsInvalidDestination()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
@@ -562,46 +585,46 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $deflatedRequest = gzdeflate($request);
         $encodedRequest = base64_encode($deflatedRequest);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest = new LogoutRequest($this->_settings, $encodedRequest);
         $this->assertTrue($logoutRequest->isValid());
 
         $this->_settings->setStrict(true);
-        $logoutRequest2 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest2 = new LogoutRequest($this->_settings, $encodedRequest);
 
         $this->assertFalse($logoutRequest2->isValid());
         $this->assertContains('The LogoutRequest was received at', $logoutRequest2->getError());
     }
 
     /**
-    * Tests the isValid method of the OneLogin_Saml2_LogoutRequest
-    * Case Invalid NotOnOrAfter
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::isValid
-    */
+     * Tests the isValid method of the LogoutRequest
+     * Case Invalid NotOnOrAfter
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::isValid
+     */
     public function testIsInvalidNotOnOrAfter()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/invalids/not_after_failed.xml');
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $request = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $request);
 
         $deflatedRequest = gzdeflate($request);
         $encodedRequest = base64_encode($deflatedRequest);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest = new LogoutRequest($this->_settings, $encodedRequest);
         $this->assertTrue($logoutRequest->isValid());
 
         $this->_settings->setStrict(true);
-        $logoutRequest2 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest2 = new LogoutRequest($this->_settings, $encodedRequest);
 
         $this->assertFalse($logoutRequest2->isValid());
         $this->assertEquals("Could not validate timestamp: expired. Check system clock.", $logoutRequest2->getError());
     }
 
     /**
-    * Tests the isValid method of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::isValid
-    */
+     * Tests the isValid method of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::isValid
+     */
     public function testIsValid()
     {
         $request = file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml');
@@ -609,39 +632,39 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $deflatedRequest = gzdeflate($request);
         $encodedRequest = base64_encode($deflatedRequest);
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest = new LogoutRequest($this->_settings, $encodedRequest);
         $this->assertTrue($logoutRequest->isValid());
 
         $this->_settings->setStrict(true);
-        $logoutRequest2 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest2 = new LogoutRequest($this->_settings, $encodedRequest);
         $this->assertFalse($logoutRequest2->isValid());
 
         $this->_settings->setStrict(false);
-        $logoutRequest3 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest3 = new LogoutRequest($this->_settings, $encodedRequest);
 
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
         $request2 = str_replace('http://stuff.com/endpoints/endpoints/sls.php', $currentURL, $request);
 
         $deflatedRequest2 = gzdeflate($request2);
         $encodedRequest2 = base64_encode($deflatedRequest2);
-        $logoutRequest4 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest2);
+        $logoutRequest4 = new LogoutRequest($this->_settings, $encodedRequest2);
         $this->assertTrue($logoutRequest4->isValid());
     }
 
     /**
-    * Tests that a 'true' value for compress => requests gets honored when we
-    * try to obtain the request payload from getRequest()
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getRequest()
-    */
+     * Tests that a 'true' value for compress => requests gets honored when we
+     * try to obtain the request payload from getRequest()
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getRequest()
+     */
     public function testWeCanChooseToCompressARequest()
     {
         //Test that we can compress.
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings);
         $payload = $logoutRequest->getRequest();
         $decoded = base64_decode($payload);
         $decompressed = gzinflate($decoded);
@@ -650,19 +673,19 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests that a 'false' value for compress => requests gets honored when we
-    * try to obtain the request payload from getRequest()
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::getRequest()
-    */
+     * Tests that a 'false' value for compress => requests gets honored when we
+     * try to obtain the request payload from getRequest()
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::getRequest()
+     */
     public function testWeCanChooseNotToCompressARequest()
     {
         //Test that we can choose not to compress the request payload.
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings2.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings);
         $payload = $logoutRequest->getRequest();
         $decoded = base64_decode($payload);
         $this->assertRegExp('#^<samlp:LogoutRequest#', $decoded);
@@ -673,7 +696,7 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
      * method to choose whether it should 'gzdeflate' the body
      * of the request.
      *
-     * @covers OneLogin_Saml2_LogoutRequest::getRequest()
+     * @covers OneLogin\Saml2\LogoutRequest::getRequest()
      */
     public function testWeCanChooseToDeflateARequestBody()
     {
@@ -682,8 +705,8 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         include $settingsDir.'settings1.php';
 
         //Compression is currently turned on in settings.
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings);
         $payload = $logoutRequest->getRequest(false);
         $decoded = base64_decode($payload);
         $this->assertRegExp('#^<samlp:LogoutRequest#', $decoded);
@@ -693,8 +716,8 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         include $settingsDir.'settings2.php';
 
         //Compression is currently turned off in settings.
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings);
         $payload = $logoutRequest->getRequest(true);
         $decoded = base64_decode($payload);
         $decompressed = gzinflate($decoded);
@@ -702,16 +725,16 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the isValid method of the OneLogin_Saml2_LogoutRequest
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::isValid
-    */
+     * Tests the isValid method of the LogoutRequest
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::isValid
+     */
     public function testIsInValidSign()
     {
-        $currentURL = OneLogin_Saml2_Utils::getSelfURLNoQuery();
+        $currentURL = Utils::getSelfURLNoQuery();
 
         $this->_settings->setStrict(false);
-        $_GET = array (
+        $_GET = array(
             'SAMLRequest' => 'lVLBitswEP0Vo7tjeWzJtki8LIRCYLvbNksPewmyPc6K2pJqyXQ/v1LSQlroQi/DMJr33rwZbZ2cJysezNms/gt+X9H55G2etBOXlx1ZFy2MdMoJLWd0wvfieP/xQcCGCrsYb3ozkRvI+wjpHC5eGU2Sw35HTg3lA8hqZFwWFcMKsStpxbEsxoLXeQN9OdY1VAgk+YqLC8gdCUQB7tyKB+281D6UaF6mtEiBPudcABcMXkiyD26Ulv6CevXeOpFlVvlunb5ttEmV3ZjlnGn8YTRO5qx0NuBs8kzpAd829tXeucmR5NH4J/203I8el6gFRUqbFPJnyEV51Wq30by4TLW0/9ZyarYTxt4sBsjUYLMZvRykl1Fxm90SXVkfwx4P++T4KSafVzmpUcVJ/sfSrQZJPphllv79W8WKGtLx0ir8IrVTqD1pT2MH3QAMSs4KTvui71jeFFiwirOmprwPkYW063+5uRq4urHiiC4e8hCX3J5wqAEGaPpw9XB5JmkBdeDqSlkz6CmUXdl0Qae5kv2F/1384wu3PwE=',
             'RelayState' => '_1037fbc88ec82ce8e770b2bed1119747bb812a07e6',
             'SigAlg' => 'http://www.w3.org/2000/09/xmldsig#rsa-sha1',
@@ -721,11 +744,11 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $request = gzinflate(base64_decode($_GET['SAMLRequest']));
         $encodedRequest = $_GET['SAMLRequest'];
 
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest = new LogoutRequest($this->_settings, $encodedRequest);
         $this->assertTrue($logoutRequest->isValid());
 
         $this->_settings->setStrict(true);
-        $logoutRequest2 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest2 = new LogoutRequest($this->_settings, $encodedRequest);
 
         $this->assertFalse($logoutRequest2->isValid());
         $this->assertContains('The LogoutRequest was received at', $logoutRequest2->getError());
@@ -734,7 +757,7 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $oldSignature = $_GET['Signature'];
         $_GET['Signature'] = 'vfWbbc47PkP3ejx4bjKsRX7lo9Ml1WRoE5J5owF/0mnyKHfSY6XbhO1wwjBV5vWdrUVX+xp6slHyAf4YoAsXFS0qhan6txDiZY4Oec6yE+l10iZbzvie06I4GPak4QrQ4gAyXOSzwCrRmJu4gnpeUxZ6IqKtdrKfAYRAcVf3333=';
 
-        $logoutRequest3 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest);
+        $logoutRequest3 = new LogoutRequest($this->_settings, $encodedRequest);
 
         $this->assertFalse($logoutRequest3->isValid());
         $this->assertContains('Signature validation failed. Logout Request rejected', $logoutRequest3->getError());
@@ -760,13 +783,13 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $encodedRequest2 = base64_encode($deflatedRequest2);
 
         $_GET['SAMLRequest'] = $encodedRequest2;
-        $logoutRequest4 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest2);
+        $logoutRequest4 = new LogoutRequest($this->_settings, $encodedRequest2);
 
         $this->assertFalse($logoutRequest4->isValid());
         $this->assertEquals('Signature validation failed. Logout Request rejected', $logoutRequest4->getError());
 
         $this->_settings->setStrict(false);
-        $logoutRequest5 = new OneLogin_Saml2_LogoutRequest($this->_settings, $encodedRequest2);
+        $logoutRequest5 = new LogoutRequest($this->_settings, $encodedRequest2);
 
         $this->assertFalse($logoutRequest5->isValid());
         $this->assertEquals('Signature validation failed. Logout Request rejected', $logoutRequest5->getError());
@@ -782,12 +805,12 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $settingsInfo['strict'] = true;
         $settingsInfo['security']['wantMessagesSigned'] = true;
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $_GET['SigAlg'] = $oldSigAlg;
         $oldSignature = $_GET['Signature'];
         unset($_GET['Signature']);
-        $logoutRequest6 = new OneLogin_Saml2_LogoutRequest($settings, $encodedRequest2);
+        $logoutRequest6 = new LogoutRequest($settings, $encodedRequest2);
 
         $this->assertFalse($logoutRequest6->isValid());
         $this->assertEquals('The Message of the Logout Request is not signed and the SP require it', $logoutRequest6->getError());
@@ -796,22 +819,22 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['idp']['certFingerprint'] = 'afe71c28ef740bc87425be13a2263d37971da1f9';
         unset($settingsInfo['idp']['x509cert']);
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest7 = new OneLogin_Saml2_LogoutRequest($settings2, $encodedRequest2);
+        $settings2 = new Settings($settingsInfo);
+        $logoutRequest7 = new LogoutRequest($settings2, $encodedRequest2);
 
         $this->assertFalse($logoutRequest7->isValid());
         $this->assertContains('In order to validate the sign on the Logout Request, the x509cert of the IdP is required', $logoutRequest7->getError());
     }
 
     /**
-    * Tests the isValid method of the OneLogin_Saml2_LogoutRequest
-    * Case: Using x509certMulti
-    *
-    * @covers OneLogin_Saml2_LogoutRequest::isValid
-    */
+     * Tests the isValid method of the LogoutRequest
+     * Case: Using x509certMulti
+     *
+     * @covers OneLogin\Saml2\LogoutRequest::isValid
+     */
     public function testIsValidSignUsingX509certMulti()
     {
-        $_GET = array (
+        $_GET = array(
             'SAMLRequest' => 'fZJNa+MwEIb/itHdiTz6sC0SQyEsBPoB27KHXoIsj7cGW3IlGfLzV7G7kN1DL2KYmeedmRcdgp7GWT26326JP/FzwRCz6zTaoNbKkSzeKqfDEJTVEwYVjXp9eHpUsKNq9i4640Zyh3xP6BDQx8FZkp1PR3KpqexAl72QmpUCS8SW01IiZz2TVVGD4X1VQYlAsl/oQyKPJAklPIQFzzZEbWNK0YLnlOVA3wqpQCoB7yQ7pWsGq+NKfcQ4q/0+xKXvd8ZNe7Td7AYbw10UxrCbP2aSPbv4Yl/8Qx/R3+SB5bTOoXiDQvFNvjnc7lXrIr75kh+6eYdXPc0jrkMO+/umjXhOtpxP2Q/nJx2/9+uWGbq8X1tV9NqGAW0kzaVvoe1AAJeCSWqYaUVRM2SilKKuqDTpFSlszdcK29RthVm9YriZebYdXpsLdhVAB7VJzif3haYMqqTVcl0JMBR4y+s2zak3sf/4v8l/vlHzBw==',
             'RelayState' => '_1037fbc88ec82ce8e770b2bed1119747bb812a07e6',
             'SigAlg' => 'http://www.w3.org/2000/09/xmldsig#rsa-sha1',
@@ -823,13 +846,13 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
         $settingsInfo['strict'] = true;
         $settingsInfo['security']['wantMessagesSigned'] = true;
         $encodedRequest = $_GET['SAMLRequest'];
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $settings->setBaseURL("http://stuff.com/endpoints/endpoints/");
         $_SERVER['REQUEST_URI'] = "/endpoints/endpoints/sls.php";
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings, $encodedRequest);
+        $logoutRequest = new LogoutRequest($settings, $encodedRequest);
         $valid = $logoutRequest->isValid();
         unset($_SERVER['REQUEST_URI']);
-        OneLogin_Saml2_Utils::setBaseURL(null);
+        Utils::setBaseURL(null);
         $this->assertTrue($valid);
     }
 
@@ -837,19 +860,19 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
      * Tests that we can get the request XML directly without
      * going through intermediate steps
      *
-     * @covers OneLogin_Saml2_LogoutRequest::getXML()
+     * @covers OneLogin\Saml2\LogoutRequest::getXML()
      */
     public function testGetXML()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings);
         $xml = $logoutRequest->getXML();
         $this->assertRegExp('#^<samlp:LogoutRequest#', $xml);
 
-        $logoutRequestProcessed = new OneLogin_Saml2_LogoutRequest($settings, base64_encode($xml));
+        $logoutRequestProcessed = new LogoutRequest($settings, base64_encode($xml));
         $xml2 = $logoutRequestProcessed->getXML();
         $this->assertRegExp('#^<samlp:LogoutRequest#', $xml2);
     }
@@ -857,20 +880,20 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     /**
      * Tests that we can get the ID of the LogoutRequest
      *
-     * @covers OneLogin_Saml2_LogoutRequest::getID()
+     * @covers OneLogin\Saml2\LogoutRequest::getID()
      */
     public function testGetID()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings);
         $xml = $logoutRequest->getXML();
-        $id1 = OneLogin_Saml2_LogoutRequest::getID($xml);
+        $id1 = LogoutRequest::getID($xml);
         $this->assertNotNull($id1);
 
-        $logoutRequestProcessed = new OneLogin_Saml2_LogoutRequest($settings, base64_encode($xml));
+        $logoutRequestProcessed = new LogoutRequest($settings, base64_encode($xml));
         $id2 = $logoutRequestProcessed->id;
         $this->assertEquals($id1, $id2);
     }
@@ -878,19 +901,18 @@ class OneLogin_Saml2_LogoutRequestTest extends PHPUnit_Framework_TestCase
     /**
      * Tests that the LogoutRequest throws an exception
      *
-     * @covers OneLogin_Saml2_LogoutRequest::getID()
+     * @covers OneLogin\Saml2\LogoutRequest::getID()
      *
-     * @expectedException OneLogin_Saml2_Error
+     * @expectedException OneLogin\Saml2\Error
      * @expectedExceptionMessage LogoutRequest could not be processed
      */
     public function testGetIDException()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
-
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
-        $logoutRequest = new OneLogin_Saml2_LogoutRequest($settings);
+        $settings = new Settings($settingsInfo);
+        $logoutRequest = new LogoutRequest($settings);
         $xml = $logoutRequest->getXML();
-        $id1 = OneLogin_Saml2_LogoutRequest::getID($xml.'<garbage>');
+        $id1 = LogoutRequest::getID($xml.'<garbage>');
     }
 }
