@@ -1,23 +1,32 @@
 <?php
 
+namespace OneLogin\Saml2\Tests;
+
+use OneLogin\Saml2\Error;
+use OneLogin\Saml2\Metadata;
+use OneLogin\Saml2\Settings;
+use OneLogin\Saml2\Utils;
+
+use Exception;
+
 /**
  * Unit tests for Setting class
  */
-class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
+class SettingsTest extends \PHPUnit\Framework\TestCase
 {
 
     /**
-    * Tests the OneLogin_Saml2_Settings Constructor.
-    * Case load setting from array
-    *
-    * @covers OneLogin_Saml2_Settings
-    */
+     * Tests the Settings Constructor.
+     * Case load setting from array
+     *
+     * @covers OneLogin\Saml2\Settings
+     */
     public function testLoadSettingsFromArray()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $this->assertEmpty($settings->getErrors());
 
@@ -30,202 +39,176 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         unset($settingsInfo['idp']);
 
         try {
-            $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings2 = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('Invalid array settings', $e->getMessage());
         }
 
         include $settingsDir.'settings2.php';
 
-        $settings3 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings3 = new Settings($settingsInfo);
         $this->assertEmpty($settings3->getErrors());
     }
 
     /**
-    * Tests the OneLogin_Saml2_Settings Constructor.
-    * Case load setting from OneLogin_Saml_Settings's object
-    *
-    * @covers OneLogin_Saml2_Settings
-    */
-    public function testLoadSettingsFromObject()
-    {
-        $settingsObj = new OneLogin_Saml_Settings;
-        $settingsObj->idpSingleSignOnUrl = 'http://stuff.com';
-        $settingsObj->spReturnUrl = 'http://sp.stuff.com';
-        $cert = file_get_contents(TEST_ROOT . '/data/customPath/certs/sp.crt');
-        $settingsObj->idpPublicCertificate = $cert;
-
-        $settings = new OneLogin_Saml2_Settings($settingsObj);
-
-        $this->assertEmpty($settings->getErrors());
-    }
-
-    /**
-    * Tests the OneLogin_Saml2_Settings Constructor.
-    * Case load setting from file
-    *
-    * @covers OneLogin_Saml2_Settings
-    */
+     * Tests the Settings Constructor.
+     * Case load setting from file
+     *
+     * @covers OneLogin\Saml2\Settings
+     */
     public function testLoadSettingsFromFile()
     {
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
 
         $this->assertEmpty($settings->getErrors());
     }
 
     /**
-    * Tests getCertPath method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getBasePath
-    * @covers OneLogin_Saml2_Settings::getCertPath
-    */
+     * Tests getCertPath method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getBasePath
+     * @covers OneLogin\Saml2\Settings::getCertPath
+     */
     public function testGetCertPath()
     {
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
 
         $this->assertEquals(ONELOGIN_CUSTOMPATH.'certs/', $settings->getCertPath());
     }
 
     /**
-    * Tests getLibPath method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getLibPath
-    */
+     * Tests getLibPath method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getLibPath
+     */
     public function testGetLibPath()
     {
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
         $base = $settings->getBasePath();
 
-        $this->assertEquals($base.'lib/Saml2/', $settings->getLibPath());
+        $this->assertEquals($base.'src/Saml2/', $settings->getLibPath());
     }
 
     /**
-    * Tests getExtLibPath method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getExtLibPath
-    */
-    public function testGetExtLibPath()
-    {
-        $settings = new OneLogin_Saml2_Settings();
-        $base = $settings->getBasePath();
-
-        $this->assertEquals($base.'extlib/', $settings->getExtLibPath());
-    }
-
-    /**
-    * Tests getSchemasPath method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getSchemasPath
-    */
+     * Tests getSchemasPath method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getSchemasPath
+     */
     public function testGetSchemasPath()
     {
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
         $base = $settings->getBasePath();
 
-        $this->assertEquals($base.'lib/Saml2/schemas/', $settings->getSchemasPath());
+        $this->assertEquals($base.'src/Saml2/schemas/', $settings->getSchemasPath());
 
     }
 
     /**
      * Tests getSchemasPath method of the Settings
      *
-     * @covers OneLogin_Saml2_Settings::setSchemasPath
+     * @covers OneLogin\Saml2\Settings::setSchemasPath
      */
     public function testSetSchemasPath()
     {
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
         $base = $settings->getBasePath();
-        $this->assertEquals($base.'lib/Saml2/schemas/', $settings->getSchemasPath());
+
+        $this->assertEquals($base.'src/Saml2/schemas/', $settings->getSchemasPath());
+
         $settings->setSchemasPath('custompath/');
         $this->assertEquals('custompath/', $settings->getSchemasPath());
     }
 
     /**
-    * Tests shouldCompressRequests method of OneLogin_Saml2_Settings.
-    *
-    * @covers OneLogin_Saml2_settings::shouldCompressRequests
-    */
+     * Tests shouldCompressRequests method of Settings.
+     *
+     * @covers OneLogin\Saml2\Settings::shouldCompressRequests
+     */
     public function testShouldCompressRequests()
     {
         //The default value should be true.
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
         $this->assertTrue($settings->shouldCompressRequests());
 
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
         //settings1.php contains a true value for compress => requests.
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $this->assertTrue($settings->shouldCompressRequests());
 
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings2.php';
 
         //settings2 contains a false value for compress => requests.
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $this->assertFalse($settings->shouldCompressRequests());
     }
 
     /**
-    * Tests shouldCompressResponses method of OneLogin_Saml2_Settings.
-    *
-    * @covers OneLogin_Saml2_settings::shouldCompressResponses
-    */
+     * Tests shouldCompressResponses method of Settings.
+     *
+     * @covers OneLogin\Saml2\Settings::shouldCompressResponses
+     */
     public function testShouldCompressResponses()
     {
         //The default value should be true.
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
         $this->assertTrue($settings->shouldCompressResponses());
 
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
         //settings1.php contains a true value for compress => responses.
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $this->assertTrue($settings->shouldCompressResponses());
 
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings2.php';
 
         //settings2 contains a false value for compress => responses.
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $this->assertFalse($settings->shouldCompressResponses());
     }
 
     /**
-     * Tests the checkCompressionSettings method of OneLogin_Saml2_settings.
+     * Tests the checkCompressionSettings method of Settings.
+     *
      * @dataProvider invalidCompressSettingsProvider
-     * @covers OneLogin_Saml2_settings::checkCompressionSettings
+     * @param string $invalidValue invalidCompressSettingsProvider
+     *
+     * @covers OneLogin\Saml2\Settings::checkCompressionSettings
      */
     public function testNonArrayCompressionSettingsCauseSyntaxError($invalidValue)
     {
-
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
         $settingsInfo['compress'] = $invalidValue;
 
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $expectedMessage = "Invalid array settings: invalid_syntax";
             $this->assertEquals($expectedMessage, $e->getMessage());
             return;
         }
 
-        $this->fail("An OneLogin_Saml2_Error should have been caught.");
+        $this->fail("An Error should have been caught.");
     }
 
     /**
-     * Tests the checkCompressionSettings method of OneLogin_Saml2_settings.
+     * Tests the checkCompressionSettings method of Settings.
+     *
      * @dataProvider invalidCompressSettingsProvider
-     * @covers OneLogin_Saml2_settings::checkCompressionSettings
+     * @param string $invalidValue invalidCompressSettingsProvider
+     *
+     * @covers OneLogin\Saml2\Settings::checkCompressionSettings
      */
     public function testThatOnlyBooleansCanBeUsedForCompressionSettings($invalidValue)
     {
-
         $requestsIsInvalid = false;
         $responsesIsInvalid = false;
 
@@ -234,9 +217,9 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
             include $settingsDir.'settings1.php';
 
             $settingsInfo['compress']['requests'] = $invalidValue;
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
              $expectedMessage = "Invalid array settings: 'compress'=>'requests' values must be true or false.";
              $this->assertEquals($expectedMessage, $e->getMessage());
              $requestsIsInvalid = true;
@@ -247,9 +230,9 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
             include $settingsDir.'settings1.php';
 
             $settingsInfo['compress']['responses'] = $invalidValue;
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
              $expectedMessage = "Invalid array settings: 'compress'=>'responses' values must be true or false.";
              $this->assertEquals($expectedMessage, $e->getMessage());
              $responsesIsInvalid = true;
@@ -270,16 +253,16 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the checkSPCerts method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::checkSPCerts
-    * @covers OneLogin_Saml2_Settings::getSPcert
-    * @covers OneLogin_Saml2_Settings::getSPcertNew
-    * @covers OneLogin_Saml2_Settings::getSPkey
-    */
+     * Tests the checkSPCerts method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::checkSPCerts
+     * @covers OneLogin\Saml2\Settings::getSPcert
+     * @covers OneLogin\Saml2\Settings::getSPcertNew
+     * @covers OneLogin\Saml2\Settings::getSPkey
+     */
     public function testCheckSPCerts()
     {
-        $settings = new OneLogin_Saml2_Settings();
+        $settings = new Settings();
 
         $this->assertTrue($settings->checkSPCerts());
 
@@ -287,7 +270,7 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings2.php';
 
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
         $this->assertTrue($settings2->checkSPCerts());
 
         $this->assertEquals($settings2->getSPkey(), $settings->getSPkey());
@@ -295,7 +278,7 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertNull($settings2->getSPcertNew());
 
         include $settingsDir.'settings5.php';
-        $settings3 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings3 = new Settings($settingsInfo);
         $this->assertTrue($settings3->checkSPCerts());
 
         $this->assertEquals($settings3->getSPkey(), $settings->getSPkey());
@@ -305,27 +288,27 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the checkSettings method of the OneLogin_Saml2_Settings
-    * The checkSettings method is private and is used at the constructor
-    *
-    * @covers OneLogin_Saml2_Settings::checkSettings
-    */
+     * Tests the checkSettings method of the Settings
+     * The checkSettings method is private and is used at the constructor
+     *
+     * @covers OneLogin\Saml2\Settings::checkSettings
+     */
     public function testCheckSettings()
     {
         $settingsInfo = array();
 
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('Invalid array settings: invalid_syntax', $e->getMessage());
         }
 
         $settingsInfo['strict'] = true;
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('idp_not_found', $e->getMessage());
             $this->assertContains('sp_not_found', $e->getMessage());
         }
@@ -337,9 +320,9 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $settingsInfo['security'] = array();
         $settingsInfo['security']['signMetadata'] = false;
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('idp_entityId_not_found', $e->getMessage());
             $this->assertContains('idp_sso_not_found', $e->getMessage());
             $this->assertContains('sp_entityId_not_found', $e->getMessage());
@@ -353,9 +336,9 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $settingsInfo['sp']['assertionConsumerService']['url'] = 'invalid_value';
         $settingsInfo['sp']['singleLogoutService']['url'] = 'invalid_value';
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('idp_sso_url_invalid', $e->getMessage());
             $this->assertContains('idp_slo_url_invalid', $e->getMessage());
             $this->assertContains('idp_slo_response_url_invalid', $e->getMessage());
@@ -365,17 +348,17 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
 
         $settingsInfo['security']['wantAssertionsSigned'] = true;
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('idp_cert_or_fingerprint_not_found_and_required', $e->getMessage());
         }
 
         $settingsInfo['security']['nameIdEncrypted'] = true;
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('idp_cert_not_found_and_required', $e->getMessage());
         }
 
@@ -383,36 +366,37 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         include $settingsDir.'settings1.php';
 
         $settingsInfo['security']['signMetadata']['keyFileName'] = 'metadata.key';
-        $settingsInfo['organization'] = array (
+        $settingsInfo['organization'] = array(
             'en-US' => array(
                 'name' => 'miss_information'
             )
         );
 
-        $settingsInfo['contactPerson'] = array (
-            'support' => array (
+        $settingsInfo['contactPerson'] = array(
+            'support' => array(
                 'givenName' => 'support_name'
             ),
-            'auxiliar' => array (
+            'auxiliar' => array(
                 'givenName' => 'auxiliar_name',
                 'emailAddress' => 'auxiliar@example.com',
             ),
         );
 
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $settings = new Settings($settingsInfo);
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('sp_signMetadata_invalid', $e->getMessage());
             $this->assertContains('organization_not_enought_data', $e->getMessage());
             $this->assertContains('contact_type_invalid', $e->getMessage());
         }
 
-        $settingsInfo['security']['signMetadata'] = array('privateKey' => file_get_contents(TEST_ROOT . '/data/customPath/certs/metadata.key'));
+        $settingsInfo['security']['signMetadata'] = ['privateKey' => file_get_contents(TEST_ROOT . '/data/customPath/certs/metadata.key')];
+
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
+            $settings = new Settings($settingsInfo);
             $this->fail('Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+        } catch (Error $e) {
             $this->assertContains('sp_signMetadata_invalid', $e->getMessage());
             $this->assertContains('organization_not_enought_data', $e->getMessage());
             $this->assertContains('contact_type_invalid', $e->getMessage());
@@ -420,77 +404,77 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getIdPSSOurl method of the OneLogin_Saml2_Settings class
+    * Tests the getIdPSSOurl method of the Settings class
     *
-    * @covers OneLogin_Saml2_Settings::getIdPSSOurl
+    * @covers OneLogin\Saml2\Settings::getIdPSSOurl
     */
     public function testGetIdPSSOurl()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $ssoUrl = "http://idp.example.com/SSOService.php";
         $this->assertEquals($settings->getIdPSSOUrl(), $ssoUrl);
     }
 
     /**
-    * Tests the getIdPSLOurl method of the OneLogin_Saml2_Settings class
+    * Tests the getIdPSLOurl method of the Settings class
     *
-    * @covers OneLogin_Saml2_Settings::getIdPSLOurl
+    * @covers OneLogin\Saml2\Settings::getIdPSLOurl
     */
     public function testGetIdPSLOurl()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $sloUrl = "http://idp.example.com/SingleLogoutService.php";
         $this->assertEquals($settings->getIdPSLOUrl(), $sloUrl);
 
         include $settingsDir.'settings2.php';
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
 
         $sloUrl = "http://idp.example.com/SingleLogoutService.php";
         $this->assertEquals($settings2->getIdPSLOUrl(), $sloUrl);
     }
 
     /**
-     * Tests the getIdPSLOResponseUrl method of the OneLogin_Saml2_Settings class
+     * Tests the getIdPSLOResponseUrl method of the Settings class
      *
-     * @covers OneLogin_Saml2_Settings::getIdPSLOResponseUrl
+     * @covers OneLogin\Saml2\Settings::getIdPSLOResponseUrl
      */
     public function testGetIdPSLOResponseUrl()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $sloUrl = "http://idp.example.com/SingleLogoutServiceResponse.php";
         $this->assertEquals($settings->getIdPSLOResponseUrl(), $sloUrl);
 
         include $settingsDir.'settings2.php';
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
 
         $sloUrl = "http://idp.example.com/SingleLogoutService.php";
         $this->assertEquals($settings2->getIdPSLOUrl(), $sloUrl);
     }
 
     /**
-    * Tests the getSPMetadata method of the OneLogin_Saml2_Settings
-    * Case unsigned metadata
-    *
-    * @covers OneLogin_Saml2_Settings::getSPMetadata
-    */
+     * Tests the getSPMetadata method of the Settings
+     * Case unsigned metadata
+     *
+     * @covers OneLogin\Saml2\Settings::getSPMetadata
+     */
     public function testGetSPMetadata()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = $settings->getSPMetadata();
 
@@ -506,100 +490,96 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getSPMetadata method of the OneLogin_Saml2_Settings
-    * Case with x509certNew
-    *
-    * @covers OneLogin_Saml2_Settings::getSPMetadata
-    * @dataProvider testGetSPMetadataWithX509CertNewDataProvider
-    */
+     * Tests the getSPMetadata method of the Settings
+     * Case with x509certNew
+     *
+     * @covers OneLogin\Saml2\Settings::getSPMetadata
+     * @dataProvider getSPMetadataWithX509CertNewDataProvider
+     */
     public function testGetSPMetadataWithX509CertNew($alwaysIncludeEncryption, $wantNameIdEncrypted, $wantAssertionsEncrypted, $expectEncryptionKeyDescriptor)
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings5.php';
-
         $settingsInfo['security']['wantNameIdEncrypted'] = $wantNameIdEncrypted;
         $settingsInfo['security']['wantAssertionsEncrypted'] = $wantAssertionsEncrypted;
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $metadata = $settings->getSPMetadata($alwaysIncludeEncryption);
-
         $this->assertEquals($expectEncryptionKeyDescriptor ? 4 : 2, substr_count($metadata, "<md:KeyDescriptor"));
-
         // signing KeyDescriptor should always be included
         $this->assertEquals(2, substr_count($metadata, '<md:KeyDescriptor use="signing"'));
-
         $this->assertEquals($expectEncryptionKeyDescriptor ? 2 : 0, substr_count($metadata, '<md:KeyDescriptor use="encryption"'));
     }
 
-    public function testGetSPMetadataWithX509CertNewDataProvider()
+    public function getSPMetadataWithX509CertNewDataProvider()
     {
-        return array(
-            'settings do not require encryption' => array(
+        return [
+            'settings do not require encryption' => [
                 'alwaysIncludeEncryption' => false,
                 'wantNameIdEncrypted' => false,
                 'wantAssertionsEncrypted' => false,
                 'expectEncryptionKeyDescriptor' => false,
-            ),
-            'wantNameIdEncrypted setting enabled' => array(
+            ],
+            'wantNameIdEncrypted setting enabled' => [
                 'alwaysIncludeEncryption' => false,
                 'wantNameIdEncrypted' => true,
                 'wantAssertionsEncrypted' => false,
                 'expectEncryptionKeyDescriptor' => true,
-            ),
-            'wantAssertionsEncrypted setting enabled' => array(
+            ],
+            'wantAssertionsEncrypted setting enabled' => [
                 'alwaysIncludeEncryption' => false,
                 'wantNameIdEncrypted' => false,
                 'wantAssertionsEncrypted' => true,
                 'expectEncryptionKeyDescriptor' => true,
-            ),
-            'both settings enabled'=> array(
+            ],
+            'both settings enabled'=> [
                 'alwaysIncludeEncryption' => false,
                 'wantNameIdEncrypted' => true,
                 'wantAssertionsEncrypted' => true,
                 'expectEncryptionKeyDescriptor' => true,
-            ),
-            'metadata requested with encryption' => array(
+            ],
+            'metadata requested with encryption' => [
                 'alwaysIncludeEncryption' => true,
                 'wantNameIdEncrypted' => false,
                 'wantAssertionsEncrypted' => false,
                 'expectEncryptionKeyDescriptor' => true,
-            ),
-            'metadata requested with encryption and wantNameIdEncrypted setting enabled' => array(
+            ],
+            'metadata requested with encryption and wantNameIdEncrypted setting enabled' => [
                 'alwaysIncludeEncryption' => true,
                 'wantNameIdEncrypted' => true,
                 'wantAssertionsEncrypted' => false,
                 'expectEncryptionKeyDescriptor' => true,
-            ),
-            'metadata requested with encryption and wantAssertionsEncrypted setting enabled' => array(
+            ],
+            'metadata requested with encryption and wantAssertionsEncrypted setting enabled' => [
                 'alwaysIncludeEncryption' => true,
                 'wantNameIdEncrypted' => false,
                 'wantAssertionsEncrypted' => true,
                 'expectEncryptionKeyDescriptor' => true,
-            ),
-            'metadata requested with encryption and both settings enabled' => array(
+            ],
+            'metadata requested with encryption and both settings enabled' => [
                 'alwaysIncludeEncryption' => true,
                 'wantNameIdEncrypted' => true,
                 'wantAssertionsEncrypted' => true,
                 'expectEncryptionKeyDescriptor' => true,
-            ),
-        );
+            ],
+        ];
     }
 
     /**
     * Tests the getSPMetadata method of the OneLogin_Saml2_Settings
     * Case ValidUntil CacheDuration
     *
-    * @covers OneLogin_Saml2_Settings::getSPMetadata
+    * @covers OneLogin\Saml2\Settings::getSPMetadata
     */
     public function testGetSPMetadataTiming()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
-        $currentValidUntil =  time() + OneLogin_Saml2_Metadata::TIME_VALID;
+        $currentValidUntil =  time() + Metadata::TIME_VALID;
         $currentValidUntilStr =  gmdate('Y-m-d\TH:i:s\Z', $currentValidUntil);
-        $defaultCacheDuration = OneLogin_Saml2_Metadata::TIME_CACHED;
+        $defaultCacheDuration = Metadata::TIME_CACHED;
 
         $metadata = $settings->getSPMetadata();
         $this->assertContains('validUntil="'.$currentValidUntilStr.'"', $metadata);
@@ -614,11 +594,11 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getSPMetadata method of the OneLogin_Saml2_Settings
-    * Case signed metadata
-    *
-    * @covers OneLogin_Saml2_Settings::getSPMetadata
-    */
+     * Tests the getSPMetadata method of the Settings
+     * Case signed metadata
+     *
+     * @covers OneLogin\Saml2\Settings::getSPMetadata
+     */
     public function testGetSPMetadataSigned()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -628,7 +608,7 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
             $settingsInfo['security'] = array();
         }
         $settingsInfo['security']['signMetadata'] = true;
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = $settings->getSPMetadata();
 
@@ -643,7 +623,7 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertContains('<md:NameIDFormat>urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified</md:NameIDFormat>', $metadata);
 
         $this->assertContains('<ds:SignedInfo><ds:CanonicalizationMethod Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#"/>', $metadata);
-        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2000/09/xmldsig#rsa-sha1"/>', $metadata);
+        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $metadata);
         $this->assertContains('<ds:Reference', $metadata);
         $this->assertContains('<ds:KeyInfo><ds:X509Data><ds:X509Certificate>', $metadata);
 
@@ -655,7 +635,7 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         }
         $settingsInfo['security']['signMetadata'] = true;
 
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
 
         $metadata2 = $settings2->getSPMetadata();
 
@@ -670,17 +650,18 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertContains('<md:NameIDFormat>urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified</md:NameIDFormat>', $metadata2);
 
         $this->assertContains('<ds:SignedInfo><ds:CanonicalizationMethod Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#"/>', $metadata2);
-        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2000/09/xmldsig#rsa-sha1"/>', $metadata2);
+        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $metadata2);
         $this->assertContains('<ds:Reference', $metadata2);
         $this->assertContains('<ds:KeyInfo><ds:X509Data><ds:X509Certificate>', $metadata2);
 
         $cert = file_get_contents(TEST_ROOT . '/data/customPath/certs/metadata.crt');
-        $settingsInfo['security']['signMetadata'] = array(
+        $settingsInfo['security']['signMetadata'] = [
             'privateKey' => file_get_contents(TEST_ROOT . '/data/customPath/certs/metadata.key'),
             'x509cert' => $cert,
-        );
-        $settings3 = new OneLogin_Saml2_Settings($settingsInfo);
+        ];
+        $settings3 = new Settings($settingsInfo);
         $metadata3 = $settings3->getSPMetadata();
+
         $this->assertNotEmpty($metadata3);
         $this->assertContains('<md:SPSSODescriptor', $metadata3);
         $this->assertContains('entityID="http://stuff.com/endpoints/metadata.php"', $metadata3);
@@ -689,19 +670,20 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertContains('<md:AssertionConsumerService Binding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST" Location="http://stuff.com/endpoints/endpoints/acs.php" index="1"/>', $metadata3);
         $this->assertContains('<md:SingleLogoutService Binding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect" Location="http://stuff.com/endpoints/endpoints/sls.php"/>', $metadata3);
         $this->assertContains('<md:NameIDFormat>urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified</md:NameIDFormat>', $metadata3);
+
         $this->assertContains('<ds:SignedInfo><ds:CanonicalizationMethod Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#"/>', $metadata3);
-        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2000/09/xmldsig#rsa-sha1"/>', $metadata3);
+        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $metadata3);
         $this->assertContains('<ds:Reference', $metadata3);
         $this->assertContains('<ds:KeyInfo><ds:X509Data><ds:X509Certificate>', $metadata3);
-        $this->assertContains(OneLogin_Saml2_Utils::formatCert($cert, false), $metadata3);
+        $this->assertContains(Utils::formatCert($cert, false), $metadata3);
     }
 
     /**
-    * Tests the getSPMetadata method of the OneLogin_Saml2_Settings
-    * Case signed metadata with specific certs
-    *
-    * @covers OneLogin_Saml2_Settings::getSPMetadata
-    */
+     * Tests the getSPMetadata method of the Settings
+     * Case signed metadata with specific certs
+     *
+     * @covers OneLogin\Saml2\Settings::getSPMetadata
+     */
     public function testGetSPMetadataSignedNoMetadataCert()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -710,50 +692,50 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         if (!isset($settingsInfo['security'])) {
             $settingsInfo['security'] = array();
         }
-        $settingsInfo['security']['signMetadata'] = array ();
+        $settingsInfo['security']['signMetadata'] = array();
 
         try {
-            $settings = new OneLogin_Saml2_Settings($settingsInfo);
+            $settings = new Settings($settingsInfo);
             $metadata = $settings->getSPMetadata();
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('sp_signMetadata_invalid', $e->getMessage());
         }
 
 
-        $settingsInfo['security']['signMetadata'] = array (
+        $settingsInfo['security']['signMetadata'] = array(
             'keyFileName' => 'noexist.key',
             'certFileName' => 'sp.crt'
         );
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         try {
             $metadata = $settings->getSPMetadata();
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('Private key file not found', $e->getMessage());
         }
 
-        $settingsInfo['security']['signMetadata'] = array (
+        $settingsInfo['security']['signMetadata'] = array(
             'keyFileName' => 'sp.key',
             'certFileName' => 'noexist.crt'
         );
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         try {
             $metadata = $settings->getSPMetadata();
-            $this->fail('OneLogin_Saml2_Error was not raised');
-        } catch (OneLogin_Saml2_Error $e) {
+            $this->fail('Error was not raised');
+        } catch (Error $e) {
             $this->assertContains('Public cert file not found', $e->getMessage());
         }
     }
 
 
     /**
-    * Tests the setIdPCert method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::setIdPCert
-    */
+     * Tests the setIdPCert method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::setIdPCert
+     */
     public function testSetIdPCert()
     {
         $settingsDir = TEST_ROOT .'/settings/';
@@ -763,7 +745,7 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $settingsInfo['idp']['certFingerprint'] = 'AF:E7:1C:28:EF:74:0B:C8:74:25:BE:13:A2:26:3D:37:97:1D:A1:F9';
         unset($settingsInfo['idp']['x509cert']);
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $idpData = $settings->getIdPData();
         $this->assertEquals($idpData['x509cert'], '');
 
@@ -772,22 +754,22 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertNotEquals($idpData2['x509cert'], '');
         $this->assertNotEquals($idpData2['x509cert'], $cert);
 
-        $formatedCert = OneLogin_Saml2_Utils::formatCert($cert);
+        $formatedCert = Utils::formatCert($cert);
         $this->assertEquals($idpData2['x509cert'], $formatedCert);
     }
 
     /**
-    * Tests the validateMetadata method of the OneLogin_Saml2_Settings
-    * Case valid metadata
-    *
-    * @covers OneLogin_Saml2_Settings::validateMetadata
-    */
+     * Tests the validateMetadata method of the Settings
+     * Case valid metadata
+     *
+     * @covers OneLogin\Saml2\Settings::validateMetadata
+     */
     public function testValidateMetadata()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = $settings->getSPMetadata();
 
@@ -795,17 +777,17 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the validateMetadata method of the OneLogin_Saml2_Settings
-    * Case valid signed metadata
-    *
-    * @covers OneLogin_Saml2_Settings::validateMetadata
-    */
+     * Tests the validateMetadata method of the Settings
+     * Case valid signed metadata
+     *
+     * @covers OneLogin\Saml2\Settings::validateMetadata
+     */
     public function testValidateSignedMetadata()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = file_get_contents(TEST_ROOT . '/data/metadata/signed_metadata_settings1.xml');
 
@@ -813,17 +795,17 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the validateMetadata method of the OneLogin_Saml2_Settings
-    * Case expired metadata
-    *
-    * @covers OneLogin_Saml2_Settings::validateMetadata
-    */
+     * Tests the validateMetadata method of the Settings
+     * Case expired metadata
+     *
+     * @covers OneLogin\Saml2\Settings::validateMetadata
+     */
     public function testValidateMetadataExpired()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = file_get_contents(TEST_ROOT . '/data/metadata/expired_metadata_settings1.xml');
 
@@ -833,17 +815,17 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the validateMetadata method of the OneLogin_Saml2_Settings
-    * Case no metadata
-    *
-    * @covers OneLogin_Saml2_Settings::validateMetadata
-    */
+     * Tests the validateMetadata method of the Settings
+     * Case no metadata
+     *
+     * @covers OneLogin\Saml2\Settings::validateMetadata
+     */
     public function testValidateMetadataNoXML()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = '';
         try {
@@ -861,17 +843,17 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the validateMetadata method of the OneLogin_Saml2_Settings
-    * Case invalid xml metadata: No entity
-    *
-    * @covers OneLogin_Saml2_Settings::validateMetadata
-    */
+     * Tests the validateMetadata method of the Settings
+     * Case invalid xml metadata: No entity
+     *
+     * @covers OneLogin\Saml2\Settings::validateMetadata
+     */
     public function testValidateMetadataNoEntity()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = file_get_contents(TEST_ROOT . '/data/metadata/noentity_metadata_settings1.xml');
 
@@ -881,17 +863,17 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the validateMetadata method of the OneLogin_Saml2_Settings
-    * Case invalid xml metadata: Wrong order
-    *
-    * @covers OneLogin_Saml2_Settings::validateMetadata
-    */
+     * Tests the validateMetadata method of the Settings
+     * Case invalid xml metadata: Wrong order
+     *
+     * @covers OneLogin\Saml2\Settings::validateMetadata
+     */
     public function testValidateMetadataWrongOrder()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $metadata = file_get_contents(TEST_ROOT . '/data/metadata/metadata_bad_order_settings1.xml');
 
@@ -901,16 +883,16 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getIdPData method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getIdPData
-    */
+     * Tests the getIdPData method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getIdPData
+     */
     public function testGetIdPData()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $idpData = $settings->getIdPData();
         $this->assertNotEmpty($idpData);
@@ -923,12 +905,12 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('http://idp.example.com/SSOService.php', $idpData['singleSignOnService']['url']);
         $this->assertEquals('http://idp.example.com/SingleLogoutService.php', $idpData['singleLogoutService']['url']);
         $x509cert = 'MIICgTCCAeoCCQCbOlrWDdX7FTANBgkqhkiG9w0BAQUFADCBhDELMAkGA1UEBhMCTk8xGDAWBgNVBAgTD0FuZHJlYXMgU29sYmVyZzEMMAoGA1UEBxMDRm9vMRAwDgYDVQQKEwdVTklORVRUMRgwFgYDVQQDEw9mZWlkZS5lcmxhbmcubm8xITAfBgkqhkiG9w0BCQEWEmFuZHJlYXNAdW5pbmV0dC5ubzAeFw0wNzA2MTUxMjAxMzVaFw0wNzA4MTQxMjAxMzVaMIGEMQswCQYDVQQGEwJOTzEYMBYGA1UECBMPQW5kcmVhcyBTb2xiZXJnMQwwCgYDVQQHEwNGb28xEDAOBgNVBAoTB1VOSU5FVFQxGDAWBgNVBAMTD2ZlaWRlLmVybGFuZy5ubzEhMB8GCSqGSIb3DQEJARYSYW5kcmVhc0B1bmluZXR0Lm5vMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDivbhR7P516x/S3BqKxupQe0LONoliupiBOesCO3SHbDrl3+q9IbfnfmE04rNuMcPsIxB161TdDpIesLCn7c8aPHISKOtPlAeTZSnb8QAu7aRjZq3+PbrP5uW3TcfCGPtKTytHOge/OlJbo078dVhXQ14d1EDwXJW1rRXuUt4C8QIDAQABMA0GCSqGSIb3DQEBBQUAA4GBACDVfp86HObqY+e8BUoWQ9+VMQx1ASDohBjwOsg2WykUqRXF+dLfcUH9dWR63CtZIKFDbStNomPnQz7nbK+onygwBspVEbnHuUihZq3ZUdmumQqCw4Uvs/1Uvq3orOo/WJVhTyvLgFVK2QarQ4/67OZfHd7R+POBXhophSMv1ZOo';
-        $formatedx509cert = OneLogin_Saml2_Utils::formatCert($x509cert);
+        $formatedx509cert = Utils::formatCert($x509cert);
         $this->assertEquals($formatedx509cert, $idpData['x509cert']);
 
         include $settingsDir.'settings6.php';
 
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
         $idpData2 = $settings2->getIdPData();
         $this->assertNotEmpty($idpData2);
         $this->assertArrayHasKey('x509certMulti', $idpData2);
@@ -936,7 +918,7 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertArrayHasKey('encryption', $idpData2['x509certMulti']);
 
         $x509cert2 = 'MIICbDCCAdWgAwIBAgIBADANBgkqhkiG9w0BAQ0FADBTMQswCQYDVQQGEwJ1czETMBEGA1UECAwKQ2FsaWZvcm5pYTEVMBMGA1UECgwMT25lbG9naW4gSW5jMRgwFgYDVQQDDA9pZHAuZXhhbXBsZS5jb20wHhcNMTQwOTIzMTIyNDA4WhcNNDIwMjA4MTIyNDA4WjBTMQswCQYDVQQGEwJ1czETMBEGA1UECAwKQ2FsaWZvcm5pYTEVMBMGA1UECgwMT25lbG9naW4gSW5jMRgwFgYDVQQDDA9pZHAuZXhhbXBsZS5jb20wgZ8wDQYJKoZIhvcNAQEBBQADgY0AMIGJAoGBAOWA+YHU7cvPOrBOfxCscsYTJB+kH3MaA9BFrSHFS+KcR6cw7oPSktIJxUgvDpQbtfNcOkE/tuOPBDoech7AXfvH6d7Bw7xtW8PPJ2mB5Hn/HGW2roYhxmfh3tR5SdwN6i4ERVF8eLkvwCHsNQyK2Ref0DAJvpBNZMHCpS24916/AgMBAAGjUDBOMB0GA1UdDgQWBBQ77/qVeiigfhYDITplCNtJKZTM8DAfBgNVHSMEGDAWgBQ77/qVeiigfhYDITplCNtJKZTM8DAMBgNVHRMEBTADAQH/MA0GCSqGSIb3DQEBDQUAA4GBAJO2j/1uO80E5C2PM6Fk9mzerrbkxl7AZ/mvlbOn+sNZE+VZ1AntYuG8ekbJpJtG1YfRfc7EA9mEtqvv4dhv7zBy4nK49OR+KpIBjItWB5kYvrqMLKBa32sMbgqqUqeF1ENXKjpvLSuPdfGJZA3dNa/+Dyb8GGqWe707zLyc5F8m';
-        $formatedx509cert2 = OneLogin_Saml2_Utils::formatCert($x509cert2);
+        $formatedx509cert2 = Utils::formatCert($x509cert2);
 
         $idpCertSign = $idpData2['x509certMulti']['signing'];
         $this->assertEquals($formatedx509cert2, $idpData2['x509certMulti']['signing'][0]);
@@ -945,16 +927,16 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getSPData method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getSPData
-    */
+     * Tests the getSPData method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getSPData
+     */
     public function testGetSPData()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $spData = $settings->getSPData();
         $this->assertNotEmpty($spData);
@@ -970,16 +952,16 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getSecurityData method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getSecurityData
-    */
+     * Tests the getSecurityData method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getSecurityData
+     */
     public function testGetSecurityData()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $security = $settings->getSecurityData();
         $this->assertNotEmpty($security);
@@ -996,21 +978,22 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
         $this->assertArrayHasKey('wantXMLValidation', $security);
         $this->assertArrayHasKey('wantNameId', $security);
         $this->assertArrayHasKey('relaxDestinationValidation', $security);
+        $this->assertArrayHasKey('rejectUnsolicitedResponsesWithInResponseTo', $security);
         $this->assertArrayHasKey('destinationStrictlyMatches', $security);
     }
 
     /**
-    * Tests default values of Security advanced sesettings
-    *
-    * @covers OneLogin_Saml2_Settings::getSecurityData
-    */
+     * Tests default values of Security advanced sesettings
+     *
+     * @covers OneLogin\Saml2\Settings::getSecurityData
+     */
     public function testGetDefaultSecurityValues()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
         unset($settingsInfo['security']);
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $security = $settings->getSecurityData();
         $this->assertNotEmpty($security);
@@ -1053,16 +1036,16 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getContacts method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getContacts
-    */
+     * Tests the getContacts method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getContacts
+     */
     public function testGetContacts()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $contacts = $settings->getContacts();
         $this->assertNotEmpty($contacts);
@@ -1073,16 +1056,16 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getOrganization method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::getOrganization
-    */
+     * Tests the getOrganization method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::getOrganization
+     */
     public function testGetOrganization()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $organization = $settings->getOrganization();
         $this->assertNotEmpty($organization);
@@ -1092,24 +1075,23 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the setStrict method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::setStrict
-    */
+     * Tests the setStrict method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::setStrict
+     */
     public function testSetStrict()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
-        $settingsInfo['strict'] = false;
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
+        $this->assertFalse($settings->isStrict());
+
+        $settings->setStrict(false);
         $this->assertFalse($settings->isStrict());
 
         $settings->setStrict(true);
         $this->assertTrue($settings->isStrict());
-
-        $settings->setStrict(false);
-        $this->assertFalse($settings->isStrict());
 
         try {
             $settings->setStrict('a');
@@ -1120,48 +1102,48 @@ class OneLogin_Saml2_SettingsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the isStrict method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::isStrict
-    */
+     * Tests the isStrict method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::isStrict
+     */
     public function testIsStrict()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
         unset($settingsInfo['strict']);
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $this->assertTrue($settings->isStrict());
 
         $settingsInfo['strict'] = false;
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
         $this->assertFalse($settings2->isStrict());
 
         $settingsInfo['strict'] = true;
-        $settings3 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings3 = new Settings($settingsInfo);
         $this->assertTrue($settings3->isStrict());
     }
 
     /**
-    * Tests the isDebugActive method of the OneLogin_Saml2_Settings
-    *
-    * @covers OneLogin_Saml2_Settings::isDebugActive
-    */
+     * Tests the isDebugActive method of the Settings
+     *
+     * @covers OneLogin\Saml2\Settings::isDebugActive
+     */
     public function testIsDebugActive()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
         unset($settingsInfo['debug']);
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $this->assertFalse($settings->isDebugActive());
 
         $settingsInfo['debug'] = false;
-        $settings2 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings2 = new Settings($settingsInfo);
         $this->assertFalse($settings2->isDebugActive());
 
         $settingsInfo['debug'] = true;
-        $settings3 = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings3 = new Settings($settingsInfo);
         $this->assertTrue($settings3->isDebugActive());
     }
 }
