@@ -1,65 +1,53 @@
 <?php
 
+namespace OneLogin\Saml2\Tests;
+
+use OneLogin\Saml2\Constants;
+use OneLogin\Saml2\Settings;
+use OneLogin\Saml2\Utils;
+use OneLogin\Saml2\ValidationError;
+
+use RobRichards\XMLSecLibs\XMLSecurityKey;
+use RobRichards\XMLSecLibs\XMLSecurityDSig;
+
+use DOMDocument;
+use Exception;
+
 /**
  * Unit tests for Utils class
  *
  * @backupStaticAttributes enabled
  */
-class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
+class UtilsTest extends \PHPUnit\Framework\TestCase
 {
 
     /**
-    * Tests the t method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::t
-    */
-/*
-    public function testT()
-    {
-	setlocale(LC_MESSAGES, 'en_US');
-
-        $msg = 'test';
-        $translatedMsg = OneLogin_Saml2_Utils::t($msg);
-        $this->assertEquals('test', $translatedMsg);
-
-        setlocale(LC_MESSAGES, 'es_ES');
-
-        $translatedMsg = OneLogin_Saml2_Utils::t($msg);
-        $this->assertEquals('prueba', $translatedMsg);
-
-        $newmsg = 'test2: %s';
-        $translatedMsgArgs = OneLogin_Saml2_Utils::t($newmsg, array('arg'));
-        $this->assertEquals('prueba2: arg', $translatedMsgArgs);
-    }
-*/
-
-    /**
-    * Tests the loadXML method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::loadXML
-    */
+     * Tests the loadXML method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::loadXML
+     */
     public function testLoadXML()
     {
         $dom = new DOMDocument();
 
         $metadataUnloaded = '<xml><EntityDescriptor>';
-        $res1 = OneLogin_Saml2_Utils::loadXML($dom, $metadataUnloaded);
+        $res1 = Utils::loadXML($dom, $metadataUnloaded);
         $this->assertFalse($res1);
 
         $metadataInvalid = file_get_contents(TEST_ROOT .'/data/metadata/noentity_metadata_settings1.xml');
-        $res2 = OneLogin_Saml2_Utils::loadXML($dom, $metadataInvalid);
+        $res2 = Utils::loadXML($dom, $metadataInvalid);
         $this->assertTrue($res2 instanceof DOMDocument);
 
         $metadataOk = file_get_contents(TEST_ROOT .'/data/metadata/metadata_settings1.xml');
-        $res3 = OneLogin_Saml2_Utils::loadXML($dom, $metadataOk);
+        $res3 = Utils::loadXML($dom, $metadataOk);
         $this->assertTrue($res3 instanceof DOMDocument);
     }
 
     /**
-    * Tests the loadXML method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::loadXML
-    */
+     * Tests the loadXML method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::loadXML
+     */
     public function testXMLAttacks()
     {
         $dom = new DOMDocument();
@@ -69,8 +57,8 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
                       <!ELEMENT foo ANY >
                       <!ENTITY xxe SYSTEM "file:///etc/passwd" >]><foo>&xxe;</foo>';
         try {
-            $res = OneLogin_Saml2_Utils::loadXML($dom, $attackXXE);
-            $this->assertFalse($res);
+            $res = Utils::loadXML($dom, $attackXXE);
+            $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertEquals('Detected use of DOCTYPE/ENTITY in XML, disabled to prevent XXE/XEE attacks', $e->getMessage());
         }
@@ -84,7 +72,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
                             <result>test</result>
                           </results>';
         try {
-            $res2 = OneLogin_Saml2_Utils::loadXML($dom, $xmlWithDTD);
+            $res2 = Utils::loadXML($dom, $xmlWithDTD);
             $this->assertFalse($res2);
         } catch (Exception $e) {
             $this->assertEquals('Detected use of DOCTYPE/ENTITY in XML, disabled to prevent XXE/XEE attacks', $e->getMessage());
@@ -96,7 +84,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
                         <result>This result is &harmless;</result>
                       </results>';
         try {
-            $res3 = OneLogin_Saml2_Utils::loadXML($dom, $attackXEE);
+            $res3 = Utils::loadXML($dom, $attackXEE);
             $this->assertFalse($res3);
         } catch (Exception $e) {
             $this->assertEquals('Detected use of DOCTYPE/ENTITY in XML, disabled to prevent XXE/XEE attacks', $e->getMessage());
@@ -111,7 +99,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
             'UTF-16'
         );
         try {
-            $res4 = OneLogin_Saml2_Utils::loadXML($dom, $attackXEEutf16);
+            $res4 = Utils::loadXML($dom, $attackXEEutf16);
             $this->assertFalse($res4);
         } catch (Exception $e) {
             $this->assertEquals('Detected use of DOCTYPE/ENTITY in XML, disabled to prevent XXE/XEE attacks', $e->getMessage());
@@ -119,66 +107,66 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the validateXML method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::validateXML
-    */
+     * Tests the validateXML method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::validateXML
+     */
     public function testValidateXML()
     {
         $metadataUnloaded = '<xml><EntityDescriptor>';
-        $this->assertEquals(OneLogin_Saml2_Utils::validateXML($metadataUnloaded, 'saml-schema-metadata-2.0.xsd'), 'unloaded_xml');
+        $this->assertEquals(Utils::validateXML($metadataUnloaded, 'saml-schema-metadata-2.0.xsd'), 'unloaded_xml');
 
         $metadataInvalid = file_get_contents(TEST_ROOT .'/data/metadata/noentity_metadata_settings1.xml');
-        $this->assertEquals(OneLogin_Saml2_Utils::validateXML($metadataInvalid, 'saml-schema-metadata-2.0.xsd'), 'invalid_xml');
+        $this->assertEquals(Utils::validateXML($metadataInvalid, 'saml-schema-metadata-2.0.xsd'), 'invalid_xml');
 
         $metadataExpired = file_get_contents(TEST_ROOT .'/data/metadata/expired_metadata_settings1.xml');
-        $res = OneLogin_Saml2_Utils::validateXML($metadataExpired, 'saml-schema-metadata-2.0.xsd');
+        $res = Utils::validateXML($metadataExpired, 'saml-schema-metadata-2.0.xsd');
         $this->assertTrue($res instanceof DOMDocument);
 
         $metadataOk = file_get_contents(TEST_ROOT .'/data/metadata/metadata_settings1.xml');
-        $res2 = OneLogin_Saml2_Utils::validateXML($metadataOk, 'saml-schema-metadata-2.0.xsd');
+        $res2 = Utils::validateXML($metadataOk, 'saml-schema-metadata-2.0.xsd');
         $this->assertTrue($res2 instanceof DOMDocument);
 
         $metadataBadOrder = file_get_contents(TEST_ROOT .'/data/metadata/metadata_bad_order_settings1.xml');
-        $res3 = OneLogin_Saml2_Utils::validateXML($metadataBadOrder, 'saml-schema-metadata-2.0.xsd');
+        $res3 = Utils::validateXML($metadataBadOrder, 'saml-schema-metadata-2.0.xsd');
         $this->assertFalse($res3 instanceof DOMDocument);
 
         $metadataSigned = file_get_contents(TEST_ROOT .'/data/metadata/signed_metadata_settings1.xml');
-        $res4 = OneLogin_Saml2_Utils::validateXML($metadataSigned, 'saml-schema-metadata-2.0.xsd');
+        $res4 = Utils::validateXML($metadataSigned, 'saml-schema-metadata-2.0.xsd');
         $this->assertTrue($res4 instanceof DOMDocument);
 
         $dom = new DOMDocument;
-        OneLogin_Saml2_Utils::loadXML($dom, $metadataOk);
-        $res5 = OneLogin_Saml2_Utils::validateXML($dom, 'saml-schema-metadata-2.0.xsd');
+        Utils::loadXML($dom, $metadataOk);
+        $res5 = Utils::validateXML($dom, 'saml-schema-metadata-2.0.xsd');
         $this->assertTrue($res5 instanceof DOMDocument);
     }
 
     /**
-    * Tests the formatCert method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::formatCert
-    */
+     * Tests the formatCert method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::formatCert
+     */
     public function testFormatCert()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings2.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $cert = $settingsInfo['idp']['x509cert'];
         $this->assertNotContains('-----BEGIN CERTIFICATE-----', $cert);
         $this->assertNotContains('-----END CERTIFICATE-----', $cert);
         $this->assertEquals(strlen($cert), 860);
 
-        $formatedCert1 = OneLogin_Saml2_Utils::formatCert($cert);
+        $formatedCert1 = Utils::formatCert($cert);
         $this->assertContains('-----BEGIN CERTIFICATE-----', $formatedCert1);
         $this->assertContains('-----END CERTIFICATE-----', $formatedCert1);
 
-        $formatedCert2 = OneLogin_Saml2_Utils::formatCert($cert, true);
+        $formatedCert2 = Utils::formatCert($cert, true);
         $this->assertEquals($formatedCert1, $formatedCert2);
 
 
-        $formatedCert3 = OneLogin_Saml2_Utils::formatCert($cert, false);
+        $formatedCert3 = Utils::formatCert($cert, false);
         $this->assertNotContains('-----BEGIN CERTIFICATE-----', $formatedCert3);
         $this->assertNotContains('-----END CERTIFICATE-----', $formatedCert3);
         $this->assertEquals(strlen($cert), 860);
@@ -189,15 +177,15 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $this->assertNotContains('-----END CERTIFICATE-----', $cert);
         $this->assertEquals(strlen($cert), 860);
 
-        $formatedCert4 = OneLogin_Saml2_Utils::formatCert($cert);
+        $formatedCert4 = Utils::formatCert($cert);
         $this->assertContains('-----BEGIN CERTIFICATE-----', $formatedCert4);
         $this->assertContains('-----END CERTIFICATE-----', $formatedCert4);
 
-        $formatedCert5 = OneLogin_Saml2_Utils::formatCert($cert, true);
+        $formatedCert5 = Utils::formatCert($cert, true);
         $this->assertEquals($formatedCert4, $formatedCert5);
 
 
-        $formatedCert6 = OneLogin_Saml2_Utils::formatCert($cert, false);
+        $formatedCert6 = Utils::formatCert($cert, false);
         $this->assertNotContains('-----BEGIN CERTIFICATE-----', $formatedCert6);
         $this->assertNotContains('-----END CERTIFICATE-----', $formatedCert6);
         $this->assertEquals(strlen($cert2), 860);
@@ -205,16 +193,16 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the formatPrivateKey method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::formatPrivateKey
-    */
+     * Tests the formatPrivateKey method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::formatPrivateKey
+     */
     public function testFormatPrivateKey()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings2.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $key = $settingsInfo['sp']['privateKey'];
 
@@ -222,15 +210,15 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $this->assertNotContains('-----END RSA PRIVATE KEY-----', $key);
         $this->assertEquals(strlen($key), 816);
 
-        $formatedKey1 = OneLogin_Saml2_Utils::formatPrivateKey($key);
+        $formatedKey1 = Utils::formatPrivateKey($key);
         $this->assertContains('-----BEGIN RSA PRIVATE KEY-----', $formatedKey1);
         $this->assertContains('-----END RSA PRIVATE KEY-----', $formatedKey1);
 
-        $formatedKey2 = OneLogin_Saml2_Utils::formatPrivateKey($key, true);
+        $formatedKey2 = Utils::formatPrivateKey($key, true);
         $this->assertEquals($formatedKey1, $formatedKey2);
 
 
-        $formatedKey3 = OneLogin_Saml2_Utils::formatPrivateKey($key, false);
+        $formatedKey3 = Utils::formatPrivateKey($key, false);
 
         $this->assertNotContains('-----BEGIN RSA PRIVATE KEY-----', $formatedKey3);
         $this->assertNotContains('-----END RSA PRIVATE KEY-----', $formatedKey3);
@@ -238,19 +226,19 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the redirect method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::redirect
-    */
+     * Tests the redirect method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::redirect
+     */
     public function testRedirect()
     {
         // Check relative and absolute
-        $hostname = OneLogin_Saml2_Utils::getSelfHost();
+        $hostname = Utils::getSelfHost();
         $url = "http://$hostname/example";
         $url2 = '/example';
 
-        $targetUrl = OneLogin_Saml2_Utils::redirect($url, array(), true);
-        $targetUrl2 = OneLogin_Saml2_Utils::redirect($url2, array(), true);
+        $targetUrl = Utils::redirect($url, array(), true);
+        $targetUrl2 = Utils::redirect($url2, array(), true);
 
         $this->assertEquals($targetUrl, $targetUrl2);
 
@@ -258,81 +246,90 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $url3 = "https://$hostname/example?test=true";
         $url4 = "ftp://$hostname/example";
 
-        $targetUrl3 = OneLogin_Saml2_Utils::redirect($url3, array(), true);
+        $targetUrl3 = Utils::redirect($url3, array(), true);
 
         try {
-            $targetUrl4 = OneLogin_Saml2_Utils::redirect($url4, array(), true);
+            $targetUrl4 = Utils::redirect($url4, array(), true);
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Redirect to invalid URL', $e->getMessage());
         }
 
         // Review parameter prefix
-        $parameters1 = array ('value1' => 'a');
+        $parameters1 = array('value1' => 'a');
 
-        $targetUrl5 = OneLogin_Saml2_Utils::redirect($url, $parameters1, true);
+        $targetUrl5 = Utils::redirect($url, $parameters1, true);
         $this->assertEquals("http://$hostname/example?value1=a", $targetUrl5);
 
-        $targetUrl6 = OneLogin_Saml2_Utils::redirect($url3, $parameters1, true);
+        $targetUrl6 = Utils::redirect($url3, $parameters1, true);
         $this->assertEquals("https://$hostname/example?test=true&value1=a", $targetUrl6);
 
         // Review parameters
-        $parameters2 = array (
+        $parameters2 = array(
             'alphavalue' => 'a',
-            'numvalue' => array ('1', '2'),
+            'numvalue' => array('1', '2'),
             'testing' => null,
         );
 
-        $targetUrl7 = OneLogin_Saml2_Utils::redirect($url, $parameters2, true);
+        $targetUrl7 = Utils::redirect($url, $parameters2, true);
         $this->assertEquals("http://$hostname/example?alphavalue=a&numvalue[]=1&numvalue[]=2&testing", $targetUrl7);
 
-        $parameters3 = array (
+        $parameters3 = array(
             'alphavalue' => 'a',
-            'emptynumvaluelist' => array (),
-            'numvaluelist' => array (''),
+            'emptynumvaluelist' => array(),
+            'numvaluelist' => array(''),
         );
 
-        $targetUrl8 = OneLogin_Saml2_Utils::redirect($url, $parameters3, true);
+        $targetUrl8 = Utils::redirect($url, $parameters3, true);
         $this->assertEquals("http://$hostname/example?alphavalue=a&numvaluelist[]=", $targetUrl8);
     }
 
     /**
-     * @covers OneLogin_Saml2_Utils::setSelfHost
+     * @covers OneLogin\Saml2\Utils::setSelfHost
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testSetselfhost()
     {
         $_SERVER['HTTP_HOST'] = 'example.org';
-        $this->assertEquals('example.org', OneLogin_Saml2_Utils::getSelfHost());
+        $this->assertEquals('example.org', Utils::getSelfHost());
 
-        OneLogin_Saml2_Utils::setSelfHost('example.com');
-        $this->assertEquals('example.com', OneLogin_Saml2_Utils::getSelfHost());
+        Utils::setSelfHost('example.com');
+        $this->assertEquals('example.com', Utils::getSelfHost());
     }
 
     /**
-     * @covers OneLogin_Saml2_Utils::setProxyVars()
-     * @covers OneLogin_Saml2_Utils::getProxyVars()
+     * @covers OneLogin\Saml2\Utils::setProxyVars()
+     * @covers OneLogin\Saml2\Utils::getProxyVars()
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testProxyvars()
     {
-        $this->assertFalse(OneLogin_Saml2_Utils::getProxyVars());
+        $this->assertFalse(Utils::getProxyVars());
 
-        OneLogin_Saml2_Utils::setProxyVars(true);
-        $this->assertTrue(OneLogin_Saml2_Utils::getProxyVars());
+        Utils::setProxyVars(true);
+        $this->assertTrue(Utils::getProxyVars());
 
         $_SERVER['HTTP_X_FORWARDED_PROTO'] = 'https';
         $_SERVER['SERVER_PORT'] = '80';
 
-        $this->assertTrue(OneLogin_Saml2_Utils::isHTTPS());
+        $this->assertTrue(Utils::isHTTPS());
 
-        OneLogin_Saml2_Utils::setProxyVars(false);
-        $this->assertFalse(OneLogin_Saml2_Utils::isHTTPS());
+        Utils::setProxyVars(false);
+        $this->assertFalse(Utils::isHTTPS());
     }
 
     /**
-    * Tests the getSelfHost method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::getSelfHost
-    */
+     * Tests the getSelfHost method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::getSelfHost
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testGetselfhost()
     {
         if (function_exists('gethostname')) {
@@ -341,119 +338,137 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
             $hostname = php_uname("n");
         }
 
-        $this->assertEquals($hostname, OneLogin_Saml2_Utils::getSelfHost());
+        $this->assertEquals($hostname, Utils::getSelfHost());
 
         $_SERVER['SERVER_NAME'] = 'example.com';
-        $this->assertEquals('example.com', OneLogin_Saml2_Utils::getSelfHost());
+        $this->assertEquals('example.com', Utils::getSelfHost());
 
         $_SERVER['HTTP_HOST'] = 'example.org';
-        $this->assertEquals('example.org', OneLogin_Saml2_Utils::getSelfHost());
+        $this->assertEquals('example.org', Utils::getSelfHost());
 
         $_SERVER['HTTP_HOST'] = 'example.org:443';
-        $this->assertEquals('example.org', OneLogin_Saml2_Utils::getSelfHost());
+        $this->assertEquals('example.org', Utils::getSelfHost());
 
         $_SERVER['HTTP_HOST'] = 'example.org:ok';
-        $this->assertEquals('example.org', OneLogin_Saml2_Utils::getSelfHost());
+        $this->assertEquals('example.org', Utils::getSelfHost());
 
         $_SERVER['HTTP_X_FORWARDED_HOST'] = 'example.net';
-        $this->assertNotEquals('example.net', OneLogin_Saml2_Utils::getSelfHost());
+        $this->assertNotEquals('example.net', Utils::getSelfHost());
 
-        OneLogin_Saml2_Utils::setProxyVars(true);
-        $this->assertEquals('example.net', OneLogin_Saml2_Utils::getSelfHost());
+        Utils::setProxyVars(true);
+        $this->assertEquals('example.net', Utils::getSelfHost());
     }
 
     /**
-    * Tests the isHTTPS method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::isHTTPS
-    */
+     * Tests the isHTTPS method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::isHTTPS
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testisHTTPS()
     {
-        $this->assertFalse(OneLogin_Saml2_Utils::isHTTPS());
+        $this->assertFalse(Utils::isHTTPS());
 
         $_SERVER['HTTPS'] = 'on';
-        $this->assertTrue(OneLogin_Saml2_Utils::isHTTPS());
+        $this->assertTrue(Utils::isHTTPS());
 
         unset($_SERVER['HTTPS']);
-        $this->assertFalse(OneLogin_Saml2_Utils::isHTTPS());
+        $this->assertFalse(Utils::isHTTPS());
         $_SERVER['HTTP_HOST'] = 'example.com:443';
-        $this->assertTrue(OneLogin_Saml2_Utils::isHTTPS());
+        $this->assertTrue(Utils::isHTTPS());
     }
 
     /**
-     * @covers OneLogin_Saml2_Utils::getSelfURLhost
+     * @covers OneLogin\Saml2\Utils::getSelfURLhost
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetselfurlhostdoubleport()
     {
-        OneLogin_Saml2_Utils::setProxyVars(true);
+        Utils::setProxyVars(true);
         $_SERVER['HTTP_HOST'] = 'example.com:8080';
         $_SERVER['HTTP_X_FORWARDED_PORT'] = 82;
-        $this->assertEquals('http://example.com:82', OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals('http://example.com:82', Utils::getSelfURLhost());
 
         $_SERVER['HTTP_HOST'] = 'example.com:ok';
         $_SERVER['HTTP_X_FORWARDED_PORT'] = 82;
-        $this->assertEquals('http://example.com:82', OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals('http://example.com:82', Utils::getSelfURLhost());
     }
 
     /**
-     * @covers OneLogin_Saml2_Utils::getSelfPort
+     * @covers OneLogin\Saml2\Utils::getSelfPort
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testGetselfPort()
     {
-        $this->assertNull(OneLogin_Saml2_Utils::getSelfPort());
+        $this->assertNull(Utils::getSelfPort());
 
         $_SERVER['HTTP_HOST'] = 'example.org:ok';
-        $this->assertNull(OneLogin_Saml2_Utils::getSelfPort());
+        $this->assertNull(Utils::getSelfPort());
 
         $_SERVER['HTTP_HOST'] = 'example.org:8080';
-        $this->assertEquals(8080, OneLogin_Saml2_Utils::getSelfPort());
+        $this->assertEquals(8080, Utils::getSelfPort());
 
         $_SERVER["SERVER_PORT"] = 80;
-        $this->assertEquals(80, OneLogin_Saml2_Utils::getSelfPort());
+        $this->assertEquals(80, Utils::getSelfPort());
 
         $_SERVER["HTTP_X_FORWARDED_PORT"] = 443;
-        $this->assertEquals(80, OneLogin_Saml2_Utils::getSelfPort());
+        $this->assertEquals(80, Utils::getSelfPort());
 
-        OneLogin_Saml2_Utils::setProxyVars(true);
-        $this->assertEquals(443, OneLogin_Saml2_Utils::getSelfPort());
+        Utils::setProxyVars(true);
+        $this->assertEquals(443, Utils::getSelfPort());
 
-        OneLogin_Saml2_Utils::setSelfPort(8080);
-        $this->assertEquals(8080, OneLogin_Saml2_Utils::getSelfPort());
+        Utils::setSelfPort(8080);
+        $this->assertEquals(8080, Utils::getSelfPort());
     }
 
     /**
-     * @covers OneLogin_Saml2_Utils::setSelfProtocol
+     * @covers OneLogin\Saml2\Utils::setSelfProtocol
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testSetselfprotocol()
     {
-        $this->assertFalse(OneLogin_Saml2_Utils::isHTTPS());
+        $this->assertFalse(Utils::isHTTPS());
 
-        OneLogin_Saml2_Utils::setSelfProtocol('https');
-        $this->assertTrue(OneLogin_Saml2_Utils::isHTTPS());
+        Utils::setSelfProtocol('https');
+        $this->assertTrue(Utils::isHTTPS());
     }
 
     /**
-     * @covers OneLogin_Saml2_Utils::setBaseURLPath
+     * @covers OneLogin\Saml2\Utils::setBaseURLPath
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testSetBaseURLPath()
     {
-        $this->assertNull(OneLogin_Saml2_Utils::getBaseURLPath());
+        $this->assertNull(Utils::getBaseURLPath());
 
-        OneLogin_Saml2_Utils::setBaseURLPath('sp');
-        $this->assertEquals('/sp/', OneLogin_Saml2_Utils::getBaseURLPath());
+        Utils::setBaseURLPath('sp');
+        $this->assertEquals('/sp/', Utils::getBaseURLPath());
 
-        OneLogin_Saml2_Utils::setBaseURLPath('sp/');
-        $this->assertEquals('/sp/', OneLogin_Saml2_Utils::getBaseURLPath());
+        Utils::setBaseURLPath('sp/');
+        $this->assertEquals('/sp/', Utils::getBaseURLPath());
 
-        OneLogin_Saml2_Utils::setBaseURLPath('/sp');
-        $this->assertEquals('/sp/', OneLogin_Saml2_Utils::getBaseURLPath());
+        Utils::setBaseURLPath('/sp');
+        $this->assertEquals('/sp/', Utils::getBaseURLPath());
 
-        OneLogin_Saml2_Utils::setBaseURLPath('/sp/');
-        $this->assertEquals('/sp/', OneLogin_Saml2_Utils::getBaseURLPath());
+        Utils::setBaseURLPath('/sp/');
+        $this->assertEquals('/sp/', Utils::getBaseURLPath());
     }
 
     /**
-     * @covers OneLogin_Saml2_Utils::setBaseURL
+     * @covers OneLogin\Saml2\Utils::setBaseURL
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
      */
     public function testSetBaseURL()
     {
@@ -468,157 +483,169 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $expectedRoutedUrlNQ = 'https://sp.example.com/example1/route.php';
         $expectedUrl = 'https://sp.example.com/example1/route.php?x=test';
 
-        OneLogin_Saml2_Utils::setBaseURL("no-valid-url");
-        $this->assertEquals('https', OneLogin_Saml2_Utils::getSelfProtocol());
-        $this->assertEquals('sp.example.com', OneLogin_Saml2_Utils::getSelfHost());
-        $this->assertNull(OneLogin_Saml2_Utils::getSelfPort());
-        $this->assertNull(OneLogin_Saml2_Utils::getBaseURLPath());
+        Utils::setBaseURL("no-valid-url");
+        $this->assertEquals('https', Utils::getSelfProtocol());
+        $this->assertEquals('sp.example.com', Utils::getSelfHost());
+        $this->assertNull(Utils::getSelfPort());
+        $this->assertNull(Utils::getBaseURLPath());
 
-        $this->assertEquals($expectedUrlNQ, OneLogin_Saml2_Utils::getSelfURLNoQuery());
-        $this->assertEquals($expectedRoutedUrlNQ, OneLogin_Saml2_Utils::getSelfRoutedURLNoQuery());
-        $this->assertEquals($expectedUrl, OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($expectedUrlNQ, Utils::getSelfURLNoQuery());
+        $this->assertEquals($expectedRoutedUrlNQ, Utils::getSelfRoutedURLNoQuery());
+        $this->assertEquals($expectedUrl, Utils::getSelfURL());
 
-        OneLogin_Saml2_Utils::setBaseURL("http://anothersp.example.com:81/example2/");
+        Utils::setBaseURL("http://anothersp.example.com:81/example2/");
         $expectedUrlNQ2 = 'http://anothersp.example.com:81/example2/route.php';
         $expectedRoutedUrlNQ2 = 'http://anothersp.example.com:81/example2/route.php';
         $expectedUrl2 = 'http://anothersp.example.com:81/example2/route.php?x=test';
 
-        $this->assertEquals('http', OneLogin_Saml2_Utils::getSelfProtocol());
-        $this->assertEquals('anothersp.example.com', OneLogin_Saml2_Utils::getSelfHost());
-        $this->assertEquals('81', OneLogin_Saml2_Utils::getSelfPort());
-        $this->assertEquals('/example2/', OneLogin_Saml2_Utils::getBaseURLPath());
+        $this->assertEquals('http', Utils::getSelfProtocol());
+        $this->assertEquals('anothersp.example.com', Utils::getSelfHost());
+        $this->assertEquals('81', Utils::getSelfPort());
+        $this->assertEquals('/example2/', Utils::getBaseURLPath());
 
-        $this->assertEquals($expectedUrlNQ2, OneLogin_Saml2_Utils::getSelfURLNoQuery());
-        $this->assertEquals($expectedRoutedUrlNQ2, OneLogin_Saml2_Utils::getSelfRoutedURLNoQuery());
-        $this->assertEquals($expectedUrl2, OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($expectedUrlNQ2, Utils::getSelfURLNoQuery());
+        $this->assertEquals($expectedRoutedUrlNQ2, Utils::getSelfRoutedURLNoQuery());
+        $this->assertEquals($expectedUrl2, Utils::getSelfURL());
 
         $_SERVER['PATH_INFO'] = '/test';
         $expectedUrlNQ2 = 'http://anothersp.example.com:81/example2/route.php/test';
 
-        $this->assertEquals($expectedUrlNQ2, OneLogin_Saml2_Utils::getSelfURLNoQuery());
-        $this->assertEquals($expectedRoutedUrlNQ2, OneLogin_Saml2_Utils::getSelfRoutedURLNoQuery());
-        $this->assertEquals($expectedUrl2, OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($expectedUrlNQ2, Utils::getSelfURLNoQuery());
+        $this->assertEquals($expectedRoutedUrlNQ2, Utils::getSelfRoutedURLNoQuery());
+        $this->assertEquals($expectedUrl2, Utils::getSelfURL());
     }
 
     /**
-    * Tests the getSelfURLhost method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::getSelfURLhost
-    */
+     * Tests the getSelfURLhost method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::getSelfURLhost
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testGetSelfURLhost()
     {
-        $hostname = OneLogin_Saml2_Utils::getSelfHost();
+        $hostname = Utils::getSelfHost();
 
-        $this->assertEquals("http://$hostname", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("http://$hostname", Utils::getSelfURLhost());
 
         $_SERVER['SERVER_PORT'] = '80';
-        $this->assertEquals("http://$hostname", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("http://$hostname", Utils::getSelfURLhost());
 
         $_SERVER['SERVER_PORT'] = '81';
-        $this->assertEquals("http://$hostname:81", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("http://$hostname:81", Utils::getSelfURLhost());
 
         $_SERVER['SERVER_PORT'] = '443';
-        $this->assertEquals("https://$hostname", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("https://$hostname", Utils::getSelfURLhost());
 
         unset($_SERVER['SERVER_PORT']);
         $_SERVER['HTTPS'] = 'on';
-        $this->assertEquals("https://$hostname", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("https://$hostname", Utils::getSelfURLhost());
 
         $_SERVER['SERVER_PORT'] = '444';
-        $this->assertEquals("https://$hostname:444", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("https://$hostname:444", Utils::getSelfURLhost());
 
         $_SERVER['SERVER_PORT'] = '443';
         $_SERVER['REQUEST_URI'] = '/onelogin';
-        $this->assertEquals("https://$hostname", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("https://$hostname", Utils::getSelfURLhost());
 
         $_SERVER['REQUEST_URI'] = 'https://$hostname/onelogin/sso';
-        $this->assertEquals("https://$hostname", OneLogin_Saml2_Utils::getSelfURLhost());
+        $this->assertEquals("https://$hostname", Utils::getSelfURLhost());
     }
 
     /**
-    * Tests the getSelfURL method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::getSelfURL
-    */
+     * Tests the getSelfURL method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::getSelfURL
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testGetSelfURL()
     {
-        $url = OneLogin_Saml2_Utils::getSelfURLhost();
+        $url = Utils::getSelfURLhost();
 
-        $this->assertEquals($url, OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($url, Utils::getSelfURL());
 
         $_SERVER['REQUEST_URI'] = '/index.php';
-        $this->assertEquals($url.'/index.php', OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($url.'/index.php', Utils::getSelfURL());
 
         $_SERVER['REQUEST_URI'] = '/test/index.php?testing';
-        $this->assertEquals($url.'/test/index.php?testing', OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($url.'/test/index.php?testing', Utils::getSelfURL());
 
         $_SERVER['REQUEST_URI'] = '/test/index.php?testing';
-        $this->assertEquals($url.'/test/index.php?testing', OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($url.'/test/index.php?testing', Utils::getSelfURL());
 
         $_SERVER['REQUEST_URI'] = 'https://example.com/testing';
-        $this->assertEquals($url.'/testing', OneLogin_Saml2_Utils::getSelfURL());
+        $this->assertEquals($url.'/testing', Utils::getSelfURL());
     }
 
     /**
-    * Tests the getSelfURLNoQuery method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::getSelfURLNoQuery
-    */
+     * Tests the getSelfURLNoQuery method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::getSelfURLNoQuery
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function testGetSelfURLNoQuery()
     {
-        $url = OneLogin_Saml2_Utils::getSelfURLhost();
+        $url = Utils::getSelfURLhost();
         $url .= $_SERVER['SCRIPT_NAME'];
 
-        $this->assertEquals($url, OneLogin_Saml2_Utils::getSelfURLNoQuery());
+        $this->assertEquals($url, Utils::getSelfURLNoQuery());
 
         $_SERVER['PATH_INFO'] = '/test';
-        $this->assertEquals($url.'/test', OneLogin_Saml2_Utils::getSelfURLNoQuery());
+        $this->assertEquals($url.'/test', Utils::getSelfURLNoQuery());
     }
 
     /**
-    * Tests the getSelfRoutedURLNoQuery method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::getSelfRoutedURLNoQuery
-    */
+     * Tests the getSelfRoutedURLNoQuery method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::getSelfRoutedURLNoQuery
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
     public function getSelfRoutedURLNoQuery()
     {
-        $url = OneLogin_Saml2_Utils::getSelfURLhost();
+        $url = Utils::getSelfURLhost();
         $_SERVER['REQUEST_URI'] = 'example1/route?x=test';
         $_SERVER['QUERY_STRING'] = '?x=test';
 
         $url .= 'example1/route';
 
-        $this->assertEquals($url, OneLogin_Saml2_Utils::getSelfRoutedURLNoQuery());
+        $this->assertEquals($url, Utils::getSelfRoutedURLNoQuery());
     }
 
     /**
-    * Gets the status of a message
-    *
-    * @covers OneLogin_Saml2_Utils::getStatus
-    */
+     * Gets the status of a message
+     *
+     * @covers OneLogin\Saml2\Utils::getStatus
+     */
     public function testGetStatus()
     {
         $xml = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/response1.xml.base64'));
         $dom = new DOMDocument();
         $dom->loadXML($xml);
 
-        $status = OneLogin_Saml2_Utils::getStatus($dom);
-        $this->assertEquals(OneLogin_Saml2_Constants::STATUS_SUCCESS, $status['code']);
+        $status = Utils::getStatus($dom);
+        $this->assertEquals(Constants::STATUS_SUCCESS, $status['code']);
 
         $xml2 = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/invalids/status_code_responder.xml.base64'));
         $dom2 = new DOMDocument();
         $dom2->loadXML($xml2);
 
-        $status2 = OneLogin_Saml2_Utils::getStatus($dom2);
-        $this->assertEquals(OneLogin_Saml2_Constants::STATUS_RESPONDER, $status2['code']);
+        $status2 = Utils::getStatus($dom2);
+        $this->assertEquals(Constants::STATUS_RESPONDER, $status2['code']);
         $this->assertEmpty($status2['msg']);
 
         $xml3 = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/invalids/status_code_responer_and_msg.xml.base64'));
         $dom3 = new DOMDocument();
         $dom3->loadXML($xml3);
 
-        $status3 = OneLogin_Saml2_Utils::getStatus($dom3);
-        $this->assertEquals(OneLogin_Saml2_Constants::STATUS_RESPONDER, $status3['code']);
+        $status3 = Utils::getStatus($dom3);
+        $this->assertEquals(Constants::STATUS_RESPONDER, $status3['code']);
         $this->assertEquals('something_is_wrong', $status3['msg']);
 
         $xmlInv = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/invalids/no_status.xml.base64'));
@@ -626,9 +653,9 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $domInv->loadXML($xmlInv);
 
         try {
-            $statusInv = OneLogin_Saml2_Utils::getStatus($domInv);
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $statusInv = Utils::getStatus($domInv);
+            $this->fail('ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertEquals('Missing Status on response', $e->getMessage());
         }
 
@@ -637,60 +664,60 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $domInv2->loadXML($xmlInv2);
 
         try {
-            $statusInv2 = OneLogin_Saml2_Utils::getStatus($domInv2);
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $statusInv2 = Utils::getStatus($domInv2);
+            $this->fail('ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertEquals('Missing Status Code on response', $e->getMessage());
         }
     }
 
     /**
-    * Tests the parseDuration method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::parseDuration
-    */
+     * Tests the parseDuration method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::parseDuration
+     */
     public function testParseDuration()
     {
         $duration = 'PT1393462294S';
         $timestamp = 1393876825;
 
-        $parsedDuration = OneLogin_Saml2_Utils::parseDuration($duration, $timestamp);
+        $parsedDuration = Utils::parseDuration($duration, $timestamp);
         $this->assertEquals(2787339119, $parsedDuration);
 
-        $parsedDuration2 = OneLogin_Saml2_Utils::parseDuration($duration);
+        $parsedDuration2 = Utils::parseDuration($duration);
 
         $this->assertTrue($parsedDuration2 > $parsedDuration);
 
         $invalidDuration = 'PT1Y';
         try {
-            $parsedDuration3 = OneLogin_Saml2_Utils::parseDuration($invalidDuration);
+            $parsedDuration3 = Utils::parseDuration($invalidDuration);
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Invalid ISO 8601 duration', $e->getMessage());
         }
 
         $newDuration = 'P1Y1M';
-        $parsedDuration4 = OneLogin_Saml2_Utils::parseDuration($newDuration, $timestamp);
+        $parsedDuration4 = Utils::parseDuration($newDuration, $timestamp);
         $this->assertEquals(1428091225, $parsedDuration4);
 
         $negDuration = '-P14M';
-        $parsedDuration5 = OneLogin_Saml2_Utils::parseDuration($negDuration, $timestamp);
+        $parsedDuration5 = Utils::parseDuration($negDuration, $timestamp);
         $this->assertEquals(1357243225, $parsedDuration5);
     }
 
     /**
-    * Tests the parseSAML2Time method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::parseSAML2Time
-    */
+     * Tests the parseSAML2Time method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::parseSAML2Time
+     */
     public function testParseSAML2Time()
     {
         $time = 1386650371;
         $SAMLTime = '2013-12-10T04:39:31Z';
-        $this->assertEquals($time, OneLogin_Saml2_Utils::parseSAML2Time($SAMLTime));
+        $this->assertEquals($time, Utils::parseSAML2Time($SAMLTime));
 
         try {
-            OneLogin_Saml2_Utils::parseSAML2Time('invalidSAMLTime');
+            Utils::parseSAML2Time('invalidSAMLTime');
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Invalid SAML2 timestamp passed', $e->getMessage());
@@ -698,22 +725,22 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
 
         // Now test if toolkit supports miliseconds
         $SAMLTime2 = '2013-12-10T04:39:31.120Z';
-        $this->assertEquals($time, OneLogin_Saml2_Utils::parseSAML2Time($SAMLTime2));
+        $this->assertEquals($time, Utils::parseSAML2Time($SAMLTime2));
     }
 
     /**
-    * Tests the parseTime2SAML method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::parseTime2SAML
-    */
+     * Tests the parseTime2SAML method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::parseTime2SAML
+     */
     public function testParseTime2SAML()
     {
         $time = 1386650371;
         $SAMLTime = '2013-12-10T04:39:31Z';
-        $this->assertEquals($SAMLTime, OneLogin_Saml2_Utils::parseTime2SAML($time));
+        $this->assertEquals($SAMLTime, Utils::parseTime2SAML($time));
 
         try {
-            OneLogin_Saml2_Utils::parseTime2SAML('invalidtime');
+            Utils::parseTime2SAML('invalidtime');
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Failed to parse time string', $e->getMessage());
@@ -721,90 +748,87 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the getExpireTime method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::getExpireTime
-    */
+     * Tests the getExpireTime method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::getExpireTime
+     */
     public function testGetExpireTime()
     {
-        $this->assertNull(OneLogin_Saml2_Utils::getExpireTime());
+        $this->assertNull(Utils::getExpireTime());
 
-        $this->assertNotNull(OneLogin_Saml2_Utils::getExpireTime('PT1393462294S'));
+        $this->assertNotNull(Utils::getExpireTime('PT1393462294S'));
 
-        $this->assertEquals('1418186371', OneLogin_Saml2_Utils::getExpireTime('PT1393462294S', '2014-12-10T04:39:31Z'));
-        $this->assertEquals('1418186371', OneLogin_Saml2_Utils::getExpireTime('PT1393462294S', 1418186371));
+        $this->assertEquals('1418186371', Utils::getExpireTime('PT1393462294S', '2014-12-10T04:39:31Z'));
+        $this->assertEquals('1418186371', Utils::getExpireTime('PT1393462294S', 1418186371));
 
-        $this->assertNotEquals('1418186371', OneLogin_Saml2_Utils::getExpireTime('PT1393462294S', '2012-12-10T04:39:31Z'));
+        $this->assertNotEquals('1418186371', Utils::getExpireTime('PT1393462294S', '2012-12-10T04:39:31Z'));
     }
 
     /**
-    * Tests the query method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::query
-    */
+     * Tests the query method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::query
+     */
     public function testQuery()
     {
         $xml = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/valid_response.xml.base64'));
         $dom = new DOMDocument();
         $dom->loadXML($xml);
 
-        $assertionNodes = OneLogin_Saml2_Utils::query($dom, '/samlp:Response/saml:Assertion');
+        $assertionNodes = Utils::query($dom, '/samlp:Response/saml:Assertion');
         $this->assertEquals(1, $assertionNodes->length);
         $assertion = $assertionNodes->item(0);
         $this->assertEquals('saml:Assertion', $assertion->tagName);
 
-        $attributeStatementNodes = OneLogin_Saml2_Utils::query($dom, '/samlp:Response/saml:Assertion/saml:AttributeStatement');
+        $attributeStatementNodes = Utils::query($dom, '/samlp:Response/saml:Assertion/saml:AttributeStatement');
         $this->assertEquals(1, $attributeStatementNodes->length);
         $attributeStatement = $attributeStatementNodes->item(0);
         $this->assertEquals('saml:AttributeStatement', $attributeStatement->tagName);
 
-        $attributeStatementNodes2 = OneLogin_Saml2_Utils::query($dom, './saml:AttributeStatement', $assertion);
+        $attributeStatementNodes2 = Utils::query($dom, './saml:AttributeStatement', $assertion);
         $this->assertEquals(1, $attributeStatementNodes2->length);
         $attributeStatement2 = $attributeStatementNodes2->item(0);
         $this->assertEquals($attributeStatement, $attributeStatement2);
 
-        $signatureResNodes = OneLogin_Saml2_Utils::query($dom, '/samlp:Response/ds:Signature');
+        $signatureResNodes = Utils::query($dom, '/samlp:Response/ds:Signature');
         $this->assertEquals(1, $signatureResNodes->length);
         $signatureRes = $signatureResNodes->item(0);
         $this->assertEquals('ds:Signature', $signatureRes->tagName);
 
-        $signatureNodes = OneLogin_Saml2_Utils::query($dom, '/samlp:Response/saml:Assertion/ds:Signature');
+        $signatureNodes = Utils::query($dom, '/samlp:Response/saml:Assertion/ds:Signature');
         $this->assertEquals(1, $signatureNodes->length);
         $signature = $signatureNodes->item(0);
         $this->assertEquals('ds:Signature', $signature->tagName);
 
-        $signatureNodes2 = OneLogin_Saml2_Utils::query($dom, './ds:Signature', $assertion);
+        $signatureNodes2 = Utils::query($dom, './ds:Signature', $assertion);
         $this->assertEquals(1, $signatureNodes2->length);
         $signature2 = $signatureNodes2->item(0);
         $this->assertEquals($signature->textContent, $signature2->textContent);
         $this->assertNotEquals($signatureRes->textContent, $signature2->textContent);
 
-        $signatureNodes3 = OneLogin_Saml2_Utils::query($dom, './ds:SignatureValue', $assertion);
+        $signatureNodes3 = Utils::query($dom, './ds:SignatureValue', $assertion);
         $this->assertEquals(0, $signatureNodes3->length);
 
-        $signatureNodes4 = OneLogin_Saml2_Utils::query($dom, './ds:Signature/ds:SignatureValue', $assertion);
+        $signatureNodes4 = Utils::query($dom, './ds:Signature/ds:SignatureValue', $assertion);
         $this->assertEquals(1, $signatureNodes4->length);
 
-        $signatureNodes5 = OneLogin_Saml2_Utils::query($dom, './/ds:SignatureValue', $assertion);
+        $signatureNodes5 = Utils::query($dom, './/ds:SignatureValue', $assertion);
         $this->assertEquals(1, $signatureNodes5->length);
     }
 
     /**
-    * Tests the generateNameId method of the OneLogin_Saml2_Utils
-    * Adding a SPNameQualifier
-    *
-    * @covers OneLogin_Saml2_Utils::generateNameId
-    */
+     * Tests the generateNameId method of the Utils
+     * Adding a SPNameQualifier
+     *
+     * @covers OneLogin\Saml2\Utils::generateNameId
+     */
     public function testGenerateNameIdWithSPNameQualifier()
     {
-        //$xml = '<root xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">'.$decrypted.'</root>';
-        //$newDoc = new DOMDocument();
-
         $nameIdValue = 'ONELOGIN_ce998811003f4e60f8b07a311dc641621379cfde';
         $entityId = 'http://stuff.com/endpoints/metadata.php';
         $nameIDFormat = 'urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified';
 
-        $nameId = OneLogin_Saml2_Utils::generateNameId(
+        $nameId = Utils::generateNameId(
             $nameIdValue,
             $entityId,
             $nameIDFormat
@@ -818,9 +842,9 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         include $settingsDir.'settings1.php';
 
         $x509cert = $settingsInfo['idp']['x509cert'];
-        $key = OneLogin_Saml2_Utils::formatCert($x509cert);
+        $key = Utils::formatCert($x509cert);
 
-        $nameIdEnc = OneLogin_Saml2_Utils::generateNameId(
+        $nameIdEnc = Utils::generateNameId(
             $nameIdValue,
             $entityId,
             $nameIDFormat,
@@ -829,32 +853,44 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
 
         $nameidExpectedEnc = '<saml:EncryptedID><xenc:EncryptedData xmlns:xenc="http://www.w3.org/2001/04/xmlenc#" xmlns:dsig="http://www.w3.org/2000/09/xmldsig#" Type="http://www.w3.org/2001/04/xmlenc#Element"><xenc:EncryptionMethod Algorithm="http://www.w3.org/2001/04/xmlenc#aes128-cbc"/><dsig:KeyInfo xmlns:dsig="http://www.w3.org/2000/09/xmldsig#"><xenc:EncryptedKey><xenc:EncryptionMethod Algorithm="http://www.w3.org/2001/04/xmlenc#rsa-1_5"/><xenc:CipherData><xenc:CipherValue>';
         $this->assertContains($nameidExpectedEnc, $nameIdEnc);
+
+        // Check AES128_GCM support
+
+        $nameidExpectedEnc = '<saml:EncryptedID><xenc:EncryptedData xmlns:xenc="http://www.w3.org/2001/04/xmlenc#" xmlns:dsig="http://www.w3.org/2000/09/xmldsig#" Type="http://www.w3.org/2001/04/xmlenc#Element"><xenc:EncryptionMethod Algorithm="http://www.w3.org/2009/xmlenc11#aes128-gcm"/><dsig:KeyInfo xmlns:dsig="http://www.w3.org/2000/09/xmldsig#"><xenc:EncryptedKey><xenc:EncryptionMethod Algorithm="http://www.w3.org/2001/04/xmlenc#rsa-oaep-mgf1p"/><xenc:CipherData><xenc:CipherValue>';
+
+        $nameIdEnc = Utils::generateNameId(
+            $nameIdValue,
+            $entityId,
+            $nameIDFormat,
+            $key,
+            null,
+            XMLSecurityKey::AES128_GCM
+        );
+        $this->assertContains($nameidExpectedEnc, $nameIdEnc);
     }
 
     /**
-    * Tests the generateNameId method of the OneLogin_Saml2_Utils
+    * Tests the generateNameId method of the Utils
     *
-    * @covers OneLogin_Saml2_Utils::generateNameId
+    * @covers OneLogin\Saml2\Utils::generateNameId
     */
     public function testGenerateNameIdWithoutFormat()
     {
         $nameIdValue = 'ONELOGIN_ce998811003f4e60f8b07a311dc641621379cfde';
-
-        $nameId = OneLogin_Saml2_Utils::generateNameId(
+        $nameId = Utils::generateNameId(
             $nameIdValue,
             null,
             null
         );
-
         $expectedNameId = '<saml:NameID>ONELOGIN_ce998811003f4e60f8b07a311dc641621379cfde</saml:NameID>';
         $this->assertEquals($nameId, $expectedNameId);
     }
 
     /**
-    * Tests the generateNameId method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::generateNameId
-    */
+     * Tests the generateNameId method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::generateNameId
+     */
     public function testGenerateNameIdWithoutSPNameQualifier()
     {
         //$xml = '<root xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">'.$decrypted.'</root>';
@@ -864,7 +900,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $entityId = 'http://stuff.com/endpoints/metadata.php';
         $nameIDFormat = 'urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified';
 
-        $nameId = OneLogin_Saml2_Utils::generateNameId(
+        $nameId = Utils::generateNameId(
             $nameIdValue,
             null,
             $nameIDFormat
@@ -878,9 +914,9 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         include $settingsDir.'settings1.php';
 
         $x509cert = $settingsInfo['idp']['x509cert'];
-        $key = OneLogin_Saml2_Utils::formatCert($x509cert);
+        $key = Utils::formatCert($x509cert);
 
-        $nameIdEnc = OneLogin_Saml2_Utils::generateNameId(
+        $nameIdEnc = Utils::generateNameId(
             $nameIdValue,
             null,
             $nameIDFormat,
@@ -892,56 +928,10 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the generateNameId method of the OneLogin_Saml2_Utils
-    * Adding a NameQualifier
-    *
-    * @covers OneLogin_Saml2_Utils::generateNameId
-    */
-    public function testGenerateNameIdWithNameQualifier()
-    {
-        //$xml = '<root xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">'.$decrypted.'</root>';
-        //$newDoc = new DOMDocument();
-
-        $nameIdValue = 'ONELOGIN_ce998811003f4e60f8b07a311dc641621379cfde';
-        $entityId = 'http://stuff.com/endpoints/metadata.php';
-        $nameIDFormat = 'urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified';
-        $nameQualifier = 'http://stuff.com/endpoints/acs.php';
-
-        $nameId = OneLogin_Saml2_Utils::generateNameId(
-            $nameIdValue,
-            $entityId,
-            $nameIDFormat,
-            null,
-            $nameQualifier
-        );
-
-        $expectedNameId = '<saml:NameID SPNameQualifier="http://stuff.com/endpoints/metadata.php" NameQualifier="http://stuff.com/endpoints/acs.php" Format="urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified">ONELOGIN_ce998811003f4e60f8b07a311dc641621379cfde</saml:NameID>';
-
-        $this->assertEquals($nameId, $expectedNameId);
-
-        $settingsDir = TEST_ROOT .'/settings/';
-        include $settingsDir.'settings1.php';
-
-        $x509cert = $settingsInfo['idp']['x509cert'];
-        $key = OneLogin_Saml2_Utils::formatCert($x509cert);
-
-        $nameIdEnc = OneLogin_Saml2_Utils::generateNameId(
-            $nameIdValue,
-            $entityId,
-            $nameIDFormat,
-            $key,
-            $nameQualifier
-        );
-
-        $nameidExpectedEnc = '<saml:EncryptedID><xenc:EncryptedData xmlns:xenc="http://www.w3.org/2001/04/xmlenc#" xmlns:dsig="http://www.w3.org/2000/09/xmldsig#" Type="http://www.w3.org/2001/04/xmlenc#Element"><xenc:EncryptionMethod Algorithm="http://www.w3.org/2001/04/xmlenc#aes128-cbc"/><dsig:KeyInfo xmlns:dsig="http://www.w3.org/2000/09/xmldsig#"><xenc:EncryptedKey><xenc:EncryptionMethod Algorithm="http://www.w3.org/2001/04/xmlenc#rsa-1_5"/><xenc:CipherData><xenc:CipherValue>';
-        $this->assertContains($nameidExpectedEnc, $nameIdEnc);
-    }
-
-    /**
-    * Tests the deleteLocalSession method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::deleteLocalSession
-    */
+     * Tests the deleteLocalSession method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::deleteLocalSession
+     */
     public function testDeleteLocalSession()
     {
         if (getenv("TRAVIS")) {
@@ -956,7 +946,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
             $this->assertTrue(isset($_SESSION['samltest']));
             $this->assertTrue($_SESSION['samltest']);
 
-            OneLogin_Saml2_Utils::deleteLocalSession();
+            Utils::deleteLocalSession();
             $this->assertEmpty($_SESSION);
             $this->assertFalse(isset($_SESSION['samltest']));
 
@@ -965,94 +955,94 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
             error_reporting($prev);
 
             $_SESSION['samltest'] = true;
-            OneLogin_Saml2_Utils::deleteLocalSession();
+            Utils::deleteLocalSession();
             $this->assertEmpty($_SESSION);
             $this->assertFalse(isset($_SESSION['samltest']));
         }
     }
 
     /**
-    * Tests the isSessionStarted method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::isSessionStarted
-    */
+     * Tests the isSessionStarted method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::isSessionStarted
+     */
     public function testisSessionStarted()
     {
         if (getenv("TRAVIS")) {
             // Can't test that on TRAVIS
             $this->markTestSkipped("Can't test that on TRAVIS");
         } else {
-            $this->assertFalse(OneLogin_Saml2_Utils::isSessionStarted());
+            $this->assertFalse(Utils::isSessionStarted());
 
             $prev = error_reporting(0);
             session_start();
             error_reporting($prev);
 
-            $this->assertTrue(OneLogin_Saml2_Utils::isSessionStarted());
+            $this->assertTrue(Utils::isSessionStarted());
         }
     }
 
 
     /**
-    * Tests the calculateX509Fingerprint method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::calculateX509Fingerprint
-    */
+     * Tests the calculateX509Fingerprint method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::calculateX509Fingerprint
+     */
     public function testCalculateX509Fingerprint()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $certPath = $settings->getCertPath();
 
         $key = file_get_contents($certPath.'sp.key');
         $cert = file_get_contents($certPath.'sp.crt');
 
-        $this->assertNull(OneLogin_Saml2_Utils::calculateX509Fingerprint($key));
+        $this->assertNull(Utils::calculateX509Fingerprint($key));
 
-        $this->assertNull(OneLogin_Saml2_Utils::calculateX509Fingerprint(""));
+        $this->assertNull(Utils::calculateX509Fingerprint(""));
 
-        $this->assertNull(OneLogin_Saml2_Utils::calculateX509Fingerprint($settingsInfo['idp']['x509cert']));
+        $this->assertNull(Utils::calculateX509Fingerprint($settingsInfo['idp']['x509cert']));
 
-        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', OneLogin_Saml2_Utils::calculateX509Fingerprint(OneLogin_Saml2_Utils::formatCert($settingsInfo['idp']['x509cert'])));
+        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', Utils::calculateX509Fingerprint(Utils::formatCert($settingsInfo['idp']['x509cert'])));
 
-        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', OneLogin_Saml2_Utils::calculateX509Fingerprint($cert));
+        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', Utils::calculateX509Fingerprint($cert));
 
-        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', OneLogin_Saml2_Utils::calculateX509Fingerprint($cert, 'sha1'));
+        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', Utils::calculateX509Fingerprint($cert, 'sha1'));
 
-        $this->assertEquals('c51cfa06c7a49767f6eab18238eae1c56708e29264da3d11f538a12cd2c357ba', OneLogin_Saml2_Utils::calculateX509Fingerprint($cert, 'sha256'));
+        $this->assertEquals('c51cfa06c7a49767f6eab18238eae1c56708e29264da3d11f538a12cd2c357ba', Utils::calculateX509Fingerprint($cert, 'sha256'));
 
-        $this->assertEquals('bc5826e6f9429247254bae5e3c650e6968a36a62d23075eb168134978d88600559c10830c28711b2c29c7947c0c2eb1d', OneLogin_Saml2_Utils::calculateX509Fingerprint($cert, 'sha384'));
+        $this->assertEquals('bc5826e6f9429247254bae5e3c650e6968a36a62d23075eb168134978d88600559c10830c28711b2c29c7947c0c2eb1d', Utils::calculateX509Fingerprint($cert, 'sha384'));
 
-        $this->assertEquals('3db29251b97559c67988ea0754cb0573fc409b6f75d89282d57cfb75089539b0bbdb2dcd9ec6e032549ecbc466439d5992e18db2cf5494ca2fe1b2e16f348dff', OneLogin_Saml2_Utils::calculateX509Fingerprint($cert, 'sha512'));
+        $this->assertEquals('3db29251b97559c67988ea0754cb0573fc409b6f75d89282d57cfb75089539b0bbdb2dcd9ec6e032549ecbc466439d5992e18db2cf5494ca2fe1b2e16f348dff', Utils::calculateX509Fingerprint($cert, 'sha512'));
     }
 
     /**
-    * Tests the formatFingerPrint method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::formatFingerPrint
-    */
+     * Tests the formatFingerPrint method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::formatFingerPrint
+     */
     public function testFormatFingerPrint()
     {
         $fingerPrint1 = 'AF:E7:1C:28:EF:74:0B:C8:74:25:BE:13:A2:26:3D:37:97:1D:A1:F9';
-        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', OneLogin_Saml2_Utils::formatFingerPrint($fingerPrint1));
+        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', Utils::formatFingerPrint($fingerPrint1));
 
         $fingerPrint2 = 'afe71c28ef740bc87425be13a2263d37971da1f9';
-        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', OneLogin_Saml2_Utils::formatFingerPrint($fingerPrint2));
+        $this->assertEquals('afe71c28ef740bc87425be13a2263d37971da1f9', Utils::formatFingerPrint($fingerPrint2));
     }
 
     /**
-    * Tests the decryptElement method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::decryptElement
-    */
+     * Tests the decryptElement method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::decryptElement
+     */
     public function testDecryptElement()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
 
         $key = $settings->getSPkey();
         $seckey = new XMLSecurityKey(XMLSecurityKey::RSA_1_5, array('type'=>'private'));
@@ -1063,7 +1053,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $domNameIdEnc->loadXML($xmlNameIdEnc);
         $encryptedNameIDNodes = $domNameIdEnc->getElementsByTagName('EncryptedID');
         $encryptedData = $encryptedNameIDNodes->item(0)->firstChild;
-        $decryptedNameId = OneLogin_Saml2_Utils::decryptElement($encryptedData, $seckey);
+        $decryptedNameId = Utils::decryptElement($encryptedData, $seckey);
         $this->assertEquals('saml:NameID', $decryptedNameId->tagName);
         $this->assertEquals('2de11defd199f8d5bb63f9b7deb265ba5c675c10', $decryptedNameId->nodeValue);
 
@@ -1074,21 +1064,21 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $encryptedAssertionEncNode = $encryptedAssertionEncNodes->item(0);
         $encryptedDataAssertNodes = $encryptedAssertionEncNode->getElementsByTagName('EncryptedData');
         $encryptedDataAssert = $encryptedDataAssertNodes->item(0);
-        $decryptedAssertion = OneLogin_Saml2_Utils::decryptElement($encryptedDataAssert, $seckey);
+        $decryptedAssertion = Utils::decryptElement($encryptedDataAssert, $seckey);
 
         $this->assertEquals('saml:Assertion', $decryptedAssertion->tagName);
 
         try {
-            $res = OneLogin_Saml2_Utils::decryptElement($encryptedNameIDNodes->item(0), $seckey);
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $res = Utils::decryptElement($encryptedNameIDNodes->item(0), $seckey);
+            $this->fail('ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertContains('Algorithm mismatch between input key and key in message', $e->getMessage());
         }
 
         $key2 = file_get_contents(TEST_ROOT . '/data/misc/sp2.key');
         $seckey2 = new XMLSecurityKey(XMLSecurityKey::RSA_1_5, array('type'=>'private'));
         $seckey2->loadKey($key2);
-        $decryptedNameId2 = OneLogin_Saml2_Utils::decryptElement($encryptedData, $seckey2);
+        $decryptedNameId2 = Utils::decryptElement($encryptedData, $seckey2);
         $this->assertEquals('saml:NameID', $decryptedNameId2->tagName);
         $this->assertEquals('2de11defd199f8d5bb63f9b7deb265ba5c675c10', $decryptedNameId2->nodeValue);
 
@@ -1096,9 +1086,9 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $seckey3 = new XMLSecurityKey(XMLSecurityKey::RSA_SHA512, array('type'=>'private'));
         $seckey3->loadKey($key3);
         try {
-            $res = OneLogin_Saml2_Utils::decryptElement($encryptedData, $seckey3);
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $res = Utils::decryptElement($encryptedData, $seckey3);
+            $this->fail('ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertContains('Algorithm mismatch between input key and key used to encrypt  the symmetric key for the message', $e->getMessage());
         }
 
@@ -1108,7 +1098,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $encryptedNameIDNodes2 = $domNameIdEnc2->getElementsByTagName('EncryptedID');
         $encryptedData2 = $encryptedNameIDNodes2->item(0)->firstChild;
         try {
-            $res = OneLogin_Saml2_Utils::decryptElement($encryptedData2, $seckey);
+            $res = Utils::decryptElement($encryptedData2, $seckey);
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Unable to locate algorithm for this Encrypted Key', $e->getMessage());
@@ -1120,32 +1110,32 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $encryptedNameIDNodes3 = $domNameIdEnc3->getElementsByTagName('EncryptedID');
         $encryptedData3 = $encryptedNameIDNodes3->item(0)->firstChild;
         try {
-            $res = OneLogin_Saml2_Utils::decryptElement($encryptedData3, $seckey);
-            $this->fail('OneLogin_Saml2_ValidationError was not raised');
-        } catch (OneLogin_Saml2_ValidationError $e) {
+            $res = Utils::decryptElement($encryptedData3, $seckey);
+            $this->fail('ValidationError was not raised');
+        } catch (ValidationError $e) {
             $this->assertContains('Algorithm mismatch between input key and key in message', $e->getMessage());
         }
     }
 
     /**
-    * Tests the addSign method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::addSign
-    */
+     * Tests the addSign method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::addSign
+     */
     public function testAddSign()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $key = $settings->getSPkey();
         $cert = $settings->getSPcert();
 
         $xmlAuthn = base64_decode(file_get_contents(TEST_ROOT . '/data/requests/authn_request.xml.base64'));
-        $xmlAuthnSigned = OneLogin_Saml2_Utils::addSign($xmlAuthn, $key, $cert);
+        $xmlAuthnSigned = Utils::addSign($xmlAuthn, $key, $cert);
         $this->assertContains('<ds:SignatureValue>', $xmlAuthnSigned);
-        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2000/09/xmldsig#rsa-sha1"/>', $xmlAuthnSigned);
-        $this->assertContains('<ds:DigestMethod Algorithm="http://www.w3.org/2000/09/xmldsig#sha1"/>', $xmlAuthnSigned);
+        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $xmlAuthnSigned);
+        $this->assertContains('<ds:DigestMethod Algorithm="http://www.w3.org/2001/04/xmlenc#sha256"/>', $xmlAuthnSigned);
         $res = new DOMDocument();
         $res->loadXML($xmlAuthnSigned);
         $dsSignature = $res->firstChild->firstChild->nextSibling->nextSibling;
@@ -1153,9 +1143,9 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
 
         $dom = new DOMDocument();
         $dom->loadXML($xmlAuthn);
-        $xmlAuthnSigned2 = OneLogin_Saml2_Utils::addSign($dom, $key, $cert, XMLSecurityKey::RSA_SHA256, XMLSecurityDSig::SHA512);
+        $xmlAuthnSigned2 = Utils::addSign($dom, $key, $cert, XMLSecurityKey::RSA_SHA384, XMLSecurityDSig::SHA512);
         $this->assertContains('<ds:SignatureValue>', $xmlAuthnSigned2);
-        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $xmlAuthnSigned2);
+        $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha384"/>', $xmlAuthnSigned2);
         $this->assertContains('<ds:DigestMethod Algorithm="http://www.w3.org/2001/04/xmlenc#sha512"/>', $xmlAuthnSigned2);
         $res2 = new DOMDocument();
         $res2->loadXML($xmlAuthnSigned2);
@@ -1163,7 +1153,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $this->assertContains('ds:Signature', $dsSignature2->tagName);
 
         $xmlLogoutReq = base64_decode(file_get_contents(TEST_ROOT . '/data/logout_requests/logout_request.xml.base64'));
-        $xmlLogoutReqSigned = OneLogin_Saml2_Utils::addSign($xmlLogoutReq, $key, $cert, XMLSecurityKey::RSA_SHA256, XMLSecurityDSig::SHA512);
+        $xmlLogoutReqSigned = Utils::addSign($xmlLogoutReq, $key, $cert, XMLSecurityKey::RSA_SHA256, XMLSecurityDSig::SHA512);
         $this->assertContains('<ds:SignatureValue>', $xmlLogoutReqSigned);
         $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $xmlLogoutReqSigned);
         $this->assertContains('<ds:DigestMethod Algorithm="http://www.w3.org/2001/04/xmlenc#sha512"/>', $xmlLogoutReqSigned);
@@ -1173,7 +1163,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $this->assertContains('ds:Signature', $dsSignature3->tagName);
 
         $xmlLogoutRes = base64_decode(file_get_contents(TEST_ROOT . '/data/logout_responses/logout_response.xml.base64'));
-        $xmlLogoutResSigned = OneLogin_Saml2_Utils::addSign($xmlLogoutRes, $key, $cert, XMLSecurityKey::RSA_SHA256, XMLSecurityDSig::SHA512);
+        $xmlLogoutResSigned = Utils::addSign($xmlLogoutRes, $key, $cert, XMLSecurityKey::RSA_SHA256, XMLSecurityDSig::SHA512);
         $this->assertContains('<ds:SignatureValue>', $xmlLogoutResSigned);
         $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $xmlLogoutResSigned);
         $this->assertContains('<ds:DigestMethod Algorithm="http://www.w3.org/2001/04/xmlenc#sha512"/>', $xmlLogoutResSigned);
@@ -1183,7 +1173,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $this->assertContains('ds:Signature', $dsSignature4->tagName);
 
         $xmlMetadata = file_get_contents(TEST_ROOT . '/data/metadata/metadata_settings1.xml');
-        $xmlMetadataSigned = OneLogin_Saml2_Utils::addSign($xmlMetadata, $key, $cert, XMLSecurityKey::RSA_SHA256, XMLSecurityDSig::SHA512);
+        $xmlMetadataSigned = Utils::addSign($xmlMetadata, $key, $cert, XMLSecurityKey::RSA_SHA256, XMLSecurityDSig::SHA512);
         $this->assertContains('<ds:SignatureValue>', $xmlMetadataSigned);
         $this->assertContains('<ds:SignatureMethod Algorithm="http://www.w3.org/2001/04/xmldsig-more#rsa-sha256"/>', $xmlMetadataSigned);
         $this->assertContains('<ds:DigestMethod Algorithm="http://www.w3.org/2001/04/xmlenc#sha512"/>', $xmlMetadataSigned);
@@ -1194,47 +1184,47 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-    * Tests the validateSign method of the OneLogin_Saml2_Utils
-    *
-    * @covers OneLogin_Saml2_Utils::validateSign
-    */
+     * Tests the validateSign method of the Utils
+     *
+     * @covers OneLogin\Saml2\Utils::validateSign
+     */
     public function testValidateSign()
     {
         $settingsDir = TEST_ROOT .'/settings/';
         include $settingsDir.'settings1.php';
 
-        $settings = new OneLogin_Saml2_Settings($settingsInfo);
+        $settings = new Settings($settingsInfo);
         $idpData = $settings->getIdPData();
         $cert = $idpData['x509cert'];
-        $fingerprint = OneLogin_Saml2_Utils::calculateX509Fingerprint($cert);
-        $fingerprint256 = OneLogin_Saml2_Utils::calculateX509Fingerprint($cert, 'sha256');
+        $fingerprint = Utils::calculateX509Fingerprint($cert);
+        $fingerprint256 = Utils::calculateX509Fingerprint($cert, 'sha256');
 
         $xmlMetadataSigned = file_get_contents(TEST_ROOT . '/data/metadata/signed_metadata_settings1.xml');
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlMetadataSigned, $cert));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlMetadataSigned, null, $fingerprint));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlMetadataSigned, null, $fingerprint, 'sha1'));
-        $this->assertFalse(OneLogin_Saml2_Utils::validateSign($xmlMetadataSigned, null, $fingerprint, 'sha256'));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlMetadataSigned, null, $fingerprint256, 'sha256'));
+        $this->assertTrue(Utils::validateSign($xmlMetadataSigned, $cert));
+        $this->assertTrue(Utils::validateSign($xmlMetadataSigned, null, $fingerprint));
+        $this->assertTrue(Utils::validateSign($xmlMetadataSigned, null, $fingerprint, 'sha1'));
+        $this->assertFalse(Utils::validateSign($xmlMetadataSigned, null, $fingerprint, 'sha256'));
+        $this->assertTrue(Utils::validateSign($xmlMetadataSigned, null, $fingerprint256, 'sha256'));
 
         $xmlResponseMsgSigned = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/signed_message_response.xml.base64'));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlResponseMsgSigned, $cert));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlResponseMsgSigned, null, $fingerprint));
+        $this->assertTrue(Utils::validateSign($xmlResponseMsgSigned, $cert));
+        $this->assertTrue(Utils::validateSign($xmlResponseMsgSigned, null, $fingerprint));
 
         $xmlResponseAssertSigned = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/signed_assertion_response.xml.base64'));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlResponseAssertSigned, $cert));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlResponseAssertSigned, null, $fingerprint));
+        $this->assertTrue(Utils::validateSign($xmlResponseAssertSigned, $cert));
+        $this->assertTrue(Utils::validateSign($xmlResponseAssertSigned, null, $fingerprint));
 
         $xmlResponseDoubleSigned = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/double_signed_response.xml.base64'));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlResponseDoubleSigned, $cert));
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($xmlResponseDoubleSigned, null, $fingerprint));
+        $this->assertTrue(Utils::validateSign($xmlResponseDoubleSigned, $cert));
+        $this->assertTrue(Utils::validateSign($xmlResponseDoubleSigned, null, $fingerprint));
 
         $dom = new DOMDocument();
         $dom->loadXML($xmlResponseMsgSigned);
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($dom, $cert));
+        $this->assertTrue(Utils::validateSign($dom, $cert));
 
         $dom->firstChild->firstChild->nodeValue = 'https://example.com/other-idp';
         try {
-            $this->assertFalse(OneLogin_Saml2_Utils::validateSign($dom, $cert));
+            $this->assertFalse(Utils::validateSign($dom, $cert));
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Reference validation failed', $e->getMessage());
@@ -1243,25 +1233,25 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
         $dom2 = new DOMDocument();
         $dom2->loadXML($xmlResponseMsgSigned);
         $assertElem = $dom2->firstChild->firstChild->nextSibling->nextSibling;
-        $this->assertTrue(OneLogin_Saml2_Utils::validateSign($assertElem, $cert));
+        $this->assertTrue(Utils::validateSign($assertElem, $cert));
 
         $dom3 = new DOMDocument();
         $dom3->loadXML($xmlResponseMsgSigned);
         $dom3->firstChild->firstChild->nodeValue = 'https://example.com/other-idp';
         $assertElem2 = $dom3->firstChild->firstChild->nextSibling->nextSibling;
         try {
-            $this->assertTrue(OneLogin_Saml2_Utils::validateSign($assertElem2, $cert));
+            $this->assertTrue(Utils::validateSign($assertElem2, $cert));
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Reference validation failed', $e->getMessage());
         }
 
         $invalidFingerprint = 'afe71c34ef740bc87434be13a2263d31271da1f9';
-        $this->assertFalse(OneLogin_Saml2_Utils::validateSign($xmlMetadataSigned, null, $invalidFingerprint));
+        $this->assertFalse(Utils::validateSign($xmlMetadataSigned, null, $invalidFingerprint));
 
         $noSigned = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/invalids/no_signature.xml.base64'));
         try {
-            $this->assertFalse(OneLogin_Saml2_Utils::validateSign($noSigned, $cert));
+            $this->assertFalse(Utils::validateSign($noSigned, $cert));
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Cannot locate Signature Node', $e->getMessage());
@@ -1269,7 +1259,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
 
         $noKey = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/invalids/no_key.xml.base64'));
         try {
-            $this->assertFalse(OneLogin_Saml2_Utils::validateSign($noKey, $cert));
+            $this->assertFalse(Utils::validateSign($noKey, $cert));
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('We have no idea about the key', $e->getMessage());
@@ -1277,7 +1267,7 @@ class OneLogin_Saml2_UtilsTest extends PHPUnit_Framework_TestCase
 
         $signatureWrapping = base64_decode(file_get_contents(TEST_ROOT . '/data/responses/invalids/signature_wrapping_attack.xml.base64'));
         try {
-            $this->assertFalse(OneLogin_Saml2_Utils::validateSign($signatureWrapping, $cert));
+            $this->assertFalse(Utils::validateSign($signatureWrapping, $cert));
             $this->fail('Exception was not raised');
         } catch (Exception $e) {
             $this->assertContains('Reference validation failed', $e->getMessage());
