@@ -1,4 +1,5 @@
 <?php
+ 
 /**
  *  SAML Handler
  */
@@ -7,9 +8,12 @@ session_start();
 
 require_once dirname(__DIR__).'/_toolkit_loader.php';
 
+use OneLogin\Saml2\Auth;
+use OneLogin\Saml2\Utils;
+
 require_once 'settings.php';
 
-$auth = new OneLogin_Saml2_Auth($settingsInfo);
+$auth = new Auth($settingsInfo);
 
 if (isset($_GET['sso'])) {
     $auth->login();
@@ -27,12 +31,12 @@ if (isset($_GET['sso'])) {
     $auth->login($returnTo);
 } else if (isset($_GET['slo'])) {
     $returnTo = null;
-    $parameters = array();
+    $paramters = array();
     $nameId = null;
     $sessionIndex = null;
     $nameIdFormat = null;
-    $samlNameIdNameQualifier = null;
-    $samlNameIdSPNameQualifier = null;
+    $nameIdNameQualifier = null;
+    $nameIdSPNameQualifier = null;
 
     if (isset($_SESSION['samlNameId'])) {
         $nameId = $_SESSION['samlNameId'];
@@ -50,7 +54,7 @@ if (isset($_GET['sso'])) {
         $sessionIndex = $_SESSION['samlSessionIndex'];
     }
 
-    $auth->logout($returnTo, $parameters, $nameId, $sessionIndex, false, $nameIdFormat, $nameIdNameQualifier, $nameIdSPNameQualifier);
+    $auth->logout($returnTo, $paramters, $nameId, $sessionIndex, false, $nameIdFormat, $nameIdNameQualifier, $nameIdSPNameQualifier);
 
     # If LogoutRequest ID need to be saved in order to later validate it, do instead
     # $sloBuiltUrl = $auth->logout(null, $paramters, $nameId, $sessionIndex, true);
@@ -72,14 +76,14 @@ if (isset($_GET['sso'])) {
     $errors = $auth->getErrors();
 
     if (!empty($errors)) {
-        echo '<p>',implode(', ', $errors),'</p>';
+        echo '<p>' . implode(', ', $errors) . '</p>';
         if ($auth->getSettings()->isDebugActive()) {
             echo '<p>'.$auth->getLastErrorReason().'</p>';
         }
     }
 
     if (!$auth->isAuthenticated()) {
-        echo "<p>Not authenticated</p>";
+        echo '<p>Not authenticated</p>';
         exit();
     }
 
@@ -89,8 +93,9 @@ if (isset($_GET['sso'])) {
     $_SESSION['samlNameIdNameQualifier'] = $auth->getNameIdNameQualifier();
     $_SESSION['samlNameIdSPNameQualifier'] = $auth->getNameIdSPNameQualifier();
     $_SESSION['samlSessionIndex'] = $auth->getSessionIndex();
+
     unset($_SESSION['AuthNRequestID']);
-    if (isset($_POST['RelayState']) && OneLogin_Saml2_Utils::getSelfURL() != $_POST['RelayState']) {
+    if (isset($_POST['RelayState']) && Utils::getSelfURL() != $_POST['RelayState']) {
         $auth->redirectTo($_POST['RelayState']);
     }
 } else if (isset($_GET['sls'])) {
@@ -105,7 +110,7 @@ if (isset($_GET['sso'])) {
     if (empty($errors)) {
         echo '<p>Sucessfully logged out</p>';
     } else {
-        echo '<p>', implode(', ', $errors), '</p>';
+        echo '<p>' . implode(', ', $errors) . '</p>';
         if ($auth->getSettings()->isDebugActive()) {
             echo '<p>'.$auth->getLastErrorReason().'</p>';
         }
