@@ -6,8 +6,11 @@
 
 require_once dirname(__DIR__).'/_toolkit_loader.php';
 
+use OneLogin\Saml2\Auth;
+use OneLogin\Saml2\Error;
+
 try {
-    $auth = new OneLogin_Saml2_Auth();
+    $auth = new Auth();
     $settings = $auth->getSettings();
     $metadata = $settings->getSPMetadata();
     $errors = $settings->validateMetadata($metadata);
@@ -15,9 +18,9 @@ try {
         header('Content-Type: text/xml');
         echo $metadata;
     } else {
-        throw new OneLogin_Saml2_Error(
+        throw new Error(
             'Invalid SP metadata: '.implode(', ', $errors),
-            OneLogin_Saml2_Error::METADATA_SP_INVALID
+            Error::METADATA_SP_INVALID
         );
     }
 } catch (Exception $e) {
