@@ -1,6 +1,6 @@
 <?php
 
-$settings = array (
+$settings = array(
     // If 'strict' is True, then the PHP Toolkit will reject unsigned
     // or unencrypted messages if it expects them signed or encrypted
     // Also will reject the messages if not strictly follow the SAML
@@ -17,12 +17,12 @@ $settings = array (
     'baseurl' => null,
 
     // Service Provider Data that we are deploying
-    'sp' => array (
+    'sp' => array(
         // Identifier of the SP entity  (must be a URI)
         'entityId' => '',
         // Specifies info about where and how the <AuthnResponse> message MUST be
         // returned to the requester, in this case our SP.
-        'assertionConsumerService' => array (
+        'assertionConsumerService' => array(
             // URL Location where the <Response> from the IdP will be returned
             'url' => '',
             // SAML protocol binding to be used when returning the <Response>
@@ -48,7 +48,7 @@ $settings = array (
         ),
         // Specifies info about where and how the <Logout Response> message MUST be
         // returned to the requester, in this case our SP.
-        'singleLogoutService' => array (
+        'singleLogoutService' => array(
             // URL Location where the <Response> from the IdP will be returned
             'url' => '',
             // SAML protocol binding to be used when returning the <Response>
@@ -77,11 +77,11 @@ $settings = array (
     ),
 
     // Identity Provider Data that we want connect with our SP
-    'idp' => array (
+    'idp' => array(
         // Identifier of the IdP entity  (must be a URI)
         'entityId' => '',
         // SSO endpoint info of the IdP. (Authentication Request protocol)
-        'singleSignOnService' => array (
+        'singleSignOnService' => array(
             // URL Target of the IdP where the SP will send the Authentication Request Message
             'url' => '',
             // SAML protocol binding to be used when returning the <Response>
@@ -90,10 +90,10 @@ $settings = array (
             'binding' => 'urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect',
         ),
         // SLO endpoint info of the IdP.
-        'singleLogoutService' => array (
+        'singleLogoutService' => array(
             // URL Location of the IdP where the SP will send the SLO Request
             'url' => '',
-            // URL location of the IdP where the SP will send the SLO Response (ResponseLocation)
+            // URL location of the IdP where the SP SLO Response will be sent (ResponseLocation)
             // if not set, url for the SLO Request will be used
             'responseUrl' => '',
             // SAML protocol binding to be used when returning the <Response>
