@@ -1,28 +1,28 @@
 <?php
-    $settingsInfo = array (
+    $settingsInfo = array(
         'strict' => false,
         'debug' => false,
-        'sp' => array (
+        'sp' => array(
             'entityId' => 'http://stuff.com/endpoints/metadata.php',
-            'assertionConsumerService' => array (
+            'assertionConsumerService' => array(
                 'url' => 'http://stuff.com/endpoints/endpoints/acs.php',
             ),
-            'singleLogoutService' => array (
+            'singleLogoutService' => array(
                 'url' => 'http://stuff.com/endpoints/endpoints/sls.php',
             ),
             'NameIDFormat' => 'urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified',
-            'attributeConsumingService' => array (
+            'attributeConsumingService' => array(
                 'serviceName' => 'Service Name',
                 'serviceDescription' => 'Service Description',
-                'requestedAttributes' => array (
-                    array (
-                        'nameFormat' => \OneLogin_Saml2_Constants::ATTRNAME_FORMAT_BASIC,
+                'requestedAttributes' => array(
+                    array(
+                        'nameFormat' => OneLogin\Saml2\Constants::ATTRNAME_FORMAT_BASIC,
                         'isRequired' => false,
                         'name' => 'userType',
                         'attributeValue' => array('userType', 'admin')
                     ),
-                    array (
-                        'nameFormat' => \OneLogin_Saml2_Constants::ATTRNAME_FORMAT_URI,
+                    array(
+                        'nameFormat' => OneLogin\Saml2\Constants::ATTRNAME_FORMAT_URI,
                         'isRequired' => true,
                         'name' => 'urn:oid:0.9.2342.19200300.100.1.1',
                         'friendlyName' => 'uid'
@@ -30,34 +30,34 @@
                 )
             )
         ),
-        'idp' => array (
+        'idp' => array(
             'entityId' => 'http://idp.example.com/',
-            'singleSignOnService' => array (
+            'singleSignOnService' => array(
                 'url' => 'http://idp.example.com/SSOService.php',
             ),
-            'singleLogoutService' => array (
+            'singleLogoutService' => array(
                 'url' => 'http://idp.example.com/SingleLogoutService.php',
             ),
             'x509cert' => 'MIICgTCCAeoCCQCbOlrWDdX7FTANBgkqhkiG9w0BAQUFADCBhDELMAkGA1UEBhMCTk8xGDAWBgNVBAgTD0FuZHJlYXMgU29sYmVyZzEMMAoGA1UEBxMDRm9vMRAwDgYDVQQKEwdVTklORVRUMRgwFgYDVQQDEw9mZWlkZS5lcmxhbmcubm8xITAfBgkqhkiG9w0BCQEWEmFuZHJlYXNAdW5pbmV0dC5ubzAeFw0wNzA2MTUxMjAxMzVaFw0wNzA4MTQxMjAxMzVaMIGEMQswCQYDVQQGEwJOTzEYMBYGA1UECBMPQW5kcmVhcyBTb2xiZXJnMQwwCgYDVQQHEwNGb28xEDAOBgNVBAoTB1VOSU5FVFQxGDAWBgNVBAMTD2ZlaWRlLmVybGFuZy5ubzEhMB8GCSqGSIb3DQEJARYSYW5kcmVhc0B1bmluZXR0Lm5vMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDivbhR7P516x/S3BqKxupQe0LONoliupiBOesCO3SHbDrl3+q9IbfnfmE04rNuMcPsIxB161TdDpIesLCn7c8aPHISKOtPlAeTZSnb8QAu7aRjZq3+PbrP5uW3TcfCGPtKTytHOge/OlJbo078dVhXQ14d1EDwXJW1rRXuUt4C8QIDAQABMA0GCSqGSIb3DQEBBQUAA4GBACDVfp86HObqY+e8BUoWQ9+VMQx1ASDohBjwOsg2WykUqRXF+dLfcUH9dWR63CtZIKFDbStNomPnQz7nbK+onygwBspVEbnHuUihZq3ZUdmumQqCw4Uvs/1Uvq3orOo/WJVhTyvLgFVK2QarQ4/67OZfHd7R+POBXhophSMv1ZOo',
         ),
 
-        'security' => array (
+        'security' => array(
             'authnRequestsSigned' => false,
             'wantAssertionsSigned' => false,
             'signMetadata' => false,
         ),
-        'contactPerson' => array (
-            'technical' => array (
+        'contactPerson' => array(
+            'technical' => array(
                 'givenName' => 'technical_name',
                 'emailAddress' => 'technical@example.com',
             ),
-            'support' => array (
+            'support' => array(
                 'givenName' => 'support_name',
                 'emailAddress' => 'support@example.com',
             ),
         ),
 
-        'organization' => array (
+        'organization' => array(
             'en-US' => array(
                 'name' => 'sp_test',
                 'displayname' => 'SP test',
